rama

Module conversion

Traits

FromRef

Create Self from a reference T

RamaFrom

Alternative for From which can be implemented by external crates for external types

RamaInto

Alternative for Into which can be implemented by external crates for external types

RamaTryFrom

Alternative for TryFrom which can be implemented by external crates for external types

RamaTryInto

Alternative for TryInto which can be implemented by external crates for external types

Derive Macros

FromRef

Derive an implementation of FromRef for each field in a struct.