rama::conversion

Trait RamaFrom

pub trait RamaFrom<T, CrateMarker = ()> {
    // Required method
    fn rama_from(value: T) -> Self;
}

Alternative for From which can be implemented by external crates for external types

To implement this trait, use a crate local CrateMarker generic type More info: https://ramaproxy.org/book/intro/patterns.html#working-around-the-orphan-rule-in-specific-cases

Note: RamaFrom has a blacket implementation for types that implement From

Required Methods

fn rama_from(value: T) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.