rama::conversion

Trait RamaTryFrom

pub trait RamaTryFrom<T, CrateMarker = ()>: Sized {
    type Error;

    // Required method
    fn rama_try_from(value: T) -> Result<Self, Self::Error>;
}

Alternative for TryFrom which can be implemented by external crates for external types

To implement this trait, use a crate local CrateMarker generic type More info: https://ramaproxy.org/book/intro/patterns.html#working-around-the-orphan-rule-in-specific-cases

Note: RamaTryFrom has a blacket implementation for types that implement TryFrom

Required Associated Types

type Error

Required Methods

fn rama_try_from(value: T) -> Result<Self, Self::Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.