rama::crypto::dep::aws_lc_rs

Module encoding

Available on crate feature crypto only.

Serialization formats

Structs

Curve25519SeedBin

Serialized bytes

EcPrivateKeyBin

Serialized bytes

EcPrivateKeyRfc5915Der

Serialized bytes

EcPublicKeyCompressedBin

Serialized bytes

EcPublicKeyUncompressedBin

Serialized bytes

Pkcs8V1Der

Serialized bytes

Pkcs8V2Der

Serialized bytes

PqdsaPrivateKeyRaw

Serialized bytes

PqdsaSeedRaw

Serialized bytes

PublicKeyX509Der

Serialized bytes

Traits

AsBigEndian

Trait for values that can be serialized into a big-endian format

AsDer

Trait for types that can be serialized into a DER format.

AsRawBytes

Trait for values that can be serialized into a raw format