rama::crypto::dep::rcgen

Function date_time_ymd

pub fn date_time_ymd(year: i32, month: u8, day: u8) -> OffsetDateTime

Available on crate feature crypto only.

Helper to obtain an OffsetDateTime from year, month, day values

The year, month, day values are assumed to be in UTC.

This helper function serves two purposes: first, so that you don’t have to import the time crate yourself in order to specify date information, second so that users don’t have to type unproportionately long code just to generate an instance of OffsetDateTime.