rama::crypto::dep

Module x509_parser

Available on crate feature crypto only.

Re-export of the x509_parser crate.

Modules

certificate

X.509 Certificate object definitions and operations

certification_request

cri_attributes

error

X.509 errors

extensions

X.509 Extensions objects and types

objects

X.509 helper objects definitions and registry

pem

Decoding functions for PEM-encoded data

prelude

A “prelude” for users of the x509-parser crate.

Functions

parse_crl_derDeprecated

Parse a DER-encoded X.509 v2 CRL, and return the remaining of the input and the built object.

parse_x509_certificate

Parse a DER-encoded X.509 Certificate, and return the remaining of the input and the built object.

parse_x509_crl

Parse a DER-encoded X.509 v2 CRL, and return the remaining of the input and the built object.

parse_x509_derDeprecated

Parse a DER-encoded X.509 Certificate, and return the remaining of the input and the built