rama::crypto::dep::x509_parser

Module error

Available on crate feature crypto only.

X.509 errors

Structs

NidError

An error that can occur while converting an OID to a Nid.

Enums

PEMError

An error that can occur while parsing or validating a certificate.

X509Error

An error that can occur while parsing or validating a certificate.

Type Aliases

X509Result

Holds the result of parsing functions (X.509)