rama::crypto::dep::x509_parser

Module prelude

Available on crate feature crypto only.

A “prelude” for users of the x509-parser crate.

Modules

asn1_rs

License: MIT Apache License 2.0 docs.rs crates.io Download numbers Github CI Minimum rustc version

der_parser

License: MIT Apache License 2.0 docs.rs crates.io Download numbers dependency status Github CI Minimum rustc version

oid_registry

License: MIT Apache License 2.0 docs.rs crates.io Github CI Minimum rustc version

public_key

revocation_list

signature_algorithm

signature_value

time

utils

verifyverify or verify-aws

visitor

x509

X.509 objects and types

Structs

ASN1Time

An ASN.1 timestamp.

AccessDescription

AccessDescription ::= SEQUENCE { accessMethod OBJECT IDENTIFIER, accessLocation GeneralName }

AlgorithmIdentifier

Algorithm identifier

AttributeTypeAndValue

A generic attribute type and value

AuthorityInfoAccess

AuthorityKeyIdentifier

BasicConstraints

Identifies whether the subject of the certificate is a CA, and the max validation depth.

BasicExtension

Basic extension structure, used in search results

CRLDistributionPoint

CRLDistributionPoints

CertificateRevocationList

An X.509 v2 Certificate Revocation List (CRL).

ChallengePassword

CtExtensions

CtExtensions as defined in RFC6962 Section 3.2

CtLogID

LogID as defined in RFC6962 Section 3.2

CtVersion

Certificate Transparency Version as defined in RFC6962 Section 3.2

DigitallySigned

ExtendedKeyUsage

ExtensionRequest

Section 3.1 of rfc 5272

GeneralSubtree

Represents the structure used in the name constraints extensions. The fields minimum and maximum are not supported (openssl also has no support).

InhibitAnyPolicy

IssuerAlternativeName

IssuingDistributionPoint

KeyIdentifier

KeyUsage

NSCertType

NameConstraints

NidError

An error that can occur while converting an OID to a Nid.

Pem

Representation of PEM data

PemIterator

Iterator over PEM-encapsulated blocks

PolicyConstraints

PolicyInformation

PolicyMapping

PolicyMappings

PolicyQualifierInfo

ReasonCode

ReasonFlags

RelativeDistinguishedName

A Relative Distinguished Name element.

RevokedCertificate

SignedCertificateTimestamp

SubjectAlternativeName

SubjectInfoAccess

SubjectPublicKeyInfo

TbsCertList

The sequence TBSCertList contains information about the certificates that have been revoked by the CA that issued the CRL.

TbsCertificate

The sequence TBSCertificate contains information associated with the subject of the certificate and the CA that issued it.

TbsCertificateParser

TbsCertificate parser builder

UniqueIdentifier

Validity

X509Certificate

An X.509 v3 Certificate.

X509CertificateParser

X.509 Certificate parser

X509CertificationRequest

Certification Signing Request (CSR)

X509CertificationRequestInfo

Certification Request Info structure

X509CriAttribute

Attributes for Certification Request

X509Extension

X.509 version 3 extension

X509ExtensionParser

X509Extension parser builder

X509Name

X.509 Name (as used in Issuer and Subject fields)

X509Version

The version of the encoded certificate.

Enums

DistributionPointName

GeneralName

Represents a GeneralName as defined in RFC5280. There is no support X.400 addresses and EDIPartyName.

PEMError

An error that can occur while parsing or validating a certificate.

ParsedCriAttribute

Attributes for Certification Request

ParsedExtension

X509Error

An error that can occur while parsing or validating a certificate.

Traits

FromDer

Base trait for DER object parsers

Functions

format_serial

Formats a slice to a colon-separated hex string (for ex 01:02:ff:ff)

oid2abbrev

Return the abbreviation (for ex. CN for Common Name), or if not found, the OID short name

oid2description

Returns the description corresponding to the OID

oid2sn

Returns the short name corresponding to the OID

oid_registry

Return a reference to the default registry of known OIDs

parse_crl_derDeprecated

Parse a DER-encoded X.509 v2 CRL, and return the remaining of the input and the built object.

parse_ct_signed_certificate_timestamp

Parses as single Signed Certificate Timestamp entry

parse_ct_signed_certificate_timestamp_list

Parses a list of Signed Certificate Timestamp entries

parse_x509_certificate

Parse a DER-encoded X.509 Certificate, and return the remaining of the input and the built object.

parse_x509_crl

Parse a DER-encoded X.509 v2 CRL, and return the remaining of the input and the built object.

parse_x509_derDeprecated

Parse a DER-encoded X.509 Certificate, and return the remaining of the input and the built

parse_x509_pem

Read a PEM-encoded structure, and decode the base64 data

pem_to_derDeprecated

Type Aliases

CertificatePolicies

X509Result

Holds the result of parsing functions (X.509)