rama::crypto::dep::x509_parser::prelude::asn1_rs::nom::branch

Trait Permutation

pub trait Permutation<I, O, E> {
    // Required method
    fn permutation(&mut self, input: I) -> Result<(I, O), Err<E>>;
}

Available on crate feature crypto only.

Helper trait for the permutation() combinator.

This trait is implemented for tuples of up to 21 elements

Required Methods

fn permutation(&mut self, input: I) -> Result<(I, O), Err<E>>

Tries to apply all parsers in the tuple in various orders until all of them succeed