rama::crypto::dep::x509_parser::prelude::asn1_rs::nom::lib::std::collections

Enum Bound
·

pub enum Bound<T> {
    Included(T),
    Excluded(T),
    Unbounded,
}

Available on crate features crypto and std only.

An endpoint of a range of keys.

Examples

Bounds are range endpoints:

use std::ops::Bound::*;
use std::ops::RangeBounds;

assert_eq!((..100).start_bound(), Unbounded);
assert_eq!((1..12).start_bound(), Included(&1));
assert_eq!((1..12).end_bound(), Excluded(&12));

Using a tuple of Bounds as an argument to BTreeMap::range. Note that in most cases, it’s better to use range syntax (1..5) instead.

use std::collections::BTreeMap;
use std::ops::Bound::{Excluded, Included, Unbounded};

let mut map = BTreeMap::new();
map.insert(3, "a");
map.insert(5, "b");
map.insert(8, "c");

for (key, value) in map.range((Excluded(3), Included(8))) {
    println!("{key}: {value}");
}

assert_eq!(Some((&3, &"a")), map.range((Unbounded, Included(5))).next());

Variants

Included(T)

An inclusive bound.

Excluded(T)

An exclusive bound.

Unbounded

An infinite endpoint. Indicates that there is no bound in this direction.

Implementations

impl<T> Bound<T>

pub fn as_ref(&self) -> Bound<&T>

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts from &Bound<T> to Bound<&T>.

pub const fn as_mut(&mut self) -> Bound<&mut T>

🔬This is a nightly-only experimental API. (bound_as_ref)

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts from &mut Bound<T> to Bound<&mut T>.

pub fn map<U, F>(self, f: F) -> Bound<U>
where F: FnOnce(T) -> U,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Maps a Bound<T> to a Bound<U> by applying a function to the contained value (including both Included and Excluded), returning a Bound of the same kind.

Examples

use std::ops::Bound::*;

let bound_string = Included("Hello, World!");

assert_eq!(bound_string.map(|s| s.len()), Included(13));

use std::ops::Bound;
use Bound::*;

let unbounded_string: Bound<String> = Unbounded;

assert_eq!(unbounded_string.map(|s| s.len()), Unbounded);

impl<T> Bound<&T>
where T: Copy,

pub fn copied(self) -> Bound<T>

🔬This is a nightly-only experimental API. (bound_copied)

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Map a Bound<&T> to a Bound<T> by copying the contents of the bound.

Examples

#![feature(bound_copied)]

use std::ops::Bound::*;
use std::ops::RangeBounds;

assert_eq!((1..12).start_bound(), Included(&1));
assert_eq!((1..12).start_bound().copied(), Included(1));

impl<T> Bound<&T>
where T: Clone,

pub fn cloned(self) -> Bound<T>
where T: Clone,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Map a Bound<&T> to a Bound<T> by cloning the contents of the bound.

Examples

use std::ops::Bound::*;
use std::ops::RangeBounds;

let a1 = String::from("a");
let (a2, a3, a4) = (a1.clone(), a1.clone(), a1.clone());

assert_eq!(Included(&a1), (a2..).start_bound());
assert_eq!(Included(a3), (a4..).start_bound().cloned());