rama::crypto::dep::x509_parser::prelude::asn1_rs::nom::lib::std::collections

Module linked_list
·

Available on crate feature crypto and crate feature std and non-no_global_oom_handling only.

A doubly-linked list with owned nodes.

The LinkedList allows pushing and popping elements at either end in constant time.

NOTE: It is almost always better to use Vec or VecDeque because array-based containers are generally faster, more memory efficient, and make better use of CPU cache.

Structs

ExtractIf

An iterator produced by calling extract_if on LinkedList.

IntoIter

An owning iterator over the elements of a LinkedList.

Iter

An iterator over the elements of a LinkedList.

IterMut

A mutable iterator over the elements of a LinkedList.

LinkedList

A doubly-linked list with owned nodes.

CursorExperimental

A cursor over a LinkedList.

CursorMutExperimental

A cursor over a LinkedList with editing operations.