rama::crypto::dep::x509_parser::prelude::asn1_rs::nom::lib::std::ops

Enum ControlFlow
·

pub enum ControlFlow<B, C = ()> {
    Continue(C),
    Break(B),
}

Available on crate features crypto and std only.

Used to tell an operation whether it should exit early or go on as usual.

This is used when exposing things (like graph traversals or visitors) where you want the user to be able to choose whether to exit early. Having the enum makes it clearer – no more wondering “wait, what did false mean again?” – and allows including a value.

Similar to Option and Result, this enum can be used with the ? operator to return immediately if the Break variant is present or otherwise continue normally with the value inside the Continue variant.

Examples

Early-exiting from Iterator::try_for_each:

use std::ops::ControlFlow;

let r = (2..100).try_for_each(|x| {
    if 403 % x == 0 {
        return ControlFlow::Break(x)
    }

    ControlFlow::Continue(())
});
assert_eq!(r, ControlFlow::Break(13));

A basic tree traversal:

use std::ops::ControlFlow;

pub struct TreeNode<T> {
    value: T,
    left: Option<Box<TreeNode<T>>>,
    right: Option<Box<TreeNode<T>>>,
}

impl<T> TreeNode<T> {
    pub fn traverse_inorder<B>(&self, f: &mut impl FnMut(&T) -> ControlFlow<B>) -> ControlFlow<B> {
        if let Some(left) = &self.left {
            left.traverse_inorder(f)?;
        }
        f(&self.value)?;
        if let Some(right) = &self.right {
            right.traverse_inorder(f)?;
        }
        ControlFlow::Continue(())
    }
    fn leaf(value: T) -> Option<Box<TreeNode<T>>> {
        Some(Box::new(Self { value, left: None, right: None }))
    }
}

let node = TreeNode {
    value: 0,
    left: TreeNode::leaf(1),
    right: Some(Box::new(TreeNode {
        value: -1,
        left: TreeNode::leaf(5),
        right: TreeNode::leaf(2),
    }))
};
let mut sum = 0;

let res = node.traverse_inorder(&mut |val| {
    if *val < 0 {
        ControlFlow::Break(*val)
    } else {
        sum += *val;
        ControlFlow::Continue(())
    }
});
assert_eq!(res, ControlFlow::Break(-1));
assert_eq!(sum, 6);

Variants

Continue(C)

Move on to the next phase of the operation as normal.

Break(B)

Exit the operation without running subsequent phases.

Implementations

impl<B, C> ControlFlow<B, C>

pub fn is_break(&self) -> bool

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Returns true if this is a Break variant.

Examples

use std::ops::ControlFlow;

assert!(ControlFlow::<&str, i32>::Break("Stop right there!").is_break());
assert!(!ControlFlow::<&str, i32>::Continue(3).is_break());

pub fn is_continue(&self) -> bool

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Returns true if this is a Continue variant.

Examples

use std::ops::ControlFlow;

assert!(!ControlFlow::<&str, i32>::Break("Stop right there!").is_continue());
assert!(ControlFlow::<&str, i32>::Continue(3).is_continue());

pub fn break_value(self) -> Option<B>

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts the ControlFlow into an Option which is Some if the ControlFlow was Break and None otherwise.

Examples

use std::ops::ControlFlow;

assert_eq!(ControlFlow::<&str, i32>::Break("Stop right there!").break_value(), Some("Stop right there!"));
assert_eq!(ControlFlow::<&str, i32>::Continue(3).break_value(), None);

pub const fn break_ok(self) -> Result<B, C>

🔬This is a nightly-only experimental API. (control_flow_ok)

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts the ControlFlow into an Result which is Ok if the ControlFlow was Break and Err if otherwise.

Examples

#![feature(control_flow_ok)]

use std::ops::ControlFlow;

struct TreeNode<T> {
    value: T,
    left: Option<Box<TreeNode<T>>>,
    right: Option<Box<TreeNode<T>>>,
}

impl<T> TreeNode<T> {
    fn find<'a>(&'a self, mut predicate: impl FnMut(&T) -> bool) -> Result<&'a T, ()> {
        let mut f = |t: &'a T| -> ControlFlow<&'a T> {
            if predicate(t) {
                ControlFlow::Break(t)
            } else {
                ControlFlow::Continue(())
            }
        };

        self.traverse_inorder(&mut f).break_ok()
    }

    fn traverse_inorder<'a, B>(
        &'a self,
        f: &mut impl FnMut(&'a T) -> ControlFlow<B>,
    ) -> ControlFlow<B> {
        if let Some(left) = &self.left {
            left.traverse_inorder(f)?;
        }
        f(&self.value)?;
        if let Some(right) = &self.right {
            right.traverse_inorder(f)?;
        }
        ControlFlow::Continue(())
    }

    fn leaf(value: T) -> Option<Box<TreeNode<T>>> {
        Some(Box::new(Self {
            value,
            left: None,
            right: None,
        }))
    }
}

let node = TreeNode {
    value: 0,
    left: TreeNode::leaf(1),
    right: Some(Box::new(TreeNode {
        value: -1,
        left: TreeNode::leaf(5),
        right: TreeNode::leaf(2),
    })),
};

let res = node.find(|val: &i32| *val > 3);
assert_eq!(res, Ok(&5));

pub fn map_break<T, F>(self, f: F) -> ControlFlow<T, C>
where F: FnOnce(B) -> T,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Maps ControlFlow<B, C> to ControlFlow<T, C> by applying a function to the break value in case it exists.

pub fn continue_value(self) -> Option<C>

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts the ControlFlow into an Option which is Some if the ControlFlow was Continue and None otherwise.

Examples

use std::ops::ControlFlow;

assert_eq!(ControlFlow::<&str, i32>::Break("Stop right there!").continue_value(), None);
assert_eq!(ControlFlow::<&str, i32>::Continue(3).continue_value(), Some(3));

pub const fn continue_ok(self) -> Result<C, B>

🔬This is a nightly-only experimental API. (control_flow_ok)

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Converts the ControlFlow into an Result which is Ok if the ControlFlow was Continue and Err if otherwise.

Examples

#![feature(control_flow_ok)]

use std::ops::ControlFlow;

struct TreeNode<T> {
    value: T,
    left: Option<Box<TreeNode<T>>>,
    right: Option<Box<TreeNode<T>>>,
}

impl<T> TreeNode<T> {
    fn validate<B>(&self, f: &mut impl FnMut(&T) -> ControlFlow<B>) -> Result<(), B> {
        self.traverse_inorder(f).continue_ok()
    }

    fn traverse_inorder<B>(&self, f: &mut impl FnMut(&T) -> ControlFlow<B>) -> ControlFlow<B> {
        if let Some(left) = &self.left {
            left.traverse_inorder(f)?;
        }
        f(&self.value)?;
        if let Some(right) = &self.right {
            right.traverse_inorder(f)?;
        }
        ControlFlow::Continue(())
    }

    fn leaf(value: T) -> Option<Box<TreeNode<T>>> {
        Some(Box::new(Self {
            value,
            left: None,
            right: None,
        }))
    }
}

let node = TreeNode {
    value: 0,
    left: TreeNode::leaf(1),
    right: Some(Box::new(TreeNode {
        value: -1,
        left: TreeNode::leaf(5),
        right: TreeNode::leaf(2),
    })),
};

let res = node.validate(&mut |val| {
    if *val < 0 {
        return ControlFlow::Break("negative value detected");
    }

    if *val > 4 {
        return ControlFlow::Break("too big value detected");
    }

    ControlFlow::Continue(())
});
assert_eq!(res, Err("too big value detected"));

pub fn map_continue<T, F>(self, f: F) -> ControlFlow<B, T>
where F: FnOnce(C) -> T,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Maps ControlFlow<B, C> to ControlFlow<B, T> by applying a function to the continue value in case it exists.

impl<T> ControlFlow<T, T>

pub const fn into_value(self) -> T

🔬This is a nightly-only experimental API. (control_flow_into_value)

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Extracts the value T that is wrapped by ControlFlow<T, T>.

Examples

#![feature(control_flow_into_value)]
use std::ops::ControlFlow;

assert_eq!(ControlFlow::<i32, i32>::Break(1024).into_value(), 1024);
assert_eq!(ControlFlow::<i32, i32>::Continue(512).into_value(), 512);

impl<R> ControlFlow<R, <R as Try>::Output>
where R: Try,

This impl block contains no items.

These are used only as part of implementing the iterator adapters. They have mediocre names and non-obvious semantics, so aren’t currently on a path to potential stabilization.