rama::crypto::dep::x509_parser::prelude::asn1_rs::nom

Module multi

Available on crate feature crypto only.

Combinators applying their child parser multiple times

Functions

countalloc

Runs the embedded parser count times, gathering the results in a Vec

fill

Runs the embedded parser repeatedly, filling the given slice with results.

fold_many0

Repeats the embedded parser, calling g to gather the results.

fold_many1

Repeats the embedded parser, calling g to gather the results.

fold_many_m_n

Repeats the embedded parser m..=n times, calling g to gather the results

length_countalloc

Gets a number from the first parser, then applies the second parser that many times.

length_data

Gets a number from the parser and returns a subslice of the input of that size. If the parser returns Incomplete, length_data will return an error.

length_value

Gets a number from the first parser, takes a subslice of the input of that size, then applies the second parser on that subslice. If the second parser returns Incomplete, length_value will return an error.

many0alloc

Repeats the embedded parser, gathering the results in a Vec.

many0_count

Repeats the embedded parser, counting the results

many1alloc

Runs the embedded parser, gathering the results in a Vec.

many1_count

Runs the embedded parser, counting the results.

many_m_nalloc

Repeats the embedded parser m..=n times

many_tillalloc

Applies the parser f until the parser g produces a result.

separated_list0alloc

Alternates between two parsers to produce a list of elements.

separated_list1alloc

Alternates between two parsers to produce a list of elements until Err::Error.