rama::crypto::dep::x509_parser::prelude::der_parser::num_bigint

Trait ToBigInt

pub trait ToBigInt {
    // Required method
    fn to_bigint(&self) -> Option<BigInt>;
}

Available on crate feature crypto only.

A generic trait for converting a value to a BigInt. This may return None when converting from f32 or f64, and will always succeed when converting from any integer or unsigned primitive, or BigUint.

Required Methods

fn to_bigint(&self) -> Option<BigInt>

Converts the value of self to a BigInt.