rama::crypto::dep::x509_parser::prelude

Struct X509Certificate

pub struct X509Certificate<'a> {
    pub tbs_certificate: TbsCertificate<'a>,
    pub signature_algorithm: AlgorithmIdentifier<'a>,
    pub signature_value: BitString<'a>,
    /* private fields */
}

Available on crate feature crypto only.

An X.509 v3 Certificate.

X.509 v3 certificates are defined in RFC5280, section 4.1. This object uses the same structure for content, so for ex the subject can be accessed using the path x509.tbs_certificate.subject.

X509Certificate also contains convenience methods to access the most common fields (subject, issuer, etc.). These are provided using Deref<Target = TbsCertificate>, so documentation for these methods can be found in the TbsCertificate object.

A X509Certificate is a zero-copy view over a buffer, so the lifetime is the same as the buffer containing the binary representation.

fn display_x509_info(x509: &X509Certificate<'_>) {
     let subject = x509.subject();
     let issuer = x509.issuer();
     println!("X.509 Subject: {}", subject);
     println!("X.509 Issuer: {}", issuer);
     println!("X.509 serial: {}", x509.tbs_certificate.raw_serial_as_string());
}

Fields
tbs_certificate: TbsCertificate<'a>signature_algorithm: AlgorithmIdentifier<'a>signature_value: BitString<'a>
Implementations

impl<'a> X509Certificate<'a>

pub fn as_raw(&self) -> &'a [u8]

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Return the raw ASN.1 DER content of the complete signed certificate that was parsed.

This includes the to-be-signed (TBS) certificate, the signature algorithm, and the signature. If you want just the ASN.1 DER of the TBS certificate, prefer TbsCertificate::as_ref().

We avoid the AsRef trait in this instance to ensure the full lifetime of the X509Certificate is used.

pub fn verify_signature( &self, public_key: Option<&SubjectPublicKeyInfo<'_>>, ) -> Result<(), X509Error>

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature verify only.

Verify the cryptographic signature of this certificate

public_key is the public key of the signer. For a self-signed certificate, (for ex. a public root certificate authority), this is the key from the certificate, so you can use None.

For a leaf certificate, this is the public key of the certificate that signed it. It is usually an intermediate authority.

Not all algorithms are supported, this function is limited to what ring supports.

impl X509Certificate<'_>

pub fn walk<V>(&self, visitor: &mut V)
where V: X509CertificateVisitor,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Run the provided X509CertificateVisitor over the X.509 Certificate (self)