rama::dns::hickory::resolver

Crate proto

Available on crate feature dns only.

Hickory DNS Protocol library

Modules

op

Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used together to either query or update resource records sets.

rr

Resource record related components, e.g. Name aka label, Record, RData, …

runtimestd

Abstractions to deal with different async runtimes.

serialize

Contains serialization libraries for binary and text, txt.

tcpstd

TCP protocol related components for DNS

udpstd

UDP protocol related components for DNS

xfer

DNS high level transit implementations.

Structs

BufDnsStreamHandlestd

A buffering stream bound to a SocketAddr

DnsMultiplexer

A DNS Client implemented over futures-rs.

ForwardData

Data needed to process a SOA-record-based referral.

ForwardNSData

Data needed to process a NS-record-based referral.

ProtoError

The error type for errors that get returned in the crate

RetryDnsHandle

Can be used to reattempt queries if they fail

Enums

ProtoErrorKind

The error kind for errors that get returned in the crate

Traits

DnsHandle

A trait for implementing high level functions of DNS.

DnsStreamHandle

Implementations of Sinks for sending DNS messages