rama::dns::hickory::resolver::proto

Enum ProtoErrorKind

#[non_exhaustive]
pub enum ProtoErrorKind {

    BadQueryCount(usize),
    Busy,
    Canceled(Canceled),
    CharacterDataTooLong {
        max: usize,
        len: usize,
    },
    LabelOverlapsWithOther {
        label: usize,
        other: usize,
    },
    DnsKeyProtocolNot3(u8),
    DomainNameTooLong(usize),
    EdnsNameNotRoot(Name),
    FormError {
        header: Header,
        error: Box<ProtoError>,
    },
    IncorrectRDataLengthRead {
        read: usize,
        len: usize,
    },
    LabelBytesTooLong(usize),
    PointerNotPriorToLabel {
        idx: usize,
        ptr: u16,
    },
    MaxBufferSizeExceeded(usize),
    MaxRecordLimitExceeded {
        count: usize,
        record_type: RecordType,
    },
    Message(&'static str),
    Msg(String),
    NoConnections,
    NoError,
    NotAllRecordsWritten {
        count: usize,
    },
    NoRecordsFound {
        query: Box<Query>,
        soa: Option<Box<Record<SOA>>>,
        ns: Option<Arc<[ForwardNSData]>>,
        negative_ttl: Option<u32>,
        response_code: ResponseCode,
        trusted: bool,
        authorities: Option<Arc<[Record]>>,
    },
    UnknownAlgorithmTypeValue(u8),
    UnknownDigestTypeValue(u8),
    UnknownDnsClassStr(String),
    UnknownDnsClassValue(u16),
    UnknownRecordTypeStr(String),
    UnknownRecordTypeValue(u16),
    UnrecognizedLabelCode(u8),
    UnrecognizedNsec3Flags(u8),
    UnrecognizedCsyncFlags(u16),
    Io(Arc<Error>),
    Poisoned,
    RequestRefused,
    Timer,
    Timeout,
    UrlParsing(ParseError),
    Utf8(Utf8Error),
    FromUtf8(FromUtf8Error),
    ParseInt(ParseIntError),
    QueryCaseMismatch,

}

Available on crate feature dns only.

The error kind for errors that get returned in the crate

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BadQueryCount(usize)

Query count is not one

Busy

The underlying resource is too busy

This is a signal that an internal resource is too busy. The intended action should be tried again, ideally after waiting for a little while for the situation to improve. Alternatively, the action could be tried on another resource (for example, in a name server pool).

Canceled(Canceled)

An error caused by a canceled future

CharacterDataTooLong

Character data length exceeded the limit

Fields

max: usize

Specified maximum

len: usize

Actual length

LabelOverlapsWithOther

Overlapping labels

Fields

label: usize

Start of the label that is overlaps

other: usize

Start of the other label

DnsKeyProtocolNot3(u8)

DNS protocol version doesn’t have the expected version 3

DomainNameTooLong(usize)

A domain name was too long

EdnsNameNotRoot(Name)

EDNS resource record label is not the root label, although required

FormError

Format error in Message Parsing

Fields

header: Header

Header of the bad Message

error: Box<ProtoError>

Error that occurred while parsing the Message

IncorrectRDataLengthRead

The length of rdata read was not as expected

Fields

read: usize

The amount of read data

len: usize

The expected length of the data

LabelBytesTooLong(usize)

Label bytes exceeded the limit of 63

PointerNotPriorToLabel

Label bytes exceeded the limit of 63

Fields

idx: usize

index of the label containing this pointer

ptr: u16

location to which the pointer is directing

MaxBufferSizeExceeded(usize)

The maximum buffer size was exceeded

MaxRecordLimitExceeded

Maximum record limit was exceeded

Fields

count: usize

number of records

record_type: RecordType

The record type that triggered the error.

Message(&'static str)

An error with an arbitrary message, referenced as &’static str

Msg(String)

An error with an arbitrary message, stored as String

NoConnections

No resolvers available

NoError

No error was specified

NotAllRecordsWritten

Not all records were able to be written

Fields

count: usize

Number of records that were written before the error

NoRecordsFound

No records were found for a query

Fields

query: Box<Query>

The query for which no records were found.

soa: Option<Box<Record<SOA>>>

If an SOA is present, then this is an authoritative response or a referral to another nameserver, see the negative_type field.

ns: Option<Arc<[ForwardNSData]>>

Nameservers may be present in addition to or in lieu of an SOA for a referral The tuple struct layout is vec[(Nameserver, [vec of glue records])]

negative_ttl: Option<u32>

negative ttl, as determined from DnsResponse::negative_ttl this will only be present if the SOA was also present.

response_code: ResponseCode

ResponseCode, if NXDOMAIN, the domain does not exist (and no other types). If NoError, then the domain exists but there exist either other types at the same label, or subzones of that label.

trusted: bool

If we trust NXDOMAIN errors from this server

authorities: Option<Arc<[Record]>>

Authority records from the query. These are important to preserve for DNSSEC validation.

UnknownAlgorithmTypeValue(u8)

An unknown algorithm type was found

UnknownDigestTypeValue(u8)

An unknown digest type was found

UnknownDnsClassStr(String)

An unknown dns class was found

UnknownDnsClassValue(u16)

An unknown dns class value was found

UnknownRecordTypeStr(String)

An unknown record type string was found

UnknownRecordTypeValue(u16)

An unknown record type value was found

UnrecognizedLabelCode(u8)

An unrecognized label code was found

UnrecognizedNsec3Flags(u8)

Unrecognized nsec3 flags were found

UnrecognizedCsyncFlags(u16)

Unrecognized csync flags were found

Io(Arc<Error>)

Available on crate feature std only.

An error got returned from IO

Poisoned

Any sync poised error

RequestRefused

A request was Refused due to some access check

Timer

A tokio timer error

Timeout

A request timed out

UrlParsing(ParseError)

An url parsing error

Utf8(Utf8Error)

A utf8 parsing error

FromUtf8(FromUtf8Error)

A utf8 parsing error

ParseInt(ParseIntError)

An int parsing error

QueryCaseMismatch

Case randomization is enabled, and a server did not echo a query name back with the same case.

Implementations

impl ProtoErrorKind

pub fn is_bad_query_count(&self) -> bool

Returns true if this is a ProtoErrorKind::BadQueryCount, otherwise false

pub fn as_bad_query_count_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::BadQueryCount, otherwise None

pub fn as_bad_query_count(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::BadQueryCount, otherwise None

pub fn into_bad_query_count(self) -> Result<usize, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::BadQueryCount, otherwise returns back the enum in the Err case of the result

pub fn is_busy(&self) -> bool

Returns true if this is a ProtoErrorKind::Busy, otherwise false

pub fn is_canceled(&self) -> bool

Returns true if this is a ProtoErrorKind::Canceled, otherwise false

pub fn as_canceled_mut(&mut self) -> Option<&mut Canceled>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Canceled, otherwise None

pub fn as_canceled(&self) -> Option<&Canceled>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Canceled, otherwise None

pub fn into_canceled(self) -> Result<Canceled, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::Canceled, otherwise returns back the enum in the Err case of the result

pub fn is_character_data_too_long(&self) -> bool

Returns true if this is a ProtoErrorKind::CharacterDataTooLong, otherwise false

pub fn as_character_data_too_long_mut( &mut self, ) -> Option<(&mut usize, &mut usize)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise None

pub fn as_character_data_too_long(&self) -> Option<(&usize, &usize)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise None

pub fn into_character_data_too_long( self, ) -> Result<(usize, usize), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::CharacterDataTooLong, otherwise returns back the enum in the Err case of the result

pub fn is_label_overlaps_with_other(&self) -> bool

Returns true if this is a ProtoErrorKind::LabelOverlapsWithOther, otherwise false

pub fn as_label_overlaps_with_other_mut( &mut self, ) -> Option<(&mut usize, &mut usize)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::LabelOverlapsWithOther, otherwise None

pub fn as_label_overlaps_with_other(&self) -> Option<(&usize, &usize)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::LabelOverlapsWithOther, otherwise None

pub fn into_label_overlaps_with_other( self, ) -> Result<(usize, usize), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::LabelOverlapsWithOther, otherwise returns back the enum in the Err case of the result

pub fn is_dns_key_protocol_not3(&self) -> bool

Returns true if this is a ProtoErrorKind::DnsKeyProtocolNot3, otherwise false

pub fn as_dns_key_protocol_not3_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::DnsKeyProtocolNot3, otherwise None

pub fn as_dns_key_protocol_not3(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a ProtoErrorKind::DnsKeyProtocolNot3, otherwise None

pub fn into_dns_key_protocol_not3(self) -> Result<u8, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::DnsKeyProtocolNot3, otherwise returns back the enum in the Err case of the result

pub fn is_domain_name_too_long(&self) -> bool

Returns true if this is a ProtoErrorKind::DomainNameTooLong, otherwise false

pub fn as_domain_name_too_long_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::DomainNameTooLong, otherwise None

pub fn as_domain_name_too_long(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::DomainNameTooLong, otherwise None

pub fn into_domain_name_too_long(self) -> Result<usize, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::DomainNameTooLong, otherwise returns back the enum in the Err case of the result

pub fn is_edns_name_not_root(&self) -> bool

Returns true if this is a ProtoErrorKind::EdnsNameNotRoot, otherwise false

pub fn as_edns_name_not_root_mut(&mut self) -> Option<&mut Name>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::EdnsNameNotRoot, otherwise None

pub fn as_edns_name_not_root(&self) -> Option<&Name>

Optionally returns references to the inner fields if this is a ProtoErrorKind::EdnsNameNotRoot, otherwise None

pub fn into_edns_name_not_root(self) -> Result<Name, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::EdnsNameNotRoot, otherwise returns back the enum in the Err case of the result

pub fn is_form_error(&self) -> bool

Returns true if this is a ProtoErrorKind::FormError, otherwise false

pub fn as_form_error_mut( &mut self, ) -> Option<(&mut Header, &mut Box<ProtoError>)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::FormError, otherwise None

pub fn as_form_error(&self) -> Option<(&Header, &Box<ProtoError>)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::FormError, otherwise None

pub fn into_form_error( self, ) -> Result<(Header, Box<ProtoError>), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::FormError, otherwise returns back the enum in the Err case of the result

pub fn is_incorrect_r_data_length_read(&self) -> bool

Returns true if this is a ProtoErrorKind::IncorrectRDataLengthRead, otherwise false

pub fn as_incorrect_r_data_length_read_mut( &mut self, ) -> Option<(&mut usize, &mut usize)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::IncorrectRDataLengthRead, otherwise None

pub fn as_incorrect_r_data_length_read(&self) -> Option<(&usize, &usize)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::IncorrectRDataLengthRead, otherwise None

pub fn into_incorrect_r_data_length_read( self, ) -> Result<(usize, usize), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::IncorrectRDataLengthRead, otherwise returns back the enum in the Err case of the result

pub fn is_label_bytes_too_long(&self) -> bool

Returns true if this is a ProtoErrorKind::LabelBytesTooLong, otherwise false

pub fn as_label_bytes_too_long_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::LabelBytesTooLong, otherwise None

pub fn as_label_bytes_too_long(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::LabelBytesTooLong, otherwise None

pub fn into_label_bytes_too_long(self) -> Result<usize, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::LabelBytesTooLong, otherwise returns back the enum in the Err case of the result

pub fn is_pointer_not_prior_to_label(&self) -> bool

Returns true if this is a ProtoErrorKind::PointerNotPriorToLabel, otherwise false

pub fn as_pointer_not_prior_to_label_mut( &mut self, ) -> Option<(&mut usize, &mut u16)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::PointerNotPriorToLabel, otherwise None

pub fn as_pointer_not_prior_to_label(&self) -> Option<(&usize, &u16)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::PointerNotPriorToLabel, otherwise None

pub fn into_pointer_not_prior_to_label( self, ) -> Result<(usize, u16), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::PointerNotPriorToLabel, otherwise returns back the enum in the Err case of the result

pub fn is_max_buffer_size_exceeded(&self) -> bool

Returns true if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise false

pub fn as_max_buffer_size_exceeded_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise None

pub fn as_max_buffer_size_exceeded(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise None

pub fn into_max_buffer_size_exceeded(self) -> Result<usize, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::MaxBufferSizeExceeded, otherwise returns back the enum in the Err case of the result

pub fn is_max_record_limit_exceeded(&self) -> bool

Returns true if this is a ProtoErrorKind::MaxRecordLimitExceeded, otherwise false

pub fn as_max_record_limit_exceeded_mut( &mut self, ) -> Option<(&mut usize, &mut RecordType)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::MaxRecordLimitExceeded, otherwise None

pub fn as_max_record_limit_exceeded(&self) -> Option<(&usize, &RecordType)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::MaxRecordLimitExceeded, otherwise None

pub fn into_max_record_limit_exceeded( self, ) -> Result<(usize, RecordType), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::MaxRecordLimitExceeded, otherwise returns back the enum in the Err case of the result

pub fn is_message(&self) -> bool

Returns true if this is a ProtoErrorKind::Message, otherwise false

pub fn as_message_mut(&mut self) -> Option<&mut &'static str>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Message, otherwise None

pub fn as_message(&self) -> Option<&&'static str>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Message, otherwise None

pub fn into_message(self) -> Result<&'static str, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::Message, otherwise returns back the enum in the Err case of the result

pub fn is_msg(&self) -> bool

Returns true if this is a ProtoErrorKind::Msg, otherwise false

pub fn as_msg_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Msg, otherwise None

pub fn as_msg(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Msg, otherwise None

pub fn into_msg(self) -> Result<String, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::Msg, otherwise returns back the enum in the Err case of the result

pub fn is_no_connections(&self) -> bool

Returns true if this is a ProtoErrorKind::NoConnections, otherwise false

pub fn is_no_error(&self) -> bool

Returns true if this is a ProtoErrorKind::NoError, otherwise false

pub fn is_not_all_records_written(&self) -> bool

Returns true if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise false

pub fn as_not_all_records_written_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise None

pub fn as_not_all_records_written(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise None

pub fn into_not_all_records_written(self) -> Result<usize, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::NotAllRecordsWritten, otherwise returns back the enum in the Err case of the result

pub fn is_no_records_found(&self) -> bool

Returns true if this is a ProtoErrorKind::NoRecordsFound, otherwise false

pub fn as_no_records_found_mut( &mut self, ) -> Option<(&mut Box<Query>, &mut Option<Box<Record<SOA>>>, &mut Option<Arc<[ForwardNSData]>>, &mut Option<u32>, &mut ResponseCode, &mut bool, &mut Option<Arc<[Record]>>)>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::NoRecordsFound, otherwise None

pub fn as_no_records_found( &self, ) -> Option<(&Box<Query>, &Option<Box<Record<SOA>>>, &Option<Arc<[ForwardNSData]>>, &Option<u32>, &ResponseCode, &bool, &Option<Arc<[Record]>>)>

Optionally returns references to the inner fields if this is a ProtoErrorKind::NoRecordsFound, otherwise None

pub fn into_no_records_found( self, ) -> Result<(Box<Query>, Option<Box<Record<SOA>>>, Option<Arc<[ForwardNSData]>>, Option<u32>, ResponseCode, bool, Option<Arc<[Record]>>), ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::NoRecordsFound, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_algorithm_type_value(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownAlgorithmTypeValue, otherwise false

pub fn as_unknown_algorithm_type_value_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownAlgorithmTypeValue, otherwise None

pub fn as_unknown_algorithm_type_value(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownAlgorithmTypeValue, otherwise None

pub fn into_unknown_algorithm_type_value(self) -> Result<u8, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownAlgorithmTypeValue, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_digest_type_value(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownDigestTypeValue, otherwise false

pub fn as_unknown_digest_type_value_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownDigestTypeValue, otherwise None

pub fn as_unknown_digest_type_value(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownDigestTypeValue, otherwise None

pub fn into_unknown_digest_type_value(self) -> Result<u8, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownDigestTypeValue, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_dns_class_str(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownDnsClassStr, otherwise false

pub fn as_unknown_dns_class_str_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownDnsClassStr, otherwise None

pub fn as_unknown_dns_class_str(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownDnsClassStr, otherwise None

pub fn into_unknown_dns_class_str(self) -> Result<String, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownDnsClassStr, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_dns_class_value(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownDnsClassValue, otherwise false

pub fn as_unknown_dns_class_value_mut(&mut self) -> Option<&mut u16>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownDnsClassValue, otherwise None

pub fn as_unknown_dns_class_value(&self) -> Option<&u16>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownDnsClassValue, otherwise None

pub fn into_unknown_dns_class_value(self) -> Result<u16, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownDnsClassValue, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_record_type_str(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownRecordTypeStr, otherwise false

pub fn as_unknown_record_type_str_mut(&mut self) -> Option<&mut String>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownRecordTypeStr, otherwise None

pub fn as_unknown_record_type_str(&self) -> Option<&String>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownRecordTypeStr, otherwise None

pub fn into_unknown_record_type_str(self) -> Result<String, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownRecordTypeStr, otherwise returns back the enum in the Err case of the result

pub fn is_unknown_record_type_value(&self) -> bool

Returns true if this is a ProtoErrorKind::UnknownRecordTypeValue, otherwise false

pub fn as_unknown_record_type_value_mut(&mut self) -> Option<&mut u16>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnknownRecordTypeValue, otherwise None

pub fn as_unknown_record_type_value(&self) -> Option<&u16>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnknownRecordTypeValue, otherwise None

pub fn into_unknown_record_type_value(self) -> Result<u16, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnknownRecordTypeValue, otherwise returns back the enum in the Err case of the result

pub fn is_unrecognized_label_code(&self) -> bool

Returns true if this is a ProtoErrorKind::UnrecognizedLabelCode, otherwise false

pub fn as_unrecognized_label_code_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnrecognizedLabelCode, otherwise None

pub fn as_unrecognized_label_code(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnrecognizedLabelCode, otherwise None

pub fn into_unrecognized_label_code(self) -> Result<u8, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnrecognizedLabelCode, otherwise returns back the enum in the Err case of the result

pub fn is_unrecognized_nsec3_flags(&self) -> bool

Returns true if this is a ProtoErrorKind::UnrecognizedNsec3Flags, otherwise false

pub fn as_unrecognized_nsec3_flags_mut(&mut self) -> Option<&mut u8>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnrecognizedNsec3Flags, otherwise None

pub fn as_unrecognized_nsec3_flags(&self) -> Option<&u8>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnrecognizedNsec3Flags, otherwise None

pub fn into_unrecognized_nsec3_flags(self) -> Result<u8, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnrecognizedNsec3Flags, otherwise returns back the enum in the Err case of the result

pub fn is_unrecognized_csync_flags(&self) -> bool

Returns true if this is a ProtoErrorKind::UnrecognizedCsyncFlags, otherwise false

pub fn as_unrecognized_csync_flags_mut(&mut self) -> Option<&mut u16>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UnrecognizedCsyncFlags, otherwise None

pub fn as_unrecognized_csync_flags(&self) -> Option<&u16>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UnrecognizedCsyncFlags, otherwise None

pub fn into_unrecognized_csync_flags(self) -> Result<u16, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UnrecognizedCsyncFlags, otherwise returns back the enum in the Err case of the result

pub fn is_io(&self) -> bool

Returns true if this is a ProtoErrorKind::Io, otherwise false

pub fn as_io_mut(&mut self) -> Option<&mut Arc<Error>>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Io, otherwise None

pub fn as_io(&self) -> Option<&Arc<Error>>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Io, otherwise None

pub fn into_io(self) -> Result<Arc<Error>, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::Io, otherwise returns back the enum in the Err case of the result

pub fn is_poisoned(&self) -> bool

Returns true if this is a ProtoErrorKind::Poisoned, otherwise false

pub fn is_request_refused(&self) -> bool

Returns true if this is a ProtoErrorKind::RequestRefused, otherwise false

pub fn is_timer(&self) -> bool

Returns true if this is a ProtoErrorKind::Timer, otherwise false

pub fn is_timeout(&self) -> bool

Returns true if this is a ProtoErrorKind::Timeout, otherwise false

pub fn is_url_parsing(&self) -> bool

Returns true if this is a ProtoErrorKind::UrlParsing, otherwise false

pub fn as_url_parsing_mut(&mut self) -> Option<&mut ParseError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise None

pub fn as_url_parsing(&self) -> Option<&ParseError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise None

pub fn into_url_parsing(self) -> Result<ParseError, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::UrlParsing, otherwise returns back the enum in the Err case of the result

pub fn is_utf8(&self) -> bool

Returns true if this is a ProtoErrorKind::Utf8, otherwise false

pub fn as_utf8_mut(&mut self) -> Option<&mut Utf8Error>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::Utf8, otherwise None

pub fn as_utf8(&self) -> Option<&Utf8Error>

Optionally returns references to the inner fields if this is a ProtoErrorKind::Utf8, otherwise None

pub fn into_utf8(self) -> Result<Utf8Error, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::Utf8, otherwise returns back the enum in the Err case of the result

pub fn is_from_utf8(&self) -> bool

Returns true if this is a ProtoErrorKind::FromUtf8, otherwise false

pub fn as_from_utf8_mut(&mut self) -> Option<&mut FromUtf8Error>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise None

pub fn as_from_utf8(&self) -> Option<&FromUtf8Error>

Optionally returns references to the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise None

pub fn into_from_utf8(self) -> Result<FromUtf8Error, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::FromUtf8, otherwise returns back the enum in the Err case of the result

pub fn is_parse_int(&self) -> bool

Returns true if this is a ProtoErrorKind::ParseInt, otherwise false

pub fn as_parse_int_mut(&mut self) -> Option<&mut ParseIntError>

Optionally returns mutable references to the inner fields if this is a ProtoErrorKind::ParseInt, otherwise None

pub fn as_parse_int(&self) -> Option<&ParseIntError>

Optionally returns references to the inner fields if this is a ProtoErrorKind::ParseInt, otherwise None

pub fn into_parse_int(self) -> Result<ParseIntError, ProtoErrorKind>

Returns the inner fields if this is a ProtoErrorKind::ParseInt, otherwise returns back the enum in the Err case of the result

pub fn is_query_case_mismatch(&self) -> bool

Returns true if this is a ProtoErrorKind::QueryCaseMismatch, otherwise false