rama::dns::hickory::resolver::proto

Module op

Available on crate feature dns only.

Operations to send with a Client or server, e.g. Query, Message, or UpdateMessage can be used together to either query or update resource records sets.

Modules

header

Message metadata

message

Basic protocol message for DNS

op_code

Operation code for queries, updates, and responses

query

Query struct for looking up resource records

response_code

All defined response codes in DNS

update_message

Update related operations for Messages

Structs

Edns

Edns implements the higher level concepts for working with extended dns as it is used to create or be created from OPT record data.

EdnsFlags

EDNS flags

Header

Metadata for the Message struct.

LowerQuery

Identical to crate::op::Query, except that the Name is guaranteed to be in lower case form

Message

The basic request and response data structure, used for all DNS protocols.

MessageParts

Consumes Message giving public access to fields in Message so they can be destructured and taken by value

Query

Query struct for looking up resource records, basically a resource record without RDATA.

Enums

MessageType

Message types are either Query (also Update) or Response

OpCode

Operation code for queries, updates, and responses

ResponseCode

The status code of the response to a query.

Traits

MessageFinalizer

A trait for performing final amendments to a Message before it is sent.

UpdateMessage

To reduce errors in using the Message struct as an Update, this will do the call throughs to properly do that.

Type Aliases

MessageVerifier

Alias for a function verifying if a message is properly signed