rama::dns::hickory::resolver::proto::op

Module update_message

Available on crate feature dns only.

Update related operations for Messages

Constants

MAX_PAYLOAD_LEN

Maximum payload length for EDNS update messages

Traits

UpdateMessage

To reduce errors in using the Message struct as an Update, this will do the call throughs to properly do that.

Functions

appendstd or no-std-rand

Appends a record to an existing rrset, optionally require the rrset to exist (atomicity depends on the server)

compare_and_swapstd or no-std-rand

Compares and if it matches, swaps it for the new value (atomicity depends on the server)

createstd or no-std-rand

Sends a record to create on the server, this will fail if the record exists (atomicity depends on the server)

delete_allstd or no-std-rand

Deletes all rrsets at the specified name

delete_by_rdatastd or no-std-rand

Deletes a record (by rdata) from an rrset

delete_rrsetstd or no-std-rand

Deletes an entire rrset

zone_transferstd or no-std-rand

Download all records from a zone, or all records modified since given SOA was observed. The request will either be a AXFR Query (ask for full zone transfer) if a SOA was not provided, or a IXFR Query (incremental zone transfer) if a SOA was provided.