rama::dns::hickory::resolver::proto::op::update_message

Function create

pub fn create(rrset: RecordSet, zone_origin: Name, use_edns: bool) -> Message

Available on crate feature dns and (crate features std or no-std-rand) only.

Sends a record to create on the server, this will fail if the record exists (atomicity depends on the server)

RFC 2136, DNS Update, April 1997

 2.4.3 - RRset Does Not Exist

  No RRs with a specified NAME and TYPE (in the zone and class denoted
  by the Zone Section) can exist.

  For this prerequisite, a requestor adds to the section a single RR
  whose NAME and TYPE are equal to that of the RRset whose nonexistence
  is required.  The RDLENGTH of this record is zero (0), and RDATA
  field is therefore empty.  CLASS must be specified as NONE in order
  to distinguish this condition from a valid RR whose RDLENGTH is
  naturally zero (0) (for example, the NULL RR).  TTL must be specified
  as zero (0).

2.5.1 - Add To An RRset

   RRs are added to the Update Section whose NAME, TYPE, TTL, RDLENGTH
   and RDATA are those being added, and CLASS is the same as the zone
   class.  Any duplicate RRs will be silently ignored by the Primary
   Zone Server.

Arguments

rrset - the record(s) to create

zone_origin - the zone name to update, i.e. SOA name

use_edns - if true, EDNS options will be added to the request

The update must go to a zone authority (i.e. the server used in the ClientConnection)