rama::dns::hickory::resolver::proto

Module rr

Available on crate feature dns only.

Resource record related components, e.g. Name aka label, Record, RData, …

Modules

dns_class

class of DNS operations, in general always IN for internet

domain

Domain name associated types, such as Name and Label.

rdata

All record data structures and related serialization methods

record_data

record data enum variants

record_type

record type definitions

resource

resource record implementation

serial_number

Number type to support Serial Number Arithmetics

Structs

LowerName

TODO: all LowerNames should be stored in a global “intern” space, and then everything that uses them should be through references. As a workaround the Strings are all Rc as well as the array

Name

A domain name

Record

Resource records are storage value in DNS, into which all key/value pair data is stored.

RecordSet

Set of resource records associated to a name and type

RrKey

Accessor key for RRSets in the Authority.

SerialNumber

Wrapper type to support Serial Number Arithmetics as defined in RFC 1982. The signaure fields (expireation, inception) defined in RFC 4034, section 3.1.5 are serial numbers.

Enums

DNSClass

The DNS Record class

RData

Record data enum variants for all valid DNS data types.

RecordType

The type of the resource record.

RrsetRecords

An iterator over the RecordSet data

Traits

IntoName

Conversion into a Name

IntoRecordSetDeprecated

Types which implement this can be converted into a RecordSet

RecordData

RecordData that is stored in a DNS Record.