rama::dns::hickory::resolver::proto::rr

Module rdata

Available on crate feature dns only.

All record data structures and related serialization methods

Modules

a

IPv4 address record data

aaaa

IPv6 address record data

caa

allows a DNS domain name holder to specify one or more Certification Authorities (CAs) authorized to issue certificates for that domain.

cert

CERT record type for storing certificates in DNS

csync

CSYNC record for synchronizing data from a child zone to the parent

hinfo

HINFO record for storing host information

https

HTTPS type and related implementations

mx

mail exchange, email, record

name

Record type for all cname like records.

naptr

Dynamic Delegation Discovery System

null

null record type, generally not used except as an internal tool for representing null data

openpgpkey

OPENPGPKEY records for OpenPGP public keys

opt

option record for passing protocol options between the client and server

soa

start of authority record defining ownership and defaults for the zone

srv

service records for identify port mapping for specific services on a host

sshfp

SSHFP records for SSH public key fingerprints

svcb

SVCB records, see RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

tlsa

TLSA records for storing TLS certificate validation information

txt

text records for storing arbitrary data

Structs

A

The DNS A record type, an IPv4 address

AAAA

The DNS AAAA record type, an IPv6 address

ANAME

new type for the RecordData of ANAME

CAA

The CAA RR Type

CERT

RFC 4398, Storing Certificates in DNS, November 1987

CNAME

new type for the RecordData of CNAME

CSYNC

RFC 7477, Child-to-Parent Synchronization in DNS, March 2015

HINFO

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

HTTPS

HTTPS is really a derivation of the original SVCB record data. See SVCB for more documentation

MX

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

NAPTR

RFC 3403 DDDS DNS Database, October 2002

NS

new type for the RecordData of NS

NULL

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

OPENPGPKEY

RFC 7929

OPT

The OPT record type is used for ExtendedDNS records.

PTR

new type for the RecordData of PTR

SOA

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987

SRV

RFC 2782, DNS SRV RR, February 2000

SSHFP

RFC 4255

SVCB

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

TLSA

RFC 6698, DNS-Based Authentication for TLS

TXT

RFC 1035, DOMAIN NAMES - IMPLEMENTATION AND SPECIFICATION, November 1987