rama::dns::hickory::resolver::proto::rr::rdata

Module svcb

Available on crate feature dns only.

SVCB records, see RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

Structs

Alpn

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

EchConfigList

draft-ietf-tls-svcb-ech-01 Bootstrapping TLS Encrypted ClientHello with DNS Service Bindings, Sep 2024

IpHint

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

Mandatory

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

SVCB

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

Unknown

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

Enums

SvcParamKey

RFC 9460 SVCB and HTTPS Resource Records, Nov 2023

SvcParamValue

Warning, it is currently up to users of this type to validate the data against that expected by the key