rama::dns::hickory::resolver::proto

Module runtime

Available on crate features dns and std only.

Abstractions to deal with different async runtimes.

Structs

TokioHandle

A handle to the Tokio runtime

TokioRuntimeProvider

The Tokio Runtime for async execution

TokioTimetokio

New type which is implemented using tokio::time::{Delay, Timeout}

Traits

Executor

Generic executor.

QuicSocketBinderNon-__quic

Noop trait for when the quinn dependency is not available.

RuntimeProvider

RuntimeProvider defines which async runtime that handles IO and timers.

Spawn

A type defines the Handle which can spawn future.

Time

Generic Time for Delay and Timeout.

Functions

spawn_bgtokio

Spawn a background task, if it was present