rama::dns::hickory::resolver::proto::serialize

Module binary

Available on crate feature dns only.

Binary serialization types

Structs

BinDecoder

This is non-destructive to the inner buffer, b/c for pointer types we need to perform a reverse seek to lookup names

BinEncoder

Encode DNS messages and resource record types.

Restrict

Untrusted types will be wrapped in this type.

Verified

Verified data that can be operated on

Enums

DecodeError

An error that can occur deep in a decoder This type is kept very small so that function that use it inline often

EncodeMode

In the Verify mode there maybe some things which are encoded differently, e.g. SIG0 records should not be included in the additional count and not in the encoded data when in Verify

Traits

BinDecodable

A trait for types which are serializable to and from DNS binary formats

BinEncodable

A type which can be encoded into a DNS binary format

RestrictedMath

Common checked math operations for the Restrict type