rama::dns::hickory::resolver::proto

Module xfer

Available on crate feature dns only.

DNS high level transit implementations.

Primarily there are two types in this module of interest, the DnsMultiplexer type and the DnsHandle type. DnsMultiplexer can be thought of as the state machine responsible for sending and receiving DNS messages. DnsHandle is the type given to API users of the hickory-proto library to send messages into the DnsMultiplexer for delivery. Finally there is the DnsRequest type. This allows for customizations, through DnsRequestOptions, to the delivery of messages via a DnsMultiplexer.

TODO: this module needs some serious refactoring and normalization.

Modules

dns_handle

DnsHandle types perform conversions of the raw DNS messages before sending the messages on the specified streams.

dns_multiplexerstd

DnsMultiplexer and associated types implement the state machines for sending DNS messages while using the underlying streams.

dns_request

DnsRequest wraps a Message and associates a set of DnsRequestOptions for specifying different transfer options.

dns_response

DnsResponse wraps a Message and any associated connection details

retry_dns_handle

RetryDnsHandle allows for DnsQueries to be reattempted on failure

Structs

BufDnsRequestStreamHandlestd

Used for associating a name_server to a DnsRequestStreamHandle

BufDnsStreamHandlestd

A buffering stream bound to a SocketAddr

DnsExchange

This is a generic Exchange implemented over multiplexed DNS connection providers.

DnsExchangeBackground

This background future is responsible for driving all network operations for the DNS protocol.

DnsExchangeConnect

A wrapper for a future DnsExchange connection.

DnsExchangeSend

A Stream that will resolve to Responses after sending the request

DnsMultiplexer

A DNS Client implemented over futures-rs.

DnsMultiplexerConnect

A wrapper for a future DnsExchange connection

DnsRequest

A DNS request object

DnsRequestOptions

A set of options for expressing options to how requests should be treated

DnsResponse

A DNS response object

DnsResponseStreamstd

A stream returning DNS responses

FirstAnswerFuture

See FirstAnswer::first_answer

OneshotDnsRequeststd

A OneshotDnsRequest creates a channel for a response to message

RetryDnsHandle

Can be used to reattempt queries if they fail

SerialMessage

A DNS message in serialized form, with either the target address or source address

Enums

DnsResponseReceiverstd

A Stream that wraps a oneshot::Receiver<Stream> and resolves to items in the inner Stream

Protocol

The protocol on which a NameServer should be communicated with

Traits

DnsClientStreamstd

A non-multiplexed stream of Serialized DNS messages

DnsHandle

A trait for implementing high level functions of DNS.

DnsRequestSenderstd

Types that implement this are capable of sending a serialized DNS message on a stream

DnsStreamHandle

Implementations of Sinks for sending DNS messages

FirstAnswer

Helper trait to convert a Stream of dns response into a Future

Type Aliases

StreamReceiverstd

Receiver handle for peekable fused SerialMessage channel