rama

Module futures

Re-export of the futures and asynk-strim crates.

Exported for your convenience and because it is so fundamental to rama.

Modules

async_stream

asynk-strim

channel

Asynchronous channels.

executorexecutor

Built-in executors and related tools.

ioio and std

Asynchronous I/O.

lockalloc

Futures-powered synchronization primitives.

never

This module contains the Never type.

prelude

A “prelude” for crates using the futures crate.

task

Tools for working with tasks.

Macros

join

Polls multiple futures simultaneously, returning a tuple of all results once complete.

pending

A macro which yields to the event loop once.

pin_mut

Pins a value on the stack.

poll

A macro which returns the result of polling a future once within the current async context.

ready

Extracts the successful type of a Poll<T>.

selectstd

Polls multiple futures and streams simultaneously, executing the branch for the future that finishes first. If multiple futures are ready, one will be pseudo-randomly selected at runtime. Futures directly passed to select! must be Unpin and implement FusedFuture.

select_biased

Polls multiple futures and streams simultaneously, executing the branch for the future that finishes first. Unlike select!, if multiple futures are ready, one will be selected in order of declaration. Futures directly passed to select_biased! must be Unpin and implement FusedFuture.

stream_select

Combines several streams, all producing the same Item type, into one stream. This is similar to select_all but does not require the streams to all be the same type. It also keeps the streams inline, and does not require Box<dyn Stream>s to be allocated. Streams passed to this macro must be Unpin.

try_join

Polls multiple futures simultaneously, resolving to a Result containing either a tuple of the successful outputs or an error.

Structs

TryZip

Future for the try_zip() function.

Zip

Future for the zip() function.

Traits

AsyncBufRead

Read bytes asynchronously.

AsyncBufReadExt

An extension trait which adds utility methods to AsyncBufRead types.

AsyncRead

Read bytes asynchronously.

AsyncReadExt

An extension trait which adds utility methods to AsyncRead types.

AsyncSeek

Seek bytes asynchronously.

AsyncSeekExt

An extension trait which adds utility methods to AsyncSeek types.

AsyncWrite

Write bytes asynchronously.

AsyncWriteExt

An extension trait which adds utility methods to AsyncWrite types.

Future

A future represents an asynchronous computation, commonly obtained by use of async.

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

Sink

A Sink is a value into which other values can be sent, asynchronously.

SinkExt

An extension trait for Sinks that provides a variety of convenient combinator functions.

Stream

A stream of values produced asynchronously.

StreamExt

An extension trait for Streams that provides a variety of convenient combinator functions.

TryFuture

A convenience for futures that return Result values that includes a variety of adapters tailored to such futures.

TryFutureExt

Adapters specific to Result-returning futures

TryStream

A convenience for streams that return Result values that includes a variety of adapters tailored to such futures.

TryStreamExt

Adapters specific to Result-returning streams

Functions

try_zip

Joins two fallible futures, waiting for both to complete or one of them to error.

zip

Joins two futures, waiting for both to complete.