rama::futures::task

Trait LocalSpawn

pub trait LocalSpawn {
    // Required method
    fn spawn_local_obj(
        &self,
        future: LocalFutureObj<'static, ()>,
    ) -> Result<(), SpawnError>;

    // Provided method
    fn status_local(&self) -> Result<(), SpawnError> { ... }
}

The LocalSpawn is similar to Spawn, but allows spawning futures that don’t implement Send.

Required Methods

fn spawn_local_obj( &self, future: LocalFutureObj<'static, ()>, ) -> Result<(), SpawnError>

Spawns a future that will be run to completion.

Errors

The executor may be unable to spawn tasks. Spawn errors should represent relatively rare scenarios, such as the executor having been shut down so that it is no longer able to accept tasks.

Provided Methods

fn status_local(&self) -> Result<(), SpawnError>

Determines whether the executor is able to spawn new tasks.

This method will return Ok when the executor is likely (but not guaranteed) to accept a subsequent spawn attempt. Likewise, an Err return means that spawn is likely, but not guaranteed, to yield an error.