rama::graceful

Function default_signal

pub async fn default_signal()

Available on non-loom and (Unix or Windows) only.

Returns a Future that completes once one of the default signals.

Which on Unix is Ctrl-C (sigint) or sigterm, and on Windows is Ctrl-C, Ctrl-Close or Ctrl-Shutdown.

Exposed to you so you can easily expand it by for example chaining it with a tokio::time::sleep to have a delay before shutdown is triggered.