rama

Module http

Available on crate feature http only.

rama http support

mostly contains re-exports from rama-http and rama-http-backend.

Modules

body

extra http body types and utilities.

clienthttp-full

rama http client support

conn

HTTP connection utilities.

convert

Modules for conversions between types or data.

corehttp-full

Rama http protocol implementation and low level utilities.

header

HTTP header types

headers

Typed HTTP Headers

io

http I/O utilities, e.g. writing http requests/responses in std http format.

layer

Http Layers provided by Rama.

matcher

service::Matchers implementations to match on rama_http_types::Requests.

mime

Re-export of the mime crate.

opentelemetry

Opentelemetry utilities

proto

High level pertaining to the HTTP message protocol.

request

response

serverhttp-full

Rama HTTP server module, which provides the HttpServer type to serve HTTP requests.

service

Http Services provided by Rama.

sse

Server-Sent Events (SSE) support

tlstls

tls features provided from the http layer.

uri

utils

Utilities for HTTP.

wsws

WebSocket (WS) support for rama (RFC 6455).

Structs

Body

The body type used in rama requests and responses.

BodyDataStream

A stream of data frames.

BodyLimit

Can be used to communicate the desire to limit the size of request/response bodies.

HeaderMap

A specialized multimap for header names and values.

HeaderName

Represents an HTTP header field name

HeaderValue

Represents an HTTP header field value.

HttpError

A generic “error” for HTTP connections

InfiniteReader

A(n) (in)finite random byte stream implementing AsyncRead.

Method

The Request Method (VERB)

Request

Represents an HTTP request.

Response

Represents an HTTP response

Scheme

Represents the scheme component of a URI

StatusCode

An HTTP status code (status-code in RFC 9110 et al.).

Uri

The URI component of a request.

Version

Represents a version of the HTTP spec.

Traits

BodyExtractExt

An extension trait for StreamingBody that provides methods to extract data from it.

StreamingBody

Trait representing a streaming body of a Request or Response.

Type Aliases

HttpResult

A Result typedef to use with the http::Error type