rama::http::body

Module util

Available on crate feature http only.

Modules

combinators

Combinators for the Body trait.

Structs

BodyDataStream

A data stream created from a Body.

BodyStream

A stream created from a Body.

Collected

A collected body produced by BodyExt::collect which collects all the DATA frames and trailers.

Empty

A body that is always empty.

Full

A body that consists of a single chunk.

LengthLimitError

An error returned when body length exceeds the configured limit.

Limited

A length limited body.

StreamBody

A body created from a Stream.

Enums

Either

Sum type with two cases: Left and Right, used if a body can be one of two distinct types.

Traits

BodyExt

An extension trait for http_body::Body adding various combinators and adapters