rama::http

Module client

Available on crate features http and http-full only.

rama http client support

Contains re-exports from rama-http-backend::client and adds EasyHttpWebClient, an opiniated http web client which supports most common use cases and provides sensible defaults.

Modules

builder

proxy

Client Http Proxy Connector Support.

Structs

EasyHttpConnectorBuilder

Builder that is designed to easily create a connoector for super::EasyHttpWebClient from most basic use cases

EasyHttpWebClient

An opiniated http client that can be used to serve HTTP requests.

HttpClientService

Internal http sender used to send the actual requests.

HttpConnector

A Service which establishes an HTTP Connection.

HttpConnectorLayer

A Layer that produces an HttpConnector.

MaybeProxiedConnectionsocks5

A connection which will be proxied if a ProxyAddress was configured

ProxyConnectorsocks5

Proxy connector which supports http(s) and socks5(h) proxy address

ProxyConnectorLayersocks5

Proxy connector layer which supports http(s) and socks5(h) proxy address