rama::http::client::proxy

Module layer

Available on crate features http and http-full only.

Http Proxy Connector Layers for Rama Http Clients

Structs

HttpProxyAddressLayer

A Layer which allows you to add a ProxyAddress to the Context in order to have your client connector make a connection via this proxy (e.g. by using HttpProxyConnectorLayer).

HttpProxyAddressService

A Service which allows you to add a ProxyAddress to the Context in order to have your client connector make a connection via this proxy (e.g. by using HttpProxyConnectorLayer).

HttpProxyConnectResponseHeaders

Extension added to the [Context] by HttpProxyConnector to record the headers from a successful CONNECT response.

HttpProxyConnector

A connector which can be used to establish a connection over an HTTP Proxy.

HttpProxyConnectorLayer

A Layer which wraps the given service with a HttpProxyConnector.

MaybeHttpProxiedConnection

A connection which will be proxied if a ProxyAddress was configured

SetProxyAuthHttpHeaderLayer

A Layer which will set the http auth header in case there is a ProxyAddress in the [Context].

SetProxyAuthHttpHeaderService

A Service wwhich will set the http auth header in case there is a ProxyAddress in the [Context].

Enums

HttpProxyError

error that can be returned in case a http proxy did not manage to establish a connection