rama::http::client

Struct EasyHttpConnectorBuilder

pub struct EasyHttpConnectorBuilder<C = (), S = ()> { /* private fields */ }

Available on crate features http and http-full only.

Builder that is designed to easily create a connoector for super::EasyHttpWebClient from most basic use cases

Implementations

impl EasyHttpConnectorBuilder

pub fn new() -> Self

pub fn with_default_transport_connector( self, ) -> EasyHttpConnectorBuilder<TcpConnector, TransportStage>

pub fn with_custom_transport_connector<C>( self, connector: C, ) -> EasyHttpConnectorBuilder<C, TransportStage>

Add a custom transport connector that will be used by this client for the transport layer

impl<T, Stage> EasyHttpConnectorBuilder<T, Stage>

pub fn with_custom_connector<L>( self, connector_layer: L, ) -> EasyHttpConnectorBuilder<L::Service, Stage>
where L: Layer<T>,

Add a custom connector to this Stage.

Adding a custom connector to a stage will not change the state so this can be used to modify behaviour at a specific stage.

impl EasyHttpConnectorBuilder<TcpConnector, TransportStage>

pub fn with_dns_resolver<T: DnsResolver + Clone>( self, resolver: T, ) -> EasyHttpConnectorBuilder<TcpConnector<T>, TransportStage>

Add a custom DnsResolver that will be used by this client

impl<T> EasyHttpConnectorBuilder<T, TransportStage>

pub fn with_custom_tls_proxy_connector<L>( self, connector_layer: L, ) -> EasyHttpConnectorBuilder<L::Service, ProxyTunnelStage>
where L: Layer<T>,

Available on crate features rustls or boring only.

Add a custom proxy tls connector that will be used to setup a tls connection to the proxy

pub fn with_tls_proxy_support_using_boringssl( self, ) -> EasyHttpConnectorBuilder<TlsConnector<T, ConnectorKindTunnel>, ProxyTunnelStage>

Available on crate feature boring only.

Support a tls tunnel to the proxy itself using boringssl

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn with_tls_proxy_support_using_boringssl_config( self, config: Arc<TlsConnectorDataBuilder>, ) -> EasyHttpConnectorBuilder<TlsConnector<T, ConnectorKindTunnel>, ProxyTunnelStage>

Available on crate feature boring only.

Support a tls tunnel to the proxy itself using boringssl and the provided config

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn with_tls_proxy_support_using_rustls( self, ) -> EasyHttpConnectorBuilder<TlsConnector<T, ConnectorKindTunnel>, ProxyTunnelStage>

Available on crate feature rustls only.

Support a tls tunnel to the proxy itself using rustls

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn with_tls_proxy_support_using_rustls_config( self, config: TlsConnectorData, ) -> EasyHttpConnectorBuilder<TlsConnector<T, ConnectorKindTunnel>, ProxyTunnelStage>

Available on crate feature rustls only.

Support a tls tunnel to the proxy itself using rustls and the provided config

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn without_tls_proxy_support( self, ) -> EasyHttpConnectorBuilder<T, ProxyTunnelStage>

Don’t support a tls tunnel to the proxy itself

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

impl<T> EasyHttpConnectorBuilder<T, ProxyTunnelStage>

pub fn with_custom_proxy_connector<L>( self, connector_layer: L, ) -> EasyHttpConnectorBuilder<L::Service, ProxyStage>
where L: Layer<T>,

Add a custom proxy connector that will be used by this client

pub fn with_proxy_support( self, ) -> EasyHttpConnectorBuilder<ProxyConnector<Arc<T>>, ProxyStage>

Available on crate feature socks5 only.

Add support for usage of a http(s) and socks5(h) ProxyAddress to this client

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn with_http_proxy_support( self, ) -> EasyHttpConnectorBuilder<HttpProxyConnector<T>, ProxyStage>

Add support for usage of a http(s) ProxyAddress to this client

Note that a tls proxy is not needed to make a https connection to the final target. It only has an influence on the initial connection to the proxy itself

pub fn with_socks5_proxy_support( self, ) -> EasyHttpConnectorBuilder<Socks5ProxyConnector<T>, ProxyStage>

Available on crate feature socks5 only.

Add support for usage of a socks5(h) ProxyAddress to this client

pub fn without_proxy_support(self) -> EasyHttpConnectorBuilder<T, ProxyStage>

Make a client without proxy support

impl<T> EasyHttpConnectorBuilder<T, ProxyStage>

pub fn with_custom_tls_connector<L>( self, connector_layer: L, ) -> EasyHttpConnectorBuilder<L::Service, TlsStage>
where L: Layer<T>,

Available on crate features rustls or boring only.

Add a custom tls connector that will be used by the client

Note: when using a tls_connector you probably want to also add a RequestVersionAdapter which applies the negotiated http version from tls alpn. This can be achieved by using Self::with_custom_connector just after adding the tls connector.

pub fn with_tls_support_using_boringssl( self, config: Option<Arc<TlsConnectorDataBuilder>>, ) -> EasyHttpConnectorBuilder<RequestVersionAdapter<TlsConnector<T>>, TlsStage>

Available on crate feature boring only.

Support https connections by using boringssl for tls

Note: this also adds a RequestVersionAdapter to automatically change the request version to the one configured with tls alpn. If this is not wanted, use Self::with_custom_tls_connector instead.

pub fn with_tls_support_using_boringssl_and_default_http_version( self, config: Option<Arc<TlsConnectorDataBuilder>>, default_http_version: Version, ) -> EasyHttpConnectorBuilder<RequestVersionAdapter<TlsConnector<T>>, TlsStage>

Available on crate feature boring only.

Same as Self::with_tls_support_using_boringssl but also setting the default TargetHttpVersion in case no ALPN is negotiated.

This is a fairly important detail for proxy purposes given otherwise you might come in situations where the ingress traffic is negotiated to h2, but the egress traffic has no negotiation which would without a default http version remain on h2… In such a case you can get failed requests if the egress server does not handle multiple http versions.

pub fn with_tls_support_using_rustls( self, config: Option<TlsConnectorData>, ) -> EasyHttpConnectorBuilder<RequestVersionAdapter<TlsConnector<T>>, TlsStage>

Available on crate feature rustls only.

Support https connections by using ruslts for tls

Note: this also adds a RequestVersionAdapter to automatically change the request version to the one configured with tls alpn. If this is not wanted, use Self::with_custom_tls_connector instead.

pub fn with_tls_support_using_rustls_and_default_http_version( self, config: Option<TlsConnectorData>, default_http_version: Version, ) -> EasyHttpConnectorBuilder<RequestVersionAdapter<TlsConnector<T>>, TlsStage>

Available on crate feature rustls only.

Same as Self::with_tls_support_using_rustls but also setting the default TargetHttpVersion in case no ALPN is negotiated.

This is a fairly important detail for proxy purposes given otherwise you might come in situations where the ingress traffic is negotiated to h2, but the egress traffic has no negotiation which would without a default http version remain on h2… In such a case you can get failed requests if the egress server does not handle multiple http versions.

pub fn without_tls_support(self) -> EasyHttpConnectorBuilder<T, TlsStage>

Dont support https on this connector

impl<T> EasyHttpConnectorBuilder<T, TlsStage>

pub fn with_default_http_connector<Body>( self, ) -> EasyHttpConnectorBuilder<HttpConnector<T, Body>, HttpStage>

Add http support to this connector

pub fn with_custom_http_connector<L>( self, connector_layer: L, ) -> EasyHttpConnectorBuilder<L::Service, HttpStage>
where L: Layer<T>,

Add a custom http connector that will be run just after tls

impl<T> EasyHttpConnectorBuilder<T, HttpStage>

pub fn try_with_connection_pool<C>( self, config: HttpPooledConnectorConfig, ) -> Result<EasyHttpConnectorBuilder<RequestVersionAdapter<PooledConnector<T, LruDropPool<C, BasicHttpConId>, BasicHttpConnIdentifier>>, PoolStage>, OpaqueError>

Use the default connection pool for this super::EasyHttpWebClient

This will create a LruDropPool using the provided limits and will use BasicHttpConnIdentifier to group connection on protocol and authority, which should cover most common use cases

Use wait_for_pool_timeout to limit how long we wait for the pool to give us a connection

If you need a different pool or custom way to group connection you can use EasyHttpConnectorBuilder::with_custom_connection_pool() to provide you own.

This also applies a RequestVersionAdapter layer to make sure that request versions are adapted when pooled connections are used, which you almost always need, but in case that is unwanted, you can use Self::with_custom_connection_pool instead.

pub fn try_with_default_connection_pool<C>( self, ) -> Result<EasyHttpConnectorBuilder<RequestVersionAdapter<PooledConnector<T, LruDropPool<C, BasicHttpConId>, BasicHttpConnIdentifier>>, PoolStage>, OpaqueError>

Same as Self::try_with_connection_pool but using the default HttpPooledConnectorConfig.

pub fn with_custom_connection_pool<P, R>( self, pool: P, req_to_conn_id: R, wait_for_pool_timeout: Option<Duration>, ) -> EasyHttpConnectorBuilder<PooledConnector<T, P, R>, PoolStage>

Configure this client to use the provided Pool and ReqToConnId

Use wait_for_pool_timeout to limit how long we wait for the pool to give us a connection

Warning: this does not apply a RequestVersionAdapter layer to make sure that request versions are adapted when pooled connections are used, which you almost always. This should be manually added by using Self::with_custom_connector after configuring this pool and providing a RequestVersionAdapter there.

impl<T, S> EasyHttpConnectorBuilder<T, S>

pub fn build_client<Body, ModifiedBody, ConnResponse>( self, ) -> EasyHttpWebClient<Body, T::Output, ()>
where Body: StreamingBody<Data: Send + 'static, Error: Into<BoxError>> + Unpin + Send + 'static, ModifiedBody: StreamingBody<Data: Send + 'static, Error: Into<BoxError>> + Unpin + Send + 'static, T: Service<Request<Body>, Output = EstablishedClientConnection<ConnResponse, Request<ModifiedBody>>, Error = BoxError>, ConnResponse: ExtensionsMut,

Build a super::EasyHttpWebClient using the currently configured connector

pub fn build_connector(self) -> T

Build a connector from the currently configured setup