rama::http::convert

Module curl

Available on crate feature http only.

Convert http requests with or without its payload into a valid curl command.

Functions

cmd_for_request_parts

Create a curl Command for the given HttpRequestParts.

cmd_for_request_parts_and_payload

Create a curl Command for the given HttpRequestParts and payload bytes.

cmd_string_for_request_parts

Create a curl command string for the given HttpRequestParts.

cmd_string_for_request_parts_and_payload

Create a curl command string for the given HttpRequestParts and payload bytes.