rama::http

Crate headers

Available on crate feature http only.

Typed HTTP Headers

rama has the opinion that headers should be strongly-typed, because that’s why we’re using Rust in the first place. To set or get any header, an object must implement the Header trait from this module. Several common headers are already provided, such as Host, ContentType, UserAgent, and others.

Why Typed?

Or, why not stringly-typed? Types give the following advantages:

More difficult to typo, since typos in types should be caught by the compiler

Parsing to a proper type by default

Rama

Crate used by the end-user rama crate and rama crate authors alike.

Learn more about rama:

Github: https://github.com/plabayo/rama

Book: https://ramaproxy.org/book/

Modules

authorization

Authorization header and types.

encoding

Utility types and functions that can be used in context of encoding headers.

exotic

Non-Standard special-purpose headers that are not widely used.

forwarded

privacy

Privacy-related typed headers.

sec_websocket_extensions

Sec-WebSocket-Extensions header value types.

sec_websocket_protocol

specifier

Specifiers that can be used as part of header values.

util

x_robots_tag

Structs

Accept

Accept header, defined in RFC7231

AcceptRanges

Accept-Ranges header, defined in RFC7233

AccessControlAllowCredentials

Access-Control-Allow-Credentials header, as defined on mdn.

AccessControlAllowHeaders

Access-Control-Allow-Headers header, as defined on mdn.

AccessControlAllowMethods

Access-Control-Allow-Methods header, as defined on mdn.

AccessControlAllowOrigin

Access-Control-Allow-Origin header, as defined on mdn.

AccessControlAllowPrivateNetwork

Access-Control-Allow-Private-Network header, as documented in this draft of WICG.

AccessControlExposeHeaders

Access-Control-Expose-Headers header, as defined on mdn.

AccessControlMaxAge

Access-Control-Max-Age header, as defined on mdn.

AccessControlRequestHeaders

Access-Control-Request-Headers header, as defined on mdn.

AccessControlRequestMethod

Access-Control-Request-Method header, as defined on mdn.

AccessControlRequestPrivateNetwork

Access-Control-Request-Private-Network header, as documented in this draft of WICG.

Age

Age header, defined in RFC7234

Allow

Allow header, defined in RFC7231

Authorization

Authorization header, defined in RFC7235

CacheControl

Cache-Control header, defined in RFC7234 with extensions in RFC8246

Connection

Connection header, defined in RFC7230

ContentDisposition

A Content-Disposition header, (re)defined in RFC6266.

ContentEncoding

Content-Encoding header, defined in RFC7231

ContentLength

Content-Length header, defined in RFC7230

ContentLocation

Content-Location header, defined in RFC7231

ContentRange

Content-Range, described in RFC7233

ContentType

Content-Type header, defined in RFC7231

Cookie

Cookie header, defined in RFC6265

Date

Date header, defined in RFC7231

ETag

ETag header, defined in RFC7232

Error

Errors trying to decode a header.

Expect

The Expect header.

Expires

Expires header, defined in RFC7234

Host

The Host header.

IfMatch

If-Match header, defined in RFC7232

IfModifiedSince

If-Modified-Since header, defined in RFC7232

IfNoneMatch

If-None-Match header, defined in RFC7232

IfRange

If-Range header, defined in RFC7233

IfUnmodifiedSince

If-Unmodified-Since header, defined in RFC7232

LastEventId

Last-Event-ID header, defined in WhatWG’s SSE spec

LastModified

Last-Modified header, defined in RFC7232

Location

Location header, defined in RFC7231

Origin

The Origin header.

Pragma

The Pragma header defined by HTTP/1.0.

ProxyAuthorization

Proxy-Authorization header, defined in RFC7235

Range

Range header, defined in RFC7233

Referer

Referer header, defined in RFC7231

ReferrerPolicy

Referrer-Policy header, part of Referrer Policy

RetryAfter

The Retry-After header.

SecWebSocketAccept

The Sec-WebSocket-Accept header.

SecWebSocketExtensions

The Sec-WebSocket-Extensions header, containing one or multiple Extensions.

SecWebSocketKey

The Sec-WebSocket-Key header.

SecWebSocketProtocol

The Sec-WebSocket-Protocol header, containing one or multiple protocols.

SecWebSocketVersion

The Sec-WebSocket-Version header.

Server

Server header, defined in RFC7231

SetCookie

Set-Cookie header, defined RFC6265

StrictTransportSecurity

StrictTransportSecurity header, defined in RFC6797

Te

TE header, defined in RFC7230

TransferEncoding

Transfer-Encoding header, defined in RFC7230

Upgrade

Upgrade header, defined in RFC7230

UserAgent

User-Agent header, defined in RFC7231

Vary

Vary header, defined in RFC7231

XRobotsTag

Enums

After

ClientHint

ContentEncodingDirective

Directive for the ContentEncoding header.

TeDirective

Directive for the Te header.

TransferEncodingDirective

Directive for the TransferEncoding header.

Traits

HeaderDecode

A typed header which can be decoded from one or multiple headers.

HeaderEncode

A typed header which can be encoded into one or multiple headers.

HeaderMapExt

An extension trait adding “typed” methods to http::HeaderMap.

HttpRequestBuilderExt

An extension trait adding “typed” methods to http::request::Builder.

HttpResponseBuilderExt

An extension trait adding “typed” methods to http::response::Builder.

TypedHeader

Base trait for a typed header.

Functions

all_client_hint_header_name_strings

Returns an iterator over all client hint header name strings.

all_client_hint_header_names

Returns an iterator over all client hint header names.

all_client_hints

Returns an iterator over all client hints.