rama::http::headers

Module forwarded

Available on crate feature http only.

Structs

CFConnectingIp

CF-Connecting-IP provides the client IP address connecting to Cloudflare to the origin web server.

ClientIp

Client-Ip is used by some proxy software to set the real client Ip Address (known to them).

Forwarded

Typed header wrapper for rama_net::forwarded::Forwarded;

TrueClientIp

True-Client-IP provides the original client IP address to the origin web server (Cloudflare Enterprise).

Via

The Via general header is added by proxies, both forward and reverse.

XClientIp

X-Client-Ip is used by some proxy software to set the real client Ip Address (known to them).

XForwardedFor

The X-Forwarded-For (XFF) request header is a de-facto standard header for identifying the originating IP address of a client connecting to a web server through a proxy server.

XForwardedHost

The X-Forwarded-Host (XFH) header is a de-facto standard header for identifying the original host requested by the client in the Host HTTP request header.

XForwardedProto

The X-Forwarded-Proto (XFP) header is a de-facto standard header for identifying the protocol (HTTP or HTTPS) that a client used to connect to your proxy or load balancer.

XRealIp

X-Real-Ip is used by some proxy software to set the real client Ip Address (known to them).

Traits

ForwardHeader

A trait for types headers that is used by middleware which supports headers that can be converted into and from Forward data.