rama::http::headers::x_robots_tag

Struct RobotsTag

pub struct RobotsTag { /* private fields */ }

Available on crate feature http only.

A single element of X-Robots-Tag corresponding to the valid values for one bot name

More Information:

List of std directives

SPC: https://www.ietf.org/slides/slides-aicontrolws-server-privacy-control-a-server-to-client-privacy-opt-out-preference-signal-00.pdf

No-AI / No-Image-AI: no source that we aware of, if you know of any please open a PR

Implementations

impl RobotsTag

pub fn custom_rules(&self) -> &[CustomRule]

Custom rules defined for this tag.

pub fn new_custom_rule(rule: CustomRule) -> RobotsTag

Custom rules defined for this tag.

pub fn new_custom_rule_for_bot( rule: CustomRule, name: HeaderValueString, ) -> RobotsTag

Custom rules defined for this tag.

pub fn with_additional_custom_rule(self, rule: CustomRule) -> RobotsTag

Set an additional rule to this tag.

pub fn set_additional_custom_rule(&mut self, rule: CustomRule) -> &mut RobotsTag

Set an additional rule to this tag.

pub fn with_additional_custom_rules( self, rules: impl IntoIterator<Item = CustomRule>, ) -> RobotsTag

Set zero, one or multiple additional rules to this tag.

pub fn set_additional_custom_rules( &mut self, rules: impl IntoIterator<Item = CustomRule>, ) -> &mut RobotsTag

Set zero, one or multiple additional rules to this tag.

pub fn bot_name(&self) -> Option<&HeaderValueString>

Get a reference the robot name that is set.

pub fn maybe_with_bot_name(self, name: Option<HeaderValueString>) -> RobotsTag

Set or overwrite the robot name.

pub fn maybe_set_bot_name( &mut self, name: Option<HeaderValueString>, ) -> &mut RobotsTag

Set or overwrite the robot name.

pub fn with_bot_name(self, name: HeaderValueString) -> RobotsTag

Set or overwrite the robot name.

pub fn set_bot_name(&mut self, name: HeaderValueString) -> &mut RobotsTag

Set or overwrite the robot name.

pub fn without_bot_name(self) -> RobotsTag

Set or overwrite the robot name.

pub fn unset_bot_name(&mut self) -> &mut RobotsTag

Set or overwrite the robot name.

pub fn new_all() -> RobotsTag

No restrictions for indexing or serving in search results. This rule is the default value and has no effect if explicitly listed.

pub fn new_all_for_bot(name: HeaderValueString) -> RobotsTag

No restrictions for indexing or serving in search results. This rule is the default value and has no effect if explicitly listed.

pub fn all(&self) -> bool

No restrictions for indexing or serving in search results. This rule is the default value and has no effect if explicitly listed.

pub fn with_all(self, value: bool) -> RobotsTag

No restrictions for indexing or serving in search results. This rule is the default value and has no effect if explicitly listed.

pub fn set_all(&mut self, value: bool) -> &mut RobotsTag

No restrictions for indexing or serving in search results. This rule is the default value and has no effect if explicitly listed.

pub fn new_no_index() -> RobotsTag

Do not show this page, media, or resource in search results. If omitted, the page, media, or resource may be indexed and shown in search results.

pub fn new_no_index_for_bot(name: HeaderValueString) -> RobotsTag

Do not show this page, media, or resource in search results. If omitted, the page, media, or resource may be indexed and shown in search results.

pub fn no_index(&self) -> bool

Do not show this page, media, or resource in search results. If omitted, the page, media, or resource may be indexed and shown in search results.

pub fn with_no_index(self, value: bool) -> RobotsTag

Do not show this page, media, or resource in search results. If omitted, the page, media, or resource may be indexed and shown in search results.

pub fn set_no_index(&mut self, value: bool) -> &mut RobotsTag

Do not show this page, media, or resource in search results. If omitted, the page, media, or resource may be indexed and shown in search results.

pub fn new_no_follow() -> RobotsTag

Do not follow the links on this page. If omitted, search engines may use the links on the page to discover those linked pages.

pub fn new_no_follow_for_bot(name: HeaderValueString) -> RobotsTag

Do not follow the links on this page. If omitted, search engines may use the links on the page to discover those linked pages.

pub fn no_follow(&self) -> bool

Do not follow the links on this page. If omitted, search engines may use the links on the page to discover those linked pages.

pub fn with_no_follow(self, value: bool) -> RobotsTag

Do not follow the links on this page. If omitted, search engines may use the links on the page to discover those linked pages.

pub fn set_no_follow(&mut self, value: bool) -> &mut RobotsTag

Do not follow the links on this page. If omitted, search engines may use the links on the page to discover those linked pages.

pub fn new_none() -> RobotsTag

Equivalent to noindex, nofollow.

pub fn new_none_for_bot(name: HeaderValueString) -> RobotsTag

Equivalent to noindex, nofollow.

pub fn none(&self) -> bool

Equivalent to noindex, nofollow.

pub fn with_none(self, value: bool) -> RobotsTag

Equivalent to noindex, nofollow.

pub fn set_none(&mut self, value: bool) -> &mut RobotsTag

Equivalent to noindex, nofollow.

pub fn new_no_snippet() -> RobotsTag

Do not show a text snippet or video preview in the search results for this page. A static image thumbnail (if available) may still be visible. If omitted, search engines may generate a text snippet and video preview based on information found on the page.

To exclude certain sections of your content from appearing in search result snippets, use the data-nosnippet HTML attribute.

pub fn new_no_snippet_for_bot(name: HeaderValueString) -> RobotsTag

Do not show a text snippet or video preview in the search results for this page. A static image thumbnail (if available) may still be visible. If omitted, search engines may generate a text snippet and video preview based on information found on the page.

To exclude certain sections of your content from appearing in search result snippets, use the data-nosnippet HTML attribute.

pub fn no_snippet(&self) -> bool

Do not show a text snippet or video preview in the search results for this page. A static image thumbnail (if available) may still be visible. If omitted, search engines may generate a text snippet and video preview based on information found on the page.

To exclude certain sections of your content from appearing in search result snippets, use the data-nosnippet HTML attribute.

pub fn with_no_snippet(self, value: bool) -> RobotsTag

Do not show a text snippet or video preview in the search results for this page. A static image thumbnail (if available) may still be visible. If omitted, search engines may generate a text snippet and video preview based on information found on the page.

To exclude certain sections of your content from appearing in search result snippets, use the data-nosnippet HTML attribute.

pub fn set_no_snippet(&mut self, value: bool) -> &mut RobotsTag

Do not show a text snippet or video preview in the search results for this page. A static image thumbnail (if available) may still be visible. If omitted, search engines may generate a text snippet and video preview based on information found on the page.

To exclude certain sections of your content from appearing in search result snippets, use the data-nosnippet HTML attribute.

pub fn new_index_if_embedded() -> RobotsTag

A search engine is allowed to index the content of a page if it’s embedded in another page through iframes or similar HTML elements, in spite of a noindex rule. indexifembedded only has an effect if it’s accompanied by noindex.

pub fn new_index_if_embedded_for_bot(name: HeaderValueString) -> RobotsTag

A search engine is allowed to index the content of a page if it’s embedded in another page through iframes or similar HTML elements, in spite of a noindex rule. indexifembedded only has an effect if it’s accompanied by noindex.

pub fn index_if_embedded(&self) -> bool

A search engine is allowed to index the content of a page if it’s embedded in another page through iframes or similar HTML elements, in spite of a noindex rule. indexifembedded only has an effect if it’s accompanied by noindex.

pub fn with_index_if_embedded(self, value: bool) -> RobotsTag

A search engine is allowed to index the content of a page if it’s embedded in another page through iframes or similar HTML elements, in spite of a noindex rule. indexifembedded only has an effect if it’s accompanied by noindex.

pub fn set_index_if_embedded(&mut self, value: bool) -> &mut RobotsTag

A search engine is allowed to index the content of a page if it’s embedded in another page through iframes or similar HTML elements, in spite of a noindex rule. indexifembedded only has an effect if it’s accompanied by noindex.

pub fn new_max_snippet(value: u32) -> RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn new_max_snippet_for_bot(value: u32, name: HeaderValueString) -> RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn max_snippet(&self) -> Option<&u32>

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn maybe_with_max_snippet(self, value: Option<u32>) -> RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn maybe_set_max_snippet(&mut self, value: Option<u32>) -> &mut RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn with_max_snippet(self, value: u32) -> RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn set_max_snippet(&mut self, value: u32) -> &mut RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn without_max_snippet(self) -> RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn unset_max_snippet(&mut self) -> &mut RobotsTag

Use a maximum of characters as a textual snippet for this search result.

Ignored if no valid is specified.

pub fn new_max_image_preview(value: MaxImagePreviewSetting) -> RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn new_max_image_preview_for_bot( value: MaxImagePreviewSetting, name: HeaderValueString, ) -> RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn max_image_preview(&self) -> Option<&MaxImagePreviewSetting>

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn maybe_with_max_image_preview( self, value: Option<MaxImagePreviewSetting>, ) -> RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn maybe_set_max_image_preview( &mut self, value: Option<MaxImagePreviewSetting>, ) -> &mut RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn with_max_image_preview(self, value: MaxImagePreviewSetting) -> RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn set_max_image_preview( &mut self, value: MaxImagePreviewSetting, ) -> &mut RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn without_max_image_preview(self) -> RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn unset_max_image_preview(&mut self) -> &mut RobotsTag

The maximum size of an image preview for this page in a search results.

If omitted, search engines may show an image preview of the default size. If you don’t want search engines to use larger thumbnail images, specify a max-image-preview value of standard or none.

pub fn new_max_video_preview(value: i32) -> RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn new_max_video_preview_for_bot( value: i32, name: HeaderValueString, ) -> RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn max_video_preview(&self) -> Option<&i32>

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn maybe_with_max_video_preview(self, value: Option<i32>) -> RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn maybe_set_max_video_preview( &mut self, value: Option<i32>, ) -> &mut RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn with_max_video_preview(self, value: i32) -> RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn set_max_video_preview(&mut self, value: i32) -> &mut RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn without_max_video_preview(self) -> RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn unset_max_video_preview(&mut self) -> &mut RobotsTag

Use a maximum of seconds as a video snippet for videos on this page in search results.

If omitted, search engines may show a video snippet in search results, and the search engine decides how long a preview may be.

Ignored if no valid is specified.

Special values are as follows:

0: At most, a static image may be used, in accordance to the max-image-preview setting.

-1: No video length limit.

pub fn new_no_translate() -> RobotsTag

Don’t offer translation of this page in search results.

If omitted, search engines may translate the search result title and snippet into the language of the search query.

pub fn new_no_translate_for_bot(name: HeaderValueString) -> RobotsTag

Don’t offer translation of this page in search results.

If omitted, search engines may translate the search result title and snippet into the language of the search query.

pub fn no_translate(&self) -> bool

Don’t offer translation of this page in search results.

If omitted, search engines may translate the search result title and snippet into the language of the search query.

pub fn with_no_translate(self, value: bool) -> RobotsTag

Don’t offer translation of this page in search results.

If omitted, search engines may translate the search result title and snippet into the language of the search query.

pub fn set_no_translate(&mut self, value: bool) -> &mut RobotsTag

Don’t offer translation of this page in search results.

If omitted, search engines may translate the search result title and snippet into the language of the search query.

pub fn new_no_image_index() -> RobotsTag

Do not index images on this page.

If omitted, images on the page may be indexed and shown in search results.

pub fn new_no_image_index_for_bot(name: HeaderValueString) -> RobotsTag

Do not index images on this page.

If omitted, images on the page may be indexed and shown in search results.

pub fn no_image_index(&self) -> bool

Do not index images on this page.

If omitted, images on the page may be indexed and shown in search results.

pub fn with_no_image_index(self, value: bool) -> RobotsTag

Do not index images on this page.

If omitted, images on the page may be indexed and shown in search results.

pub fn set_no_image_index(&mut self, value: bool) -> &mut RobotsTag

Do not index images on this page.

If omitted, images on the page may be indexed and shown in search results.

pub fn new_unavailable_after(value: DirectiveDateTime) -> RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn new_unavailable_after_for_bot( value: DirectiveDateTime, name: HeaderValueString, ) -> RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn unavailable_after(&self) -> Option<&DirectiveDateTime>

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn maybe_with_unavailable_after( self, value: Option<DirectiveDateTime>, ) -> RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn maybe_set_unavailable_after( &mut self, value: Option<DirectiveDateTime>, ) -> &mut RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn with_unavailable_after(self, value: DirectiveDateTime) -> RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn set_unavailable_after( &mut self, value: DirectiveDateTime, ) -> &mut RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn without_unavailable_after(self) -> RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn unset_unavailable_after(&mut self) -> &mut RobotsTag

Requests not to show this page in search results after the specified <date/time>.

Ignored if no valid <date/time> is specified. A date must be specified in a format such as RFC 822, RFC 850, or ISO 8601.

By default there is no expiration date for content. If omitted, this page may be shown in search results indefinitely. Crawlers are expected to considerably decrease

pub fn new_no_ai() -> RobotsTag

No AI (e.g. LLM) allowed.

pub fn new_no_ai_for_bot(name: HeaderValueString) -> RobotsTag

No AI (e.g. LLM) allowed.

pub fn no_ai(&self) -> bool

No AI (e.g. LLM) allowed.

pub fn with_no_ai(self, value: bool) -> RobotsTag

No AI (e.g. LLM) allowed.

pub fn set_no_ai(&mut self, value: bool) -> &mut RobotsTag

No AI (e.g. LLM) allowed.

pub fn new_no_image_ai() -> RobotsTag

No Image AI (e.g. LLM) allowed.

pub fn new_no_image_ai_for_bot(name: HeaderValueString) -> RobotsTag

No Image AI (e.g. LLM) allowed.

pub fn no_image_ai(&self) -> bool

No Image AI (e.g. LLM) allowed.

pub fn with_no_image_ai(self, value: bool) -> RobotsTag

No Image AI (e.g. LLM) allowed.

pub fn set_no_image_ai(&mut self, value: bool) -> &mut RobotsTag

No Image AI (e.g. LLM) allowed.

pub fn new_spc() -> RobotsTag

Server Privacy Control

A do-not-sell-or-share preference is when a person requests that their data “not be sold or shared” for instance by activating a Server Privacy Control setting with their web server software or by using web server software that defaults to such a setting (possibly because this setting matches the most common expectations of that tool’s users). When set, this preference indicates that the person expects to create content for the Web with do-not-sell-or-share interactions.

pub fn new_spc_for_bot(name: HeaderValueString) -> RobotsTag

Server Privacy Control

A do-not-sell-or-share preference is when a person requests that their data “not be sold or shared” for instance by activating a Server Privacy Control setting with their web server software or by using web server software that defaults to such a setting (possibly because this setting matches the most common expectations of that tool’s users). When set, this preference indicates that the person expects to create content for the Web with do-not-sell-or-share interactions.

pub fn spc(&self) -> bool

Server Privacy Control

A do-not-sell-or-share preference is when a person requests that their data “not be sold or shared” for instance by activating a Server Privacy Control setting with their web server software or by using web server software that defaults to such a setting (possibly because this setting matches the most common expectations of that tool’s users). When set, this preference indicates that the person expects to create content for the Web with do-not-sell-or-share interactions.

pub fn with_spc(self, value: bool) -> RobotsTag

Server Privacy Control

A do-not-sell-or-share preference is when a person requests that their data “not be sold or shared” for instance by activating a Server Privacy Control setting with their web server software or by using web server software that defaults to such a setting (possibly because this setting matches the most common expectations of that tool’s users). When set, this preference indicates that the person expects to create content for the Web with do-not-sell-or-share interactions.

pub fn set_spc(&mut self, value: bool) -> &mut RobotsTag

Server Privacy Control

A do-not-sell-or-share preference is when a person requests that their data “not be sold or shared” for instance by activating a Server Privacy Control setting with their web server software or by using web server software that defaults to such a setting (possibly because this setting matches the most common expectations of that tool’s users). When set, this preference indicates that the person expects to create content for the Web with do-not-sell-or-share interactions.