rama::http

Module io

Available on crate feature http only.

http I/O utilities, e.g. writing http requests/responses in std http format.

Modules

upgrade

HTTP Upgrades

Functions

write_http_request

Write an HTTP request to a writer in std http format.

write_http_response

Write an HTTP response to a writer in std http format.