rama::http::layer

Module auth

Available on crate feature http only.

Authorization related middleware.

Modules

add_authorization

Add authorization to requests using the Authorization header.

validate_authorization

Authorize requests using ValidateRequest.

Structs

AddAuthorization

Middleware that adds authorization all requests using the Authorization header.

AddAuthorizationLayer

Layer that applies AddAuthorization which adds authorization to all requests using the Authorization header.

HttpAuthorizer

Utility type to allow you to use any Authorizer that works with Credentials to authorize the Authorization header, and return StatusCode::UNAUTHORIZED response with header::WWW_AUTHENTICATE for unauthorized request, tracing the original error for your convenience.