rama::http::layer::auth

Module add_authorization

Available on crate feature http only.

Add authorization to requests using the Authorization header.

Example

use rama_core::bytes::Bytes;

use rama_http::layer::validate_request::{ValidateRequestHeader, ValidateRequestHeaderLayer};
use rama_http::layer::auth::AddAuthorizationLayer;
use rama_http::{Body, Request, Response, StatusCode, header::AUTHORIZATION};
use rama_core::service::service_fn;
use rama_core::{Service, Layer};
use rama_core::error::BoxError;
use rama_net::user::credentials::basic;

let mut client = (
    // Use basic auth with the given username and password
    AddAuthorizationLayer::new(basic!("username", "password")),
).layer(service_that_requires_auth);

// Make a request, we don't have to add the `Authorization` header manually
let response = client
    .serve(Request::new(Body::default()))
    .await?;

assert_eq!(StatusCode::OK, response.status());

Structs

AddAuthorization

Middleware that adds authorization all requests using the Authorization header.

AddAuthorizationLayer

Layer that applies AddAuthorization which adds authorization to all requests using the Authorization header.