rama::http::layer::compression

Module predicate

Available on crate features http and compression only.

Predicates for disabling compression of responses.

Predicates are applied with Compression::compress_when or CompressionLayer::compress_when.

Structs

Always

Predicate to always compress.

And

Two predicates combined into one.

DefaultPredicate

The default predicate used by Compression and CompressionLayer.

DefaultStreamPredicate

NotForContentType

Predicate that wont allow responses with a specific content-type to be compressed.

SizeAbove

Predicate that will only allow compression of responses above a certain size.

Traits

Predicate

Predicate used to determine if a response should be compressed or not.