rama::http::layer

Module dns

Available on crate feature http only.

Layers in function of DNS.

Structs

DnsResolveMode

A vanity Extensions type for others to easily check if eager DNS resolution is enabled.

DnsResolveModeLayer

Layer which can extend Dns (see rama_core) overwrites with mappings.

DnsResolveModeService

Service to support configuring the DNS resolve mode.

DnsResolveModeUsernameParser

A parser which parses DnsResolveModes from username labels and adds it to the Context’s Extensions.