rama::http::layer::follow_redirect::policy

Trait Policy

pub trait Policy<B, E>:
    Send
    + Sync
    + 'static {
    // Required method
    fn redirect(&mut self, attempt: &Attempt<'_>) -> Result<Action, E>;

    // Provided methods
    fn on_request(&mut self, _request: &mut Request<B>) { ... }
    fn clone_body(&mut self, __body: &B) -> Option<B> { ... }
}

Available on crate feature http only.

Trait for the policy on handling redirection responses.

It is important for the policy to be cloneable, where the clone is a fresh instance of the policy, ready to be used in a new request.

Example

Detecting a cyclic redirection:

use std::collections::HashSet;
use rama_http::{Request, Uri};
use rama_http::layer::follow_redirect::policy::{Action, Attempt, Policy};

#[derive(Clone)]
pub struct DetectCycle {
    uris: HashSet<Uri>,
}

impl< B, E> Policy< B, E> for DetectCycle {
    fn redirect(&mut self, attempt: &Attempt<'_>) -> Result<Action, E> {
        if self.uris.contains(attempt.location()) {
            Ok(Action::Stop)
        } else {
            self.uris.insert(attempt.previous().clone());
            Ok(Action::Follow)
        }
    }
}

Required Methods

fn redirect(&mut self, attempt: &Attempt<'_>) -> Result<Action, E>

Invoked when the service received a response with a redirection status code (3xx).

This method returns an Action which indicates whether the service should follow the redirection.

Provided Methods

fn on_request(&mut self, _request: &mut Request<B>)

Invoked right before the service makes a request, regardless of whether it is redirected or not.

This can for example be used to remove sensitive headers from the request or prepare the request in other ways.

The default implementation does nothing.

fn clone_body(&mut self, __body: &B) -> Option<B>

Try to clone a request body before the service makes a redirected request.

If the request body cannot be cloned, return None.

This is not invoked when [B::size_hint][http_body::Body::size_hint] returns zero, in which case B::default() will be used to create a new request body.

The default implementation returns None.