rama::http::layer::har

Module spec

Available on crate feature http only.

Structs

Browser

Cache

This objects contains info about a request coming from browser cache.

CacheState

Content

This object describes details about response content

Cookie

This object contains list of all cookies

Creator

Creator and browser objects share the same structure.

Entry

This object represents a single exportred request with its response and metadata.

Header

Single HTTP Header.

Log

This object represents the root of exported data.

LogFile

This object represents the exported data structure.

Page

PageTimings

This object describes timings for various events (states) fired during the page load.

PostData

This object describes posted data,

PostParam

QueryStringPair

This object contains list of all parameters & values parsed from a query string, if any (embedded in Request object).

Request

This object contains detailed info about performed request.

Response

This object contains detailed info about the response.

Timings

This object describes various phases within request-response round trip.

Enums

HttpVersion