rama::http::layer::required_header

Module request

Available on crate feature http only.

Set required headers on the request, if they are missing.

For now this only sets Host header on http/1.1, as well as always a User-Agent for all versions.

Structs

AddRequiredRequestHeaders

Middleware that sets a header on the request.

AddRequiredRequestHeadersLayer

Layer that applies AddRequiredRequestHeaders which adds a request header.