rama::http::layer::retry

Module managed

Available on crate feature http only.

Managed retry Policy.

See ManagedPolicy for more details.

Structs

DoNotRetry

An Extensions value that can be added to the [Context] of a Request to signal that the request should not be retried.

ManagedPolicy

A managed retry Policy, which allows for an easier interface to configure retrying requests.

Undefined

A type to represent the undefined default type, which is used as the placeholder in the ManagedPolicy, when the user does not provide a specific type.

Traits

CloneInput

A trait that is used to umbrella-cover all possible implementation kinds for the cloning functionality.

RetryRule

A trait that is used to umbrella-cover all possible implementation kinds for the retry rule functionality.