rama::http::layer

Module sensitive_headers

Available on crate feature http only.

Middlewares that mark headers as sensitive.

Example

use rama_http::layer::sensitive_headers::SetSensitiveHeadersLayer;
use rama_http::{Body, Request, Response, header::AUTHORIZATION};
use rama_core::service::service_fn;
use rama_core::{Service, Layer};
use rama_core::error::BoxError;
use std::{iter::once, convert::Infallible};

async fn handle(req: Request) -> Result<Response, Infallible> {
    // ...
}

let mut service = (
    // Mark the `Authorization` header as sensitive so it doesn't show in logs
    //
    // `SetSensitiveHeadersLayer` will mark the header as sensitive on both the
    // request and response.
    //
    // The middleware is constructed from an iterator of headers to easily mark
    // multiple headers at once.
    SetSensitiveHeadersLayer::new(once(AUTHORIZATION)),
).into_layer(service_fn(handle));

// Call the service.
let response = service
    .serve(Request::new(Body::empty()))
    .await?;

Structs

SetSensitiveHeadersLayer

Mark headers as sensitive on both requests and responses.

SetSensitiveRequestHeaders

Mark request headers as sensitive.

SetSensitiveRequestHeadersLayer

Mark request headers as sensitive.

SetSensitiveResponseHeaders

Mark response headers as sensitive.

SetSensitiveResponseHeadersLayer

Mark response headers as sensitive.

Type Aliases

SetSensitiveHeaders

Mark headers as sensitive on both requests and responses.