rama::http::layer::set_header::request

Trait MakeHeaderValue

pub trait MakeHeaderValue<B>:
    Send
    + Sync
    + 'static {
    // Required method
    fn make_header_value(
        &self,
        req: Request<B>,
    ) -> impl Future<Output = (Request<B>, Option<HeaderValue>)> + Send;
}

Available on crate feature http only.

Trait for producing header values.

Used by [SetRequestHeader] and [SetResponseHeader].

This trait is implemented for closures with the correct type signature. Typically users will not have to implement this trait for their own types.

It is also implemented directly for HeaderValue. When a fixed header value should be added to all responses, it can be supplied directly to the middleware.

Required Methods

fn make_header_value( &self, req: Request<B>, ) -> impl Future<Output = (Request<B>, Option<HeaderValue>)> + Send

Try to create a header value from the request or response.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.