rama::http::layer

Module traffic_writer

Available on crate feature http only.

Middleware to write Http traffic in std format.

Can be useful for cli / debug purposes.

Structs

BidirectionalWriter

A writer that can write both requests and responses.

DoNotWriteRequest

Marker struct to indicate that the request should not be printed.

DoNotWriteResponse

Marker struct to indicate that the response should not be printed.

RequestWriterLayer

Middleware to print Http request in std format.

RequestWriterService

Middleware to print Http request in std format.

ResponseWriterLayer

Layer that applies ResponseWriterService which prints the http response in std format.

ResponseWriterService

Middleware to print Http request in std format.

Enums

BidirectionalMessage

The internal message type for the BidirectionalWriter.

WriterMode

Http writer mode.

Traits

RequestWriter

A trait for writing http requests.

ResponseWriter

A trait for writing http responses.