rama::http::layer

Module upgrade

Available on crate feature http-full only.

middleware to handle branching into http upgrade services

See UpgradeService for more details.

Modules

service

upgrade service to handle branching into http upgrade services

Structs

UpgradeLayer

UpgradeLayer is a middleware that can be used to upgrade a request.

UpgradeService

Upgrade service can be used to handle the possibility of upgrading a request, after which it will pass down the transport RW to the attached upgrade service.

Upgraded

An upgraded HTTP connection.