rama::http::layer::validate_request

Trait ValidateRequestFn

pub trait ValidateRequestFn<B, A>:
    Send
    + Sync
    + 'static {
    type ResponseBody;

    // Required method
    fn call(
        &self,
        request: Request<B>,
    ) -> impl Future<Output = Result<Request<B>, Response<Self::ResponseBody>>> + Send;
}

Available on crate feature http only.

Trait for validating requests.

Required Associated Types

type ResponseBody

The body type used for responses to unvalidated requests.

Required Methods

fn call( &self, request: Request<B>, ) -> impl Future<Output = Result<Request<B>, Response<Self::ResponseBody>>> + Send

Validate the request.

If Ok(()) is returned then the request is allowed through, otherwise not.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.