rama::http::layer

Module version_adapter

Available on crate feature http only.

Structs

RequestVersionAdapter

ConnectorService which will adapt the request version if needed.

RequestVersionAdapterLayer

ConnectorService layer which will adapt the request version if needed.

ResponseVersionAdapter

Service which will adapt the response version to the original request version.

ResponseVersionAdapterLayer

Layer which will adapt the response version to the original request version.

Functions

adapt_request_version

Adapt request to match the provided Version

adapt_response_version

Adapt response to match the provided Version