rama::http

Module matcher

Available on crate feature http only.

service::Matchers implementations to match on rama_http_types::Requests.

See service::matcher module for more information.

Modules

uri

provides a UriMatcher matcher for matching requests based on their URI.

Structs

DomainMatcher

Matcher based on the (sub)domain of the request’s URI.

HeaderMatcher

Matcher based on the Request’s headers.

HttpMatcher

A matcher that is used to match an http Request

MethodMatcher

A matcher that matches one or more HTTP methods.

PathMatcher

Matcher based on the URI path.

SubdomainTrieMatcher

A matcher that matches subdomains.

UriMatcher

Matcher the request’s URI, using a substring or regex pattern.

UriParams

parameters that are inserted in the [Context], in case the PathMatcher found a match for the given Request.

UriParamsDeserializeError

Error that can occur during the deserialization of the UriParams.

VersionMatcher

A matcher that matches one or more HTTP methods.

Enums

HttpMatcherKind

A matcher that is used to match an http Request