rama::http::proto::h1

Module headers

Available on crate feature http only.

types and functionality to preserve http1* header casing and order.

This is especially important for proxies and clients… because out there… are wild servers that care about header casing for reasons… You can think of that what you want, but they do and we have to deal with it.

Modules

original

Original order and case tracking for h1 tracking…

Structs

HeaderMapValueRemover

Utility that can be used to be able to remove headers from an HeaderMap in random order, one by one.

HeaderMapValueRemoverIntoIter

Porduced by the IntoIterator implementation for HeaderMapValueRemover.

Http1HeaderMap

Http1HeaderMapIntoIter

Http1HeaderName

Traits

IntoHttp1HeaderName

TryIntoHttp1HeaderName