rama::http::proto::h2

Module frame

Available on crate feature http only.

Structs

Continuation

Data

Data frame

EarlyFrameCapture

EarlyFrameStreamContext

GoAway

Head

Headers

Header frame

Ping

Priority

Pseudo

PushPromise

Reason

HTTP/2 error codes.

Reset

Setting

A struct representing the combination of a SettingId with its u32 value.

SettingOrder

Settings

SettingsConfig

StreamDependency

StreamId

A stream identifier, as described in Section 5.1.1 of RFC 7540.

StreamIdOverflow

WindowUpdate

Enums

EarlyFrame

A Frame which is sent often early before any data

Error

Errors that can occur during parsing an HTTP/2 frame.

Frame

Kind

PushPromiseHeaderError

SettingId

An enum that lists all valid settings that can be sent in a SETTINGS frame.

Constants

DEFAULT_INITIAL_WINDOW_SIZE

The default value of SETTINGS_INITIAL_WINDOW_SIZE

DEFAULT_MAX_FRAME_SIZE

The default value of MAX_FRAME_SIZE

DEFAULT_SETTINGS_HEADER_TABLE_SIZE

The default value of SETTINGS_HEADER_TABLE_SIZE

HEADER_LEN

MAX_MAX_FRAME_SIZE

MAX_FRAME_SIZE upper bound

Functions

parse_u64

Type Aliases

FrameSize