rama::http::service::client

Module ext

Available on crate feature http only.

Structs

RequestBuilder

A builder to construct the properties of a Request.

Traits

HttpClientExt

Extends an Http Client with high level features, to facilitate the creation and sending of http requests, in a more ergonomic way.

IntoHeaderName

A trait to try to convert some type into a HeaderName.

IntoHeaderValue

A trait to try to convert some type into a HeaderValue.

IntoUrl

A trait to try to convert some type into a Url.