rama::http::service::client::ext

Trait HttpClientExt

pub trait HttpClientExt:
    Sized
    + HttpClientExtSealed
    + Send
    + Sync
    + 'static {
    type ExecuteResponse;
    type ExecuteError;

    // Required methods
    fn get(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn post(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn put(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn patch(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn delete(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn head(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn connect(
        &self,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn request(
        &self,
        method: Method,
        url: impl IntoUrl,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>;
    fn build_from_request<Body>(
        &self,
        request: Request<Body>,
    ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>
       where Body: Into<Body>;
    fn execute(
        &self,
        request: Request,
    ) -> impl Future<Output = Result<Self::ExecuteResponse, Self::ExecuteError>>;
}

Available on crate feature http only.

Extends an Http Client with high level features, to facilitate the creation and sending of http requests, in a more ergonomic way.

Required Associated Types

type ExecuteResponse

The response type returned by the execute method.

type ExecuteError

The error type returned by the execute method.

Required Methods

fn get( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a GET request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn post( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a POST request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn put( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a PUT request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn patch( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a PATCH request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn delete( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a DELETE request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn head( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a HEAD request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn connect( &self, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Convenience method to make a CONNECT request to a URL.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn request( &self, method: Method, url: impl IntoUrl, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>

Start building a Request with the Method and Url.

Returns a RequestBuilder, which will allow setting headers and the request body before sending.

Errors

This method fails whenever the supplied Url cannot be parsed.

fn build_from_request<Body>( &self, request: Request<Body>, ) -> RequestBuilder<'_, Self, Self::ExecuteResponse>
where Body: Into<Body>,

Start building a Request, using the given Request.

Returns a RequestBuilder, which will allow setting headers and the request body before sending.

fn execute( &self, request: Request, ) -> impl Future<Output = Result<Self::ExecuteResponse, Self::ExecuteError>>

Executes a Request.

Errors

This method fails if there was an error while sending request.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.