rama::http::service::web

Module extract

Available on crate feature http only.

Extract utilities to develop endpoint services efortless.

Modules

authority

Module in function of the Authority extractor.

body

module in function of extractors for Request bodies

datastar

🚀 Datastar support extractor for rama.

extension

host

Module in function of the Host extractor.

path

Module in function of the Path extractor.

query

Module in function of the Query extractor.

typed_header

module in function of TypedHeader

Structs

Authority

Extractor that resolves the authority of the request.

Body

Extractor to get the response body.

Bytes

Extractor to get the response body, collected as Bytes.

Csv

Wrapper used to create Csv Http Responses, as well as to extract Csv from Http Request bodies.

Extension

Extractor that extracts an extension from the request extensions

Form

Wrapper used to create Form Http Responses, as well as to extract Form from Http Request bodies.

Host

Extractor that resolves the hostname of the request.

Json

Wrapper used to create Json Http Responses, as well as to extract Json from Http Request bodies.

Path

Extractor to get path parameters from the context in deserialized form.

Query

Extractor that deserializes query strings into some type.

State

Extractor for static State provided by the caller of IntoEndpointServiceWithState

Text

Extractor to get the response body, collected as String.

TypedHeader

Extractor to get a TypedHeader from the request.

Traits

FromPartsStateRefPair

Types that can be created from request parts.

FromRequest

Types that can be created from requests.

OptionalFromPartsStateRefPair

Customize the behavior of Option<Self> as a FromPartsStateRefPair extractor.

OptionalFromRequest

Customize the behavior of Option<Self> as a FromRequest extractor.