rama::http::service::web::extract

Module body

Available on crate feature http only.

module in function of extractors for Request bodies

Structs

Body

Extractor to get the response body.

Bytes

Extractor to get the response body, collected as Bytes.

BytesRejection

Rejection type used when the Bytes extractor fails to collect the request body.

Csv

Wrapper used to create Csv Http Responses, as well as to extract Csv from Http Request bodies.

FailedToDeserializeCsv

Rejection type used if the Csv deserialize the payload into the target type.

FailedToDeserializeForm

Rejection type used if the Form deserialize the form into the target type.

FailedToDeserializeJson

Rejection type used if the Json deserialize the payload into the target type.

Form

Wrapper used to create Form Http Responses, as well as to extract Form from Http Request bodies.

InvalidCsvContentType

Rejection type for Csv used if the Content-Type header is missing or its value is not text/csv.

InvalidFormContentType

Rejection type for Form used if the Content-Type header is missing or its value is not application/x-www-form-urlencoded.

InvalidJsonContentType

Rejection type for Json used if the Content-Type header is missing or its value is not application/json.

InvalidTextContentType

Rejection type for Text used if the Content-Type header is missing or its value is not text/plain.

InvalidUtf8Text

Rejection type used if the Text was not valid UTF-8.

Json

Wrapper used to create Json Http Responses, as well as to extract Json from Http Request bodies.

Text

Extractor to get the response body, collected as String.

Enums

CsvRejection

Rejection used for Csv

FormRejection

Rejection used for Form

JsonRejection

Rejection used for Json

TextRejection

Rejection used for Text

Type Aliases

JsonLines