rama::http::service::web::extract

Struct State

pub struct State<S>(pub S);

Available on crate feature http only.

Extractor for static State provided by the caller of IntoEndpointServiceWithState

S can be the exact provided State or any type that implements FromRef<State>

Tuple Fields
0: S