rama::http::sse

Module datastar

Available on crate feature http only.

🚀 Datastar support for rama.

Datastar helps you build reactive web applications with the simplicity of server-side rendering and the power of a full-stack SPA framework.

It’s the combination of a small js library which makes use of SSE among other utilities, this module implements the event data types used from the server-side to send to the client, which makes use of this JS library.

You can join the discord server of 🚀 Datastar at https://discord.gg/sGfFuw9k, after which you can join the #general-rust channel for any datastar specific help.

Combining 🚀 Datastar with 🦙 Rama (ラマ) provides a powerful foundation for your web application—one that empowers you to build and scale without limitations.

Modules

execute_script

Structs

EventDataReader

EventDataLineReader for the EventDataRead implementation of [RemoveSignals].

ExecuteScript

ExecuteScript executes JavaScript in the browser

PatchElements

PatchElements patches HTML elements into the DOM.

PatchElementsReader

EventDataLineReader for the EventDataRead implementation of PatchElements.

PatchSignals

PatchSignals patches signals into the signal store

PatchSignalsReader

EventDataLineReader for the EventDataRead implementation of PatchSignals.

Enums

ElementPatchMode

The mode in which elements are patched into the DOM.

EventData

EventType

The type protocol on top of SSE which allows for core pushed based communication between the server and the client.

Type Aliases

DatastarEvent