rama::http::sse

Module server

Available on crate feature http only.

SSE types for servers.

Structs

KeepAlive

Configure the interval between keep-alive messages, the content of each message, and the associated stream.

KeepAliveStream

A wrapper around a stream that produces keep-alive events

SseResponseBody

An SSE body stream that can be used for SSE events by the server.