rama::http

Module uri

Available on crate feature http only.

Structs

Authority

Represents the authority component of a URI.

Builder

A builder for Uris.

InvalidUri

An error resulting from a failed attempt to construct a URI.

InvalidUriParts

An error resulting from a failed attempt to construct a URI.

Parts

The various parts of a URI.

PathAndQuery

Represents the path component of a URI

Port

The port component of a URI.

Scheme

Represents the scheme component of a URI

Uri

The URI component of a request.

Functions

try_to_strip_path_prefix_from_uri