rama::http

Crate ws

Available on crate features http and ws only.

WebSocket (WS) support for rama (RFC 6455).

Rama

Crate used by the end-user rama crate and rama crate authors alike.

Learn more about rama:

Github: https://github.com/plabayo/rama

Book: https://ramaproxy.org/book/

Modules

handshake

Utilities to aid in the handshake phase of establishing a WebSocket connection.

protocol

Generic WebSocket message stream.

runtime

Async runtime for Rama WebSockets

Structs

AsyncWebSocket

A wrapper around an underlying raw stream which implements the WebSocket protocol.

Utf8Bytes

Utf8 payload.

Enums

Message

An enum representing the various forms of a WebSocket message.

ProtocolError

Indicates the specific type/cause of a protocol error.