rama::http::ws::handshake

Module client

Available on crate features http and ws only.

WebSocket client types and utilities

Structs

AcceptedWebSocketData

ClientWebSocket

Client [WebSocket], used as input-output stream.

HandshakeRequest

Request data to be used by an http client to initiate an http request.

NegotiatedHandshakeRequest

Intermediate websocket handshake created by WebSocketRequestBuilder::initiate_handshake.

WebSocketRequestBuilder

Builder that can be used by clients to initiate the WebSocket handshake.

WithService

WebSocketRequestBuilder inner wrapper type used for a builder, which includes a service, and thus is there to actually send the request as well and even follow up.

Enums

HandshakeError

Client error which can be triggered in case the handshake phase failed.

ResponseValidateError

Client error which can be triggered in case the response validation failed

Traits

HttpClientWebSocketExt

Extends an Http Client with high level features WebSocket features.

Functions

validate_http_server_response

Validate the “accept” response from the http server with whom the client is trying to establish a WebSocket connection.