rama::http::ws

Module protocol

Available on crate features http and ws only.

Generic WebSocket message stream.

Modules

frame

Utilities to work with raw WebSocket frames.

Structs

CloseFrame

A struct representing the close command.

PerMessageDeflateConfigcompression

Per-message-deflate configuration as specified in RFC7692

WebSocket

WebSocket input-output stream.

WebSocketConfig

The configuration for WebSocket connection.

WebSocketContext

A context for managing WebSocket stream.

Enums

Message

An enum representing the various forms of a WebSocket message.

ProtocolError

Indicates the specific type/cause of a protocol error.

Role

Indicates a Client or Server role of the websocket