rama::layer::limit::policy

Trait Policy

pub trait Policy<Input>:
    Send
    + Sync
    + 'static {
    type Guard: Send + 'static;
    type Error: Send + 'static;

    // Required method
    fn check(
        &self,
        input: Input,
    ) -> impl Future<Output = PolicyResult<Input, Self::Guard, Self::Error>> + Send;
}

A limit Policy is used to determine whether a input is allowed to proceed, and if not, how to handle it.

Required Associated Types

type Guard: Send + 'static

The guard type that is returned when the input is allowed to proceed.

See PolicyOutput::Ready.

type Error: Send + 'static

The error type that is returned when the input is not allowed to proceed, and should be aborted.

See PolicyOutput::Abort.

Required Methods

fn check( &self, input: Input, ) -> impl Future<Output = PolicyResult<Input, Self::Guard, Self::Error>> + Send

Check whether the input is allowed to proceed.

Optionally modify the input before it is passed to the inner service, which can be used to add metadata to the input regarding how the input was handled by this limit policy.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.