rama

Module matcher

Matcher utilities for any middleware where need to match on incoming inputs.

This module provides the Matcher trait and convenience utilities around it.

Examples of this are iterator “reducers” as made available via IteratorMatcherExt, as well as optional Matcher::or and Matcher::and trait methods.

These all serve as building blocks together with And, Or, Not and a bool to combine and transform any kind of Matcher.

And finally there is MatchFn, easily created using match_fn to create a Matcher from any compatible Fn.

Structs

And

A matcher that matches if all of the inner matchers match.

ExtensionMatcher

A matcher which allows you to match based on an extension, either by comparing with an extension (ExtensionMatcher::with_const), or using a custom predicate (ExtensionMatcher::with_fn).

MatcherRouter

Wrapper type that can be used to turn a tuple of (Matcher, Service) tuples into a single Service.

Not

A matcher that matches if the inner matcher does not match.

Or

A matcher that matches if any of the inner matchers match.

Traits

IteratorMatcherExt

Extension to apply matcher operations to an Iterator of Matchers.

MatchFn

A MatchFn is a Matcher implemented using a function.

Matcher

A condition to decide whether Input matches for router or other middleware purposes.

Functions

match_fn

Create a MatchFn from a function.