rama::matcher

Trait Matcher

pub trait Matcher<Input>:
    Send
    + Sync
    + 'static {
    // Required method
    fn matches(&self, ext: Option<&mut Extensions>, input: &Input) -> bool;

    // Provided methods
    fn or<M>(self, other: M) -> impl Matcher<Input>
       where Self: Sized,
             M: Matcher<Input> { ... }
    fn and<M>(self, other: M) -> impl Matcher<Input>
       where Self: Sized,
             M: Matcher<Input> { ... }
    fn not(self) -> impl Matcher<Input>
       where Self: Sized { ... }
}

A condition to decide whether Input matches for router or other middleware purposes.

Required Methods

fn matches(&self, ext: Option<&mut Extensions>, input: &Input) -> bool

returns true on a match, false otherwise

ext is None in case the callee is not interested in collecting potential match metadata gathered during the matching process. An example of this path parameters for an http Uri matcher.

Provided Methods

fn or<M>(self, other: M) -> impl Matcher<Input>
where Self: Sized, M: Matcher<Input>,

Provide an alternative matcher to match if the current one does not match.

fn and<M>(self, other: M) -> impl Matcher<Input>
where Self: Sized, M: Matcher<Input>,

Add another condition to match on top of the current one.

fn not(self) -> impl Matcher<Input>
where Self: Sized,

Negate the current condition.