rama

Crate net

Available on crate feature net only.

Rama network types and utilities.

Protocols such as tcp, http, tls, etc are not explicitly implemented here, see the relevant rama crates for this.

Learn more about rama:

Github: https://github.com/plabayo/rama

Book: https://ramaproxy.org/book/

Modules

address

network address types and utilities

asn

autonomous system number (ASN)

client

generic client net logic

conn

Connection utilities

fingerprinttls or http

fingerprint implementations for the network surface

forwarded

rama support for the “Forwarded HTTP Extension”

httphttp

http net support added as part of rama’s net support

httpstls and http

mode

proxy

Proxy utilities and types.

socket

stream

Utilities that operate on a Stream

test_utils

tlstls

rama common tls types

transporthttp

transport net logic

user

types and utilities for network users

Structs

Protocol

Web protocols that are relevant to Rama.