rama::net

Module address

Available on crate feature net only.

network address types and utilities

This module provides the common language to work with the different kind of formats that network addresses come in, and are used as the building stone for other parts of Rama that have to work with “addresses”, regardless if they are domains or IPs, or have ports explicitly specified or not.

Modules

ip

IP constants and utilities

Structs

Authority

A Host with optionally a port and/or user-info (Basic).

Domain

A domain.

DomainAddress

A Domain with an associated port (u16)

DomainParentMatch

Result of DomainTrie::match_parent.

DomainTrie

An efficient radix tree that can be used to match (sub)domains.

HostWithOptPort

A Host with optionally a port.

HostWithPort

A Host with an associated port.

ProxyAddress

Address of a proxy that can be connected to.

SocketAddress

An IpAddr with an associated port (u16)

Enums

Host

Either a Domain or an IpAddr.

Traits

AsDomainRef

A trait which is used by the rama-net crate for places where we wish to have access to a reference to a Domain, directly or indirectly, for non-move purposes.

IntoDomain

A trait which can be use by crates where a Domain is expected, it can however only be implemented by the rama-net rate.