rama::net::address

Trait AsDomainRef

pub trait AsDomainRef: AsDomainRefPrivate {
    // Provided method
    fn as_wildcard_parent(&self) -> Option<Domain> { ... }
}

Available on crate feature net only.

A trait which is used by the rama-net crate for places where we wish to have access to a reference to a Domain, directly or indirectly, for non-move purposes.

For example to compare it, or use it in a derived form.

Provided Methods

fn as_wildcard_parent(&self) -> Option<Domain>