rama::net

Module client

Available on crate feature net only.

generic client net logic

Modules

pool

Structs

BoxedConnectorService

A ConnectorService which only job is to Box the created Service by the inner ConnectorService.

ConnectorTarget

Target [HostWithPort] which if found in extensions is to be used by a connector such as a TCPConnector instead of the requested address, unless a proxy is requested in which case a proxy is to be used instead.

EstablishedClientConnection

The established connection to a server returned for the http client to be used.

Enums

EitherConn

A type to allow you to use multiple types as a single type.

EitherConn3

A type to allow you to use multiple types as a single type.

EitherConn4

A type to allow you to use multiple types as a single type.

EitherConn5

A type to allow you to use multiple types as a single type.

EitherConn6

A type to allow you to use multiple types as a single type.

EitherConn7

A type to allow you to use multiple types as a single type.

EitherConn8

A type to allow you to use multiple types as a single type.

EitherConn9

A type to allow you to use multiple types as a single type.

EitherConn3Connected

A type to allow you to use multiple types as a single type.

EitherConn4Connected

A type to allow you to use multiple types as a single type.

EitherConn5Connected

A type to allow you to use multiple types as a single type.

EitherConn6Connected

A type to allow you to use multiple types as a single type.

EitherConn7Connected

A type to allow you to use multiple types as a single type.

EitherConn8Connected

A type to allow you to use multiple types as a single type.

EitherConn9Connected

A type to allow you to use multiple types as a single type.

EitherConnConnected

A type to allow you to use multiple types as a single type.

Traits

ConnectorService

Glue trait that is used as the Connector trait bound for clients establishing a connection on one layer or another.