rama::net::client

Module pool

Available on crate feature net only.

Modules

httphttp

metricsopentelemetry

Structs

ActiveSlot

Active slot is able to actively use a connection to make requests. They are used to track ‘active’ connections inside the pool

LeasedConnection

LeasedConnection is a connection that is temporarily leased from a pool

LruDropPool

Connection pool that uses LRU to evict connections

NoPool

Connection pool that doesn’t store connections and has no limits.

PoolSlot

Pool slot is needed to add a connection to the pool. Poolslots have a one to one mapping to connections inside the pool, and are used to track the ‘total’ connections inside the pool

PooledConnector

PooledConnectorLayer

Enums

ConnectionResult

Result returned by a successful call to Pool::get_conn

ReuseStrategy

Traits

ConnID

ConnID is used to identify a connection in a connection pool. These IDs are not unique and multiple connections can have the same ID. IDs are used to filter which connections can be used for a specific input in a way that is independent of what an input is.

Pool

[PoolStorage] implements the storage part of a connection pool. This storage also decides which connection it returns for a given ID or when the caller asks to remove one, this results in the storage deciding which mode we use for connection reuse and dropping (eg FIFO for reuse and LRU for dropping conn when pool is full)

ReqToConnID

ReqToConnID is used to convert a Input to a connection ID. These IDs are not unique and multiple connections can have the same ID. IDs are used to filter which connections can be used for a specific input in a way that is independent of what an input is.

Type Aliases

FiFoReuseLruDropPoolDeprecated