rama::net::client::pool

Struct LeasedConnection

pub struct LeasedConnection<C, ID> { /* private fields */ }

Available on crate feature net only.

LeasedConnection is a connection that is temporarily leased from a pool

It will be returned to the pool once dropped if the user didn’t take ownership of the connection C with LeasedConnection::into_connection(). LeasedConnections are considered active pool connections until dropped or ownership is taken of the internal connection.

Implementations

impl<C, ID> LeasedConnection<C, ID>

pub fn into_connection(self) -> C

pub fn mark_as_failed(&self)