rama::net

Module fingerprint

Available on crate feature net and (crate features tls or http) only.

fingerprint implementations for the network surface

Structs

AkamaiH2

HttpRequestInput

Minimal input data structure which can be used by ja4h computation functions instead of a reference to a rama_http_types::Request.

Ja3

Data which can be hashed using Self::hash, and which is also displayed as a “ja3” hash.

Ja4

Input data for a “ja4” hash.

Ja4H

Input data for a “ja4h” hash. or displaying it.

PeetPrint

Input data for a “peetprint” fingerprint.

Enums

AkamaiH2ComputeError

Ja3ComputeError

error identifying a failure in Ja3::compute

Ja4ComputeError

error identifying a failure in Ja4::compute

Ja4HComputeError

error identifying a failure in Ja4H::compute

PeetComputeError

error identifying a failure in PeetPrint::compute

Traits

ClientHelloProvider

Sealed trait used by ja3/ja4 computation functions, to allow you to immediately compute from either a ClientHello or a ClientConfig data structure.

HttpRequestProvider

Sealed trait used by the ja4h computation functions, to allow you to immediately compute from either a rama_http_types::Request or a HttpRequestInput data structure.