rama::net

Module proxy

Available on crate feature net only.

Proxy utilities and types.

Structs

ProxyRequest

A request to proxy between source and target (stream).

ProxyTarget

Target HostWithPort for a proxy/forwarder service.

StreamForwardService

A proxy Service which takes a ProxyRequest and copies the bytes of both the source and target Streams bidirectionally.