rama::net::socket::core

Struct SockRef

pub struct SockRef<'s> { /* private fields */ }

Available on crate feature net only.

A reference to a Socket that can be used to configure socket types other than the Socket type itself.

This allows for example a TcpStream, found in the standard library, to be configured using all the additional methods found in the Socket API.

SockRef can be created from any socket type that implements AsFd (Unix) or AsSocket (Windows) using the From implementation.

Examples

Below is an example of converting a TcpStream into a SockRef.

use std::net::{TcpStream, SocketAddr};

use socket2::SockRef;

// Create `TcpStream` from the standard library.
let address: SocketAddr = "127.0.0.1:1234".parse()?;
let stream = TcpStream::connect(address)?;

// Create a `SockRef`erence to the stream.
let socket_ref = SockRef::from(&stream);
// Use `Socket::set_tcp_nodelay` on the stream.
socket_ref.set_tcp_nodelay(true)?;
drop(socket_ref);

assert_eq!(stream.nodelay()?, true);