rama::net::stream::dep

Module ipnet

Available on crate feature net only.

Re-export of the ipnet crate.

Types for IPv4 and IPv6 network addresses.

Structs

AddrParseError

An error which can be returned when parsing an IP network address.

Ipv4AddrRange

An Iterator over a range of IPv4 addresses.

Ipv4Net

An IPv4 network address.

Ipv4Subnets

An Iterator that generates IPv4 network addresses.

Ipv6AddrRange

An Iterator over a range of IPv6 addresses.

Ipv6Net

An IPv6 network address.

Ipv6Subnets

An Iterator that generates IPv6 network addresses.

PrefixLenError

An error which can be returned when the prefix length is invalid.

Enums

IpAddrRange

An Iterator over a range of IP addresses, either IPv4 or IPv6.

IpNet

An IP network address, either IPv4 or IPv6.

IpSubnets

An Iterator that generates IP network addresses, either IPv4 or IPv6.

Traits

IpAdd

Provides a saturating_add() method for Ipv4Addr and Ipv6Addr.

IpBitAnd

Provides a bitand() method for Ipv4Addr and Ipv6Addr.

IpBitOr

Provides a bitor() method for Ipv4Addr and Ipv6Addr.

IpSub

Provides a saturating_sub() method for Ipv4Addr and Ipv6Addr.

Functions

ip_mask_to_prefix

Converts a IpAddr network mask into a prefix.

ipv4_mask_to_prefix

Converts a Ipv4Addr network mask into a prefix.

ipv6_mask_to_prefix

Converts a Ipv6Addr network mask into a prefix.