rama::net::stream

Module matcher

Available on crate feature net only.

service::Matchers implementations to match on Sockets.

See service::matcher module for more information.

Modules

ip

ip matcher and utilities

Structs

IpNetMatcher

Matcher based on whether or not the IpNet contains the SocketAddr of the peer.

LoopbackMatcher

Matcher based on the ip part of the SocketAddr of the peer, matching only if the ip is a loopback address.

PortMatcher

Matcher based on the port part of the SocketAddr of the peer.

PrivateIpNetMatcher

Matcher based on the ip part of the SocketAddr of the peer, matching only if the IP is considered a private address.

SocketAddressMatcher

Matcher based on the SocketAddr of the peer.

SocketMatcher

A matcher to match on a Socket.