rama::net

Module tls

Available on crate features net and tls only.

rama common tls types

Modules

client

TLS implementation agnostic client types

keylog

Keylog facility used by any tls implementation supported by rama, and which can be used for your owns as well.

server

TLS implementation agnostic server types

Structs

SecureTransport

An Extensions value that can be added to the Context of a transport layer to signal that the transport is secure.

TlsTunnel

Context information that can be provided by tls connectors`, to configure the connection in function on an tls tunnel.

Enums

ApplicationProtocol

The Application Layer Negotiation Protocol (ALPN) identifiers as found in the IANA registry for Tls ExtensionType values.

CertificateCompressionAlgorithm

The CertificateCompressionAlgorithm TLS protocol enum, the algorithm used to compress the certificate. The algorithm MUST be one of the algorithms listed in the peer’s compress_certificate extension.

CipherSuite

The CipherSuite TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

CompressionAlgorithm

The CompressionAlgorithm TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

DataEncoding

Implementation agnostic encoding of common data such as certificates and keys.

ECPointFormat

The ECPointFormat TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

ExtensionId

The ExtensionId enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

KeyLogIntent

Intent for a (tls) keylogger to be used.

ProtocolVersion

The ProtocolVersion TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

SignatureScheme

The SignatureScheme TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

SupportedGroup

The SupportedGroup TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.