rama::net::tls

Enum ExtensionId

pub enum ExtensionId {

    SERVER_NAME,
    MAX_FRAGMENT_LENGTH,
    CLIENT_CERTIFICATE_URL,
    TRUSTED_CA_KEYS,
    TRUNCATED_HMAC,
    STATUS_REQUEST,
    USER_MAPPING,
    CLIENT_AUTHZ,
    SERVER_AUTHZ,
    CERT_TYPE,
    SUPPORTED_GROUPS,
    EC_POINT_FORMATS,
    SRP,
    SIGNATURE_ALGORITHMS,
    USE_SRTP,
    HEARTBEAT,
    APPLICATION_LAYER_PROTOCOL_NEGOTIATION,
    STATUS_REQUEST_V2,
    SIGNED_CERTIFICATE_TIMESTAMP,
    CLIENT_CERTIFICATE_TYPE,
    SERVER_CERTIFICATE_TYPE,
    PADDING,
    ENCRYPT_THEN_MAC,
    EXTENDED_MASTER_SECRET,
    TOKEN_BINDING,
    CACHED_INFO,
    TLS_LTS,
    COMPRESS_CERTIFICATE,
    RECORD_SIZE_LIMIT,
    PWD_PROTECT,
    PWD_CLEAR,
    PASSWORD_SALT,
    TICKET_PINNING,
    TLS_CERT_WITH_EXTERN_PSK,
    DELEGATED_CREDENTIAL,
    SESSION_TICKET,
    TLMSP,
    TLMSP_PROXYING,
    TLMSP_DELEGATE,
    SUPPORTED_EKT_CIPHERS,
    PRE_SHARED_KEY,
    EARLY_DATA,
    SUPPORTED_VERSIONS,
    COOKIE,
    PSK_KEY_EXCHANGE_MODES,
    CERTIFICATE_AUTHORITIES,
    OID_FILTERS,
    POST_HANDSHAKE_AUTH,
    SIGNATURE_ALGORITHMS_CERT,
    KEY_SHARE,
    TRANSPARENCY_INFO,
    CONNECTION_ID,
    EXTERNAL_ID_HASH,
    EXTERNAL_SESSION_ID,
    QUIC_TRANSPORT_PARAMETERS,
    TICKET_REQUEST,
    DNSSEC_CHAIN,
    SEQUENCE_NUMBER_ENCRYPTION_ALGORITHMS,
    RRC,
    NEXT_PROTOCOL_NEGOTIATION,
    OLD_APPLICATION_SETTINGS,
    APPLICATION_SETTINGS,
    ECH_OUTER_EXTENSIONS,
    ENCRYPTED_CLIENT_HELLO,
    RENEGOTIATION_INFO,
    Unknown(u16),

}

Available on crate features net and tls only.

The ExtensionId enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

Variants

SERVER_NAME

MAX_FRAGMENT_LENGTH

CLIENT_CERTIFICATE_URL

TRUSTED_CA_KEYS

TRUNCATED_HMAC

STATUS_REQUEST

USER_MAPPING

CLIENT_AUTHZ

SERVER_AUTHZ

CERT_TYPE

SUPPORTED_GROUPS

EC_POINT_FORMATS

SRP

SIGNATURE_ALGORITHMS

USE_SRTP

HEARTBEAT

APPLICATION_LAYER_PROTOCOL_NEGOTIATION

STATUS_REQUEST_V2

SIGNED_CERTIFICATE_TIMESTAMP

CLIENT_CERTIFICATE_TYPE

SERVER_CERTIFICATE_TYPE

PADDING

ENCRYPT_THEN_MAC

EXTENDED_MASTER_SECRET

TOKEN_BINDING

CACHED_INFO

TLS_LTS

COMPRESS_CERTIFICATE

RECORD_SIZE_LIMIT

PWD_PROTECT

PWD_CLEAR

PASSWORD_SALT

TICKET_PINNING

TLS_CERT_WITH_EXTERN_PSK

DELEGATED_CREDENTIAL

SESSION_TICKET

TLMSP

TLMSP_PROXYING

TLMSP_DELEGATE

SUPPORTED_EKT_CIPHERS

PRE_SHARED_KEY

EARLY_DATA

SUPPORTED_VERSIONS

COOKIE

PSK_KEY_EXCHANGE_MODES

CERTIFICATE_AUTHORITIES

OID_FILTERS

POST_HANDSHAKE_AUTH

SIGNATURE_ALGORITHMS_CERT

KEY_SHARE

TRANSPARENCY_INFO

CONNECTION_ID

EXTERNAL_ID_HASH

EXTERNAL_SESSION_ID

QUIC_TRANSPORT_PARAMETERS

TICKET_REQUEST

DNSSEC_CHAIN

SEQUENCE_NUMBER_ENCRYPTION_ALGORITHMS

RRC

NEXT_PROTOCOL_NEGOTIATION

OLD_APPLICATION_SETTINGS

APPLICATION_SETTINGS

ECH_OUTER_EXTENSIONS

ENCRYPTED_CLIENT_HELLO

RENEGOTIATION_INFO

Unknown(u16)

Implementations

impl ExtensionId

pub fn is_grease(&self) -> bool

Available on (crate features rustls or boring or acme) and crate feature boring only.

returns true if this id is a (tls) grease object