rama::net::tls

Module keylog

Available on crate features net and tls only.

Keylog facility used by any tls implementation supported by rama, and which can be used for your owns as well.

Center to this module is the KeyLogger which is a wrapper around a FS file

Structs

KeyLogFileHandle

Handle to a (tls) keylog file.

Functions

normalize_path

try_new_key_log_file_handle

Get a key log file handle for the given path only one file handle will be opened per unique path String.