rama::net::tls

Module server

Available on crate features net and tls only.

TLS implementation agnostic server types

Structs

DynamicIssuer

Dynamic issuer which internally contains the dyn issuer

NoTlsRejectError

non-tls connection is rejected

SelfSignedData

Data that can be used to configure the self-signed single data

ServerAuthData

Raw private key and certificate data to facilitate server authentication.

ServerCertIssuerData

ServerConfig

Common API to configure a TLS Server

SniRequest

A request ready for SNI routing, usually used in combination with SniRouter.

SniRouter

A Service router that can be used to support routing of tls traffic as well as non-tls traffic.

TlsPeekRouter

A Service router that can be used to support tls traffic as well as non-tls traffic.

Enums

CacheKind

Cache kind that will be used to cache results of certificate issuers

ClientVerifyMode

Mode of client verification by a (tls) server

ServerAuth

The kind of server auth to be used.

ServerCertIssuerKind

A type of ServerAuth which can be used to generate server certs on the fly using the given issuer

Traits

DynamicCertIssuer

Trait that needs to be implemented by cert issuers to support dynamically issueing (external) certs based on client_hello input.

Type Aliases

SniPeekStream

PeekStream alias used by SniRouter.

TlsPeekStream

PeekStream alias used by TlsPeekRouter.