rama

Module proxy

Available on crate features proxy or haproxy or socks5 only.

rama proxy support

Modules

haproxyhaproxy

rama HaProxy support

socks5socks5

SOCKS5 support for Rama.

Structs

LiveUpdateProxyDBproxy

A wrapper around a T ProxyDB which can be updated through the only linked writer LiveUpdateProxyDBSetter.

LiveUpdateProxyDBSetterproxy

Writer to set a new ProxyDB in the linked LiveUpdateProxyDB.

MemoryProxyDBproxy

A fast in-memory ProxyDatabase that is the default choice for Rama.

MemoryProxyDBInsertErrorproxy

The error type that can be returned by MemoryProxyDB when some of the proxies could not be inserted due to a proxy that had a duplicate key or was invalid for some other reason.

MemoryProxyDBQueryErrorproxy

The error type that can be returned by MemoryProxyDB when no proxy could be returned.

Proxyproxy

The selected proxy to use to connect to the proxy.

ProxyContextproxy

The context as relevant to the proxy layer.

ProxyCsvRowReaderproxy

A CSV Reader that can be used to create a Proxy database from a CSV file or raw data.

ProxyCsvRowReaderErrorproxy

An error that can occur when reading a Proxy CSV row.

ProxyDBLayerproxy

A Layer which wraps an inner Service to select a Proxy based on the given input Extensions, and insert, if a Proxy is selected, it in the input Extensions for further processing.

ProxyDBServiceproxy

A Service which selects a Proxy based on the given input Extensions.

ProxyFilterproxy

Filter to select a specific kind of proxy.

ProxyFilterUsernameParserproxy

A parser which parses ProxyFilters from username labels and adds it to the input Extensions.

ProxyIDproxy

ID of the selected proxy. To be inserted into the Context, only if that proxy is selected.

StringFilterproxy

A string filter that normalizes the string prior to consumption.

Enums

MemoryProxyDBInsertErrorKindproxy

The kind of error that MemoryProxyDBInsertError represents.

MemoryProxyDBQueryErrorKindproxy

The kind of error that MemoryProxyDBQueryError represents.

ProxyCsvRowReaderErrorKindproxy

The kind of error that can occur when reading a Proxy CSV row.

ProxyFilterModeproxy

The modus operandi to decide how to deal with a missing ProxyFilter in the input Extensions when selecting a Proxy from the ProxyDB.

Traits

ProxyDBproxy

The trait to implement to provide a proxy database to other facilities, such as connection pools, to provide a proxy based on the given [TransportContext] and ProxyFilter.

ProxyQueryPredicateproxy

Trait that is used by the ProxyDB for providing an optional filter predicate to rule out returned results.

UsernameFormatterproxy

Trait that is used to allow the formatting of a username, e.g. to allow proxy routers to have proxy config labels in the username.

Functions

proxy_db_updaterproxy

Create a new ProxyDB updater which allows you to have a (typically in-memory) ProxyDB which you can update live.