rama::proxy::haproxy::protocol::v2

Trait WriteToHeader

pub trait WriteToHeader {
    // Required method
    fn write_to(&self, writer: &mut Writer) -> Result<usize, Error>;

    // Provided method
    fn to_bytes(&self) -> Result<Vec<u8>, Error> { ... }
}

Available on (crate features proxy or haproxy or socks5) and crate feature haproxy only.

Defines how to write a type as part of a binary PROXY protocol header.

Required Methods

fn write_to(&self, writer: &mut Writer) -> Result<usize, Error>

Write this instance to the given Writer. The Writer returns an IO error when an individual byte slice is longer than u16::MAX. However, the total length of the buffer may exceed u16::MAX.

Provided Methods

fn to_bytes(&self) -> Result<Vec<u8>, Error>

Writes this instance to a temporary buffer and returns the buffer.