rama::proxy::socks5::client

Module bind

Available on (crate features proxy or haproxy or socks5) and crate feature socks5 only.

Contains bind-flow types such as the Binder, used by a socks5 Client as part of the bind-handshake.

Structs

BindError

Error that is returned in case the bind process while waiting for the (2nd) success reply was not successfull.

BindOutput

Output that is returned in case the bind process while waiting for the (2nd) success reply was successfull.

Binder

Binder is used to await for the socks5 server as it has to come back with a reply on whether or not the server has established a connection with the socks5 server.