rama::proxy::socks5::client

Module udp

Available on (crate features proxy or haproxy or socks5) and crate feature socks5 only.

Structs

UdpFramedRelay

A unified Stream and Sink interface to an underlying UdpSocketRelay, using the Encoder and Decoder traits to encode and decode frames.

UdpSocketRelay

UdpSocketRelay ready to relay udp packets.

UdpSocketRelayBinder

Udp Associate binder ready to create a UdpSocketRelay ready to proxy udp packets via the socks5 server