rama::proxy::socks5

Module server

Available on (crate features proxy or haproxy or socks5) and crate feature socks5 only.

Socks5 Server Implementation for Rama.

See Socks5Acceptor for more information, its Default implementation only supports the Command::Connect method using the DefaultConnector, but custom connectors as well as binders and udp associators are optionally possible.

For MITM socks5 proxies you can use LazyConnector as the connector service of Socks5Acceptor.

Modules

bind

udp

Structs

Binder

Only “useful” public Socks5Binder implementation, which actually is able to accept bind requests and process them.

Connector

Proxy Forward Socks5Connector implementation, which actually is able to accept connect requests and process them.

Error

Server-side error returned in case of a failure during the handshake process.

LazyConnector

Lazy Socks5Connector implementation, which accepts a connection but does delegates all the work on the egress side to the inner (stream) service.

NoSocks5RejectError

non-socks5 connection is rejected

Socks5Acceptor

Socks5 server implementation of RFC 1928

Socks5PeekRouter

A Service router that can be used to support socks5 traffic as well as non-socks5 traffic.

UdpRelay

Only “useful” public Socks5UdpAssociator implementation, which actually is able to accept udp-relay requests and process them.

Traits

Socks5Binder

Types which can be used as socks5 Command::Bind drivers on the server side.

Socks5Connector

Types which can be used as socks5 Command::Connect drivers on the server side.

Socks5UdpAssociator

Types which can be used as socks5 Command::UdpAssociate drivers on the server side.

Type Aliases

DefaultBinder

Default Binder type.

DefaultConnector

Default Connector type.

DefaultUdpRelay

Default [UdpBinder] type.

Socks5PeekStream

PeekStream alias used by Socks5PeekRouter.