rama::proxy::socks5::server

Trait Socks5Binder

pub trait Socks5Binder<S>: Socks5BinderSeal<S> { }

Available on (crate features proxy or haproxy or socks5) and crate feature socks5 only.

Types which can be used as socks5 Command::Bind drivers on the server side.

Typically used as a component part of a Socks5Acceptor.

The actual underlying trait is sealed and not exposed for usage. No custom binders can be implemented. You can however customise the individual steps as provided and used by Binder.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.