rama::proxy::socks5::server

Module udp

Available on (crate features proxy or haproxy or socks5) and crate feature socks5 only.

Structs

AsyncUdpInspector

Wrapper used for Async udp inspectors.

DefaultUdpBinder

Default [UdpBinder] implementation.

DirectUdpRelay

An UDP Relay which relays the UDP Packets without any inspection.

RelayRequest

Request to Relay, used by an async UDP inspector Service.

RelayResponse

SyncUdpInspector

Wrapper used for synchronous udp inspectors.

UdpRelay

Only “useful” public Socks5UdpAssociator implementation, which actually is able to accept udp-relay requests and process them.

Enums

RelayDirection

Direction in which we relay

UdpInspectAction

Action defined by UdpInspector for an inspected udp packet.

Traits

Socks5UdpAssociator

Types which can be used as socks5 Command::UdpAssociate drivers on the server side.

Socks5UdpAssociatorSeal

UdpInspector

Inspector of relayed udp packets, handling both north and south traffic.

Type Aliases

DefaultUdpRelay

Default [UdpBinder] type.