rama

Module service

Service type and utilities.

Service are the abstraction of (leaf) services in Rama.

Modules

handler

async fn(...) as crate.

Structs

BoxService

A boxed Service, to serve Inputs with, for where you inputuire dynamic dispatch.

MirrorService

A Service which will simply return the given input as Ok(_), with an Infallible error.

RejectError

Input rejected

RejectService

A Service which always rejects with an error.

Traits

Service

A Service that produces rama services, to serve given an input, be it transport layer Inputs or application layer http requests, or something else entirely.

Functions

service_fn

Create a ServiceFn from a function.