rama

Module stream

Modules

codec

Adaptors from AsyncRead/AsyncWrite to Stream/Sink

io

Helpers for IO related tasks.

json

ndjson support in rama

rewind

Structs

ChainReader

Reader that can be used to chain two readers together.

HeapReader

Reader for reading from a heap-allocated bytes buffer.

PeekStream

a stream which has peeked some data of the inner stream, to be read first prior to any other reading

StackReader

Reader for reading from a stack buffer

Traits

Stream

A stream is a type that implements AsyncRead, AsyncWrite and Send. This is specific to Rama and is directly linked to the supertraits of Tokio.