rama::stream

Module json

ndjson support in rama

Newline Delimited Json streams.

Use SSE If you can, ndjson if you must.

Structs

JsonDecoder

NDJson decoder decoding ndjson stream of bytes into json objects.

JsonEncoder

NDJson encoder.

JsonReadStream

Wraps a Stream of Results of data blocks, i.e. types that reference as byte array, and offers a Stream implementation over parsed NDJSON-records according to Deserialize, forwarding potential errors returned by the wrapped iterator.

JsonWriteStream

Wraps a Stream of Results of items that can be json-serialized and offers a Stream implementation of Bytes (json-encoded)-records according to Serialize, forwarding potential errors returned by the wrapped iterator.

ParseConfig

Configuration for the NDJSON-parser which controls the behavior in various situations.

Enums

EmptyLineHandling

Controls how the parser deals with lines that contain no JSON values.