rama::tcp::client

Module service

Available on crate features tcp and http only.

TCP services for Rama.

Structs

CreatedTcpStreamConnector

Contains a Connector created by a TcpStreamConnectorFactory, together with the [Context] used to create it in relation to.

Forwarder

A TCP forwarder.

TcpConnector

A connector which can be used to establish a TCP connection to a server.

TcpStreamConnectorCloneFactory

Utility implementation of a TcpStreamConnectorFactory which is implemented to allow one to use a Cloneable TcpStreamConnector as a TcpStreamConnectorFactory by cloning itself.

Traits

TcpStreamConnectorFactory

Factory to create a TcpStreamConnector. This is used by the TCP stream service to create a stream within a specific [Context].

Type Aliases

DefaultForwarder

Default Forwarder.