rama::telemetry

Module opentelemetry

Available on crate feature opentelemetry only.

openelemetry module re-exports

This module re-exports the crates supported and used by rama for (open) telemetry, such that you can make use of it for custom metrics, registries and more.

Re-exports

pub use ::opentelemetry_otlp as collector;

Modules

baggage

Primitives for sending name/value data across system boundaries.

context

Execution-scoped context propagation.

global

Utilities for working with global telemetry primitives

logslogs

OpenTelemetry Logs Bridge API

metricsmetrics

OpenTelemetry Metrics API

propagationtrace

OpenTelemetry Propagator interface

sdk

Implements the SDK component of OpenTelemetry.

semantic_conventions

OpenTelemetry semantic conventions are agreed standardized naming patterns for OpenTelemetry things. This crate aims to be the centralized place to interact with these conventions.

tracetrace

API for tracing applications and libraries.

Macros

otel_debug

Macro for logging debug messages in OpenTelemetry.

otel_error

Macro for logging error messages in OpenTelemetry.

otel_info

Note: These macros (otel_info!, otel_warn!, otel_debug!, and otel_error!) are intended to be used internally within OpenTelemetry code or for custom exporters, processors and other plugins. They are not designed for general application logging and should not be used for that purpose.

otel_warn

Macro for logging warning messages in OpenTelemetry.

Structs

Context

An execution-scoped collection of values.

ContextGuard

A guard that resets the current context to the prior context when dropped.

InstrumentationScope

Information about a library or crate providing instrumentation.

InstrumentationScopeBuilder

Configuration options for InstrumentationScope.

Key

The key part of attribute KeyValue pairs.

KeyValue

A key-value pair describing an attribute.

MeterOptions

Options that can be used to customize a meter (middleware) provided by rama.

ServiceInfo

Information about the service producing the metrics.

SpanId

An 8-byte value which identifies a given span.

StringValue

Wrapper for string-like values

TraceFlags

Flags that can be set on a SpanContext.

TraceId

A 16-byte value which identifies a given trace.

Enums

Array

A Value::Array containing homogeneous values.

Value

The value part of attribute KeyValue pairs.

Traits

AttributesFactory

Trait that can be used to implement your own attributes creator. It is used by layers as a starting point for attributes, and they will add their own attributes on top.