rama::telemetry::opentelemetry

Module baggage

Available on crate feature opentelemetry only.

Primitives for sending name/value data across system boundaries.

Baggage is used to annotate telemetry, adding context and information to metrics, traces, and logs. It is a set of name/value pairs describing user-defined properties. Each name in Baggage is associated with exactly one value.

Main types in this module are:

Baggage: A set of name/value pairs describing user-defined properties.

BaggageExt: Extensions for managing Baggage in a Context.

Baggage can be sent between systems using a baggage propagator in accordance with the W3C Baggage specification.

Note: Baggage is not automatically added to any telemetry. Users have to explicitly add baggage entries to telemetry items.

Structs

Baggage

A set of name/value pairs describing user-defined properties.

BaggageMetadata

An optional property set that can be added to Baggage values.

Iter

An iterator over the entries of a Baggage.

KeyValueMetadata

Baggage name/value pairs with their associated metadata.

Traits

BaggageExt

Methods for sorting and retrieving baggage data in a context.