rama::telemetry::opentelemetry::global

Function meter

pub fn meter(name: &'static str) -> Meter

Available on crate feature opentelemetry only.

Creates a named Meter via the currently configured global MeterProvider.

This is a more convenient way of expressing global::meter_provider().meter(name).

NOTE: Calls to meter() return a Meter backed by the global MeterProvider configured during the method invocation. If the global MeterProvider is changed after getting Meter instances from these calls, the Meter instances returned will not reflect the change.