rama::telemetry::opentelemetry

Module logs

Available on crate features opentelemetry and logs only.

OpenTelemetry Logs Bridge API

Structs

NoopLoggerProvider

A no-op implementation of a LoggerProvider.

Enums

AnyValue

Value types for representing arbitrary values in a log record. Note: The tracing and log crates only support basic types that can be converted to these core variants: i64, f64, StringValue, and bool. Any complex and custom types are supported through their Debug implementation, and converted to String. More complex types (Bytes, ListAny, and Map) are included here to meet specification requirements and are available to support custom appenders that may be implemented for other logging crates. These types allow for handling dynamic data structures, so keep in mind the potential performance overhead of using boxed vectors and maps in appenders.

Severity

A normalized severity value.

Traits

LogRecord

SDK implemented trait for managing log records

Logger

The interface for emitting LogRecords.

LoggerProvider

Interfaces that can create Logger instances.