rama::telemetry::opentelemetry

Module metrics

Available on crate features opentelemetry and metrics only.

OpenTelemetry Metrics API

Structs

AsyncInstrumentBuilder

Configuration for building an async instrument.

Counter

An instrument that records increasing values.

Gauge

An instrument that records independent values

Histogram

An instrument that records a distribution of values.

HistogramBuilder

Configuration for building a Histogram.

InstrumentBuilder

Configuration for building a sync instrument.

Meter

Provides the ability to create instruments for recording measurements or accepting callbacks to report measurements.

ObservableCounter

An async instrument that records increasing values.

ObservableGauge

An async instrument that records independent readings.

ObservableUpDownCounter

An async instrument that records increasing or decreasing values.

UpDownCounter

An instrument that records increasing or decreasing values.

Traits

AsyncInstrument

An SDK implemented instrument that records measurements via callback.

InstrumentProvider

SDK implemented trait for creating instruments

MeterProvider

Provides access to named Meter instances, for instrumenting an application or crate.

SyncInstrument

An SDK implemented instrument that records measurements synchronously.

Type Aliases

Callback

A function registered with a Meter that makes observations for the instruments it is registered with.