rama::telemetry::opentelemetry::semantic_conventions

Module trace

Available on crate feature opentelemetry only.

Trace Semantic Conventions

The trace semantic conventions define a set of standardized attributes to be used in Spans.

Usage

use opentelemetry::KeyValue;
use opentelemetry::{global, trace::Tracer as _};
use opentelemetry_semantic_conventions as semconv;

let tracer = global::tracer("my-component");
let _span = tracer
    .span_builder("span-name")
    .with_attributes([
        KeyValue::new(semconv::trace::CLIENT_ADDRESS, "example.org"),
        KeyValue::new(semconv::trace::CLIENT_PORT, 80i64),
    ])
    .start(&tracer);

Constants

ANDROID_APP_STATEsemconv_experimental

This attribute represents the state of the application.

APP_SCREEN_COORDINATE_Xsemconv_experimental

The x (horizontal) coordinate of a screen coordinate, in screen pixels.

APP_SCREEN_COORDINATE_Ysemconv_experimental

The y (vertical) component of a screen coordinate, in screen pixels.

APP_WIDGET_IDsemconv_experimental

An identifier that uniquely differentiates this widget from other widgets in the same application.

APP_WIDGET_NAMEsemconv_experimental

The name of an application widget.

AWS_BEDROCK_GUARDRAIL_IDsemconv_experimental

The unique identifier of the AWS Bedrock Guardrail. A guardrail helps safeguard and prevent unwanted behavior from model responses or user messages.

AWS_BEDROCK_KNOWLEDGE_BASE_IDsemconv_experimental

The unique identifier of the AWS Bedrock Knowledge base. A knowledge base is a bank of information that can be queried by models to generate more relevant responses and augment prompts.

AWS_DYNAMODB_ATTRIBUTES_TO_GETsemconv_experimental

The value of the AttributesToGet request parameter.

AWS_DYNAMODB_ATTRIBUTE_DEFINITIONSsemconv_experimental

The JSON-serialized value of each item in the AttributeDefinitions request field.

AWS_DYNAMODB_CONSISTENT_READsemconv_experimental

The value of the ConsistentRead request parameter.

AWS_DYNAMODB_CONSUMED_CAPACITYsemconv_experimental

The JSON-serialized value of each item in the ConsumedCapacity response field.

AWS_DYNAMODB_COUNTsemconv_experimental

The value of the Count response parameter.

AWS_DYNAMODB_EXCLUSIVE_START_TABLEsemconv_experimental

The value of the ExclusiveStartTableName request parameter.

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEXESsemconv_experimental

The JSON-serialized value of each item of the GlobalSecondaryIndexes request field

AWS_DYNAMODB_GLOBAL_SECONDARY_INDEX_UPDATESsemconv_experimental

The JSON-serialized value of each item in the GlobalSecondaryIndexUpdates request field.

AWS_DYNAMODB_INDEX_NAMEsemconv_experimental

The value of the IndexName request parameter.

AWS_DYNAMODB_ITEM_COLLECTION_METRICSsemconv_experimental

The JSON-serialized value of the ItemCollectionMetrics response field.

AWS_DYNAMODB_LIMITsemconv_experimental

The value of the Limit request parameter.

AWS_DYNAMODB_LOCAL_SECONDARY_INDEXESsemconv_experimental

The JSON-serialized value of each item of the LocalSecondaryIndexes request field.

AWS_DYNAMODB_PROJECTIONsemconv_experimental

The value of the ProjectionExpression request parameter.

AWS_DYNAMODB_PROVISIONED_READ_CAPACITYsemconv_experimental

The value of the ProvisionedThroughput.ReadCapacityUnits request parameter.

AWS_DYNAMODB_PROVISIONED_WRITE_CAPACITYsemconv_experimental

The value of the ProvisionedThroughput.WriteCapacityUnits request parameter.

AWS_DYNAMODB_SCANNED_COUNTsemconv_experimental

The value of the ScannedCount response parameter.

AWS_DYNAMODB_SCAN_FORWARDsemconv_experimental

The value of the ScanIndexForward request parameter.

AWS_DYNAMODB_SEGMENTsemconv_experimental

The value of the Segment request parameter.

AWS_DYNAMODB_SELECTsemconv_experimental

The value of the Select request parameter.

AWS_DYNAMODB_TABLE_COUNTsemconv_experimental

The number of items in the TableNames response parameter.

AWS_DYNAMODB_TABLE_NAMESsemconv_experimental

The keys in the RequestItems object field.

AWS_DYNAMODB_TOTAL_SEGMENTSsemconv_experimental

The value of the TotalSegments request parameter.

AWS_EXTENDED_REQUEST_IDsemconv_experimental

The AWS extended request ID as returned in the response header x-amz-id-2.

AWS_LAMBDA_INVOKED_ARNsemconv_experimental

The full invoked ARN as provided on the Context passed to the function (Lambda-Runtime-Invoked-Function-Arn header on the /runtime/invocation/next applicable).

AWS_LAMBDA_RESOURCE_MAPPING_IDsemconv_experimental

The UUID of the AWS Lambda EvenSource Mapping. An event source is mapped to a lambda function. It’s contents are read by Lambda and used to trigger a function. This isn’t available in the lambda execution context or the lambda runtime environtment. This is going to be populated by the AWS SDK for each language when that UUID is present. Some of these operations are Create/Delete/Get/List/Update EventSourceMapping.

AWS_REQUEST_IDsemconv_experimental

The AWS request ID as returned in the response headers x-amzn-requestid, x-amzn-request-id or x-amz-request-id.

AWS_S3_BUCKETsemconv_experimental

The S3 bucket name the request refers to. Corresponds to the --bucket parameter of the S3 API operations.

AWS_S3_COPY_SOURCEsemconv_experimental

The source object (in the form bucket/key) for the copy operation.

AWS_S3_DELETEsemconv_experimental

The delete request container that specifies the objects to be deleted.

AWS_S3_KEYsemconv_experimental

The S3 object key the request refers to. Corresponds to the --key parameter of the S3 API operations.

AWS_S3_PART_NUMBERsemconv_experimental

The part number of the part being uploaded in a multipart-upload operation. This is a positive integer between 1 and 10,000.

AWS_S3_UPLOAD_IDsemconv_experimental

Upload ID that identifies the multipart upload.

AZURE_CLIENT_IDsemconv_experimental

The unique identifier of the client instance.

AZURE_COSMOSDB_CONNECTION_MODEsemconv_experimental

Cosmos client connection mode.

AZURE_COSMOSDB_CONSISTENCY_LEVELsemconv_experimental

Account or request consistency level.

AZURE_COSMOSDB_OPERATION_CONTACTED_REGIONSsemconv_experimental

List of regions contacted during operation in the order that they were contacted. If there is more than one region listed, it indicates that the operation was performed on multiple regions i.e. cross-regional call.

AZURE_COSMOSDB_OPERATION_REQUEST_CHARGEsemconv_experimental

The number of request units consumed by the operation.

AZURE_COSMOSDB_REQUEST_BODY_SIZEsemconv_experimental

Request payload size in bytes.

AZURE_COSMOSDB_RESPONSE_SUB_STATUS_CODEsemconv_experimental

Cosmos DB sub status code.

AZURE_RESOURCE_PROVIDER_NAMESPACEsemconv_experimental

Azure Resource Provider Namespace as recognized by the client.

AZURE_SERVICE_REQUEST_IDsemconv_experimental

The unique identifier of the service request. It’s generated by the Azure service and returned with the response.

AZ_SERVICE_REQUEST_IDDeprecatedsemconv_experimental

Deprecated, use azure.service.request.id instead.

CASSANDRA_CONSISTENCY_LEVELsemconv_experimental

The consistency level of the query. Based on consistency values from CQL.

CASSANDRA_COORDINATOR_DCsemconv_experimental

The data center of the coordinating node for a query.

CASSANDRA_COORDINATOR_IDsemconv_experimental

The ID of the coordinating node for a query.

CASSANDRA_PAGE_SIZEsemconv_experimental

The fetch size used for paging, i.e. how many rows will be returned at once.

CASSANDRA_QUERY_IDEMPOTENTsemconv_experimental

Whether or not the query is idempotent.

CASSANDRA_SPECULATIVE_EXECUTION_COUNTsemconv_experimental

The number of times a query was speculatively executed. Not set or 0 if the query was not executed speculatively.

CICD_PIPELINE_ACTION_NAMEsemconv_experimental

The kind of action a pipeline run is performing.

CICD_PIPELINE_RESULTsemconv_experimental

The result of a pipeline run.

CICD_PIPELINE_TASK_NAMEsemconv_experimental

The human readable name of a task within a pipeline. Task here most closely aligns with a computing process in a pipeline. Other terms for tasks include commands, steps, and procedures.

CICD_PIPELINE_TASK_RUN_IDsemconv_experimental

The unique identifier of a task run within a pipeline.

CICD_PIPELINE_TASK_RUN_RESULTsemconv_experimental

The result of a task run.

CICD_PIPELINE_TASK_RUN_URL_FULLsemconv_experimental

The URL of the pipeline task run, providing the complete address in order to locate and identify the pipeline task run.

CLIENT_ADDRESS

Client address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.

CLIENT_PORT

Client port number.

CLOUD_REGIONsemconv_experimental

The geographical region within a cloud provider. When associated with a resource, this attribute specifies the region where the resource operates. When calling services or APIs deployed on a cloud, this attribute identifies the region where the called destination is deployed.

CLOUD_RESOURCE_IDsemconv_experimental

Cloud provider-specific native identifier of the monitored cloud resource (e.g. an ARN on AWS, a fully qualified resource ID on Azure, a full resource name on GCP)

DB_COLLECTION_NAME

The name of a collection (table, container) within the database.

DB_NAMESPACE

The name of the database, fully qualified within the server address and port.

DB_OPERATION_BATCH_SIZE

The number of queries included in a batch operation.

DB_OPERATION_NAME

The name of the operation or command being executed.

DB_OPERATION_PARAMETERsemconv_experimental

A database operation parameter, with key being the parameter name, and the attribute value being a string representation of the parameter value.

DB_QUERY_PARAMETERsemconv_experimental

A database query parameter, with key being the parameter name, and the attribute value being a string representation of the parameter value.

DB_QUERY_SUMMARY

Low cardinality summary of a database query.

DB_QUERY_TEXT

The database query being executed.

DB_RESPONSE_RETURNED_ROWSsemconv_experimental

Number of rows returned by the operation.

DB_RESPONSE_STATUS_CODE

Database response status code.

DB_STORED_PROCEDURE_NAME

The name of a stored procedure within the database.

DB_SYSTEM_NAME

The database management system (DBMS) product as identified by the client instrumentation.

DNS_ANSWERSsemconv_experimental

The list of IPv4 or IPv6 addresses resolved during DNS lookup.

DNS_QUESTION_NAMEsemconv_experimental

The name being queried.

ELASTICSEARCH_NODE_NAMEsemconv_experimental

Represents the human-readable identifier of the node/instance to which a request was routed.

ERROR_MESSAGEsemconv_experimental

A message providing more detail about an error in human-readable form.

ERROR_TYPE

Describes a class of error the operation ended with.

EXCEPTION_ESCAPEDDeprecated

Indicates that the exception is escaping the scope of the span.

EXCEPTION_MESSAGE

The exception message.

EXCEPTION_STACKTRACE

A stacktrace as a string in the natural representation for the language runtime. The representation is to be determined and documented by each language SIG.

EXCEPTION_TYPE

The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.

FAAS_COLDSTARTsemconv_experimental

A boolean that is true if the serverless function is executed for the first time (aka cold-start).

FAAS_CRONsemconv_experimental

A string containing the schedule period as Cron Expression.

FAAS_DOCUMENT_COLLECTIONsemconv_experimental

The name of the source on which the triggering operation was performed. For example, in Cloud Storage or S3 corresponds to the bucket name, and in Cosmos DB to the database name.

FAAS_DOCUMENT_NAMEsemconv_experimental

The document name/table subjected to the operation. For example, in Cloud Storage or S3 is the name of the file, and in Cosmos DB the table name.

FAAS_DOCUMENT_OPERATIONsemconv_experimental

Describes the type of the operation that was performed on the data.

FAAS_DOCUMENT_TIMEsemconv_experimental

A string containing the time when the data was accessed in the ISO 8601 format expressed in UTC.

FAAS_INVOKED_NAMEsemconv_experimental

The name of the invoked function.

FAAS_INVOKED_PROVIDERsemconv_experimental

The cloud provider of the invoked function.

FAAS_INVOKED_REGIONsemconv_experimental

The cloud region of the invoked function.

FAAS_TIMEsemconv_experimental

A string containing the function invocation time in the ISO 8601 format expressed in UTC.

FAAS_TRIGGERsemconv_experimental

Type of the trigger which caused this function invocation.

FEATURE_FLAG_CONTEXT_IDsemconv_experimental

The unique identifier for the flag evaluation context. For example, the targeting key.

FEATURE_FLAG_KEYsemconv_experimental

The lookup key of the feature flag.

FEATURE_FLAG_PROVIDER_NAMEsemconv_experimental

Identifies the feature flag provider.

FEATURE_FLAG_RESULT_REASONsemconv_experimental

The reason code which shows how a feature flag value was determined.

FEATURE_FLAG_RESULT_VALUEsemconv_experimental

The evaluated value of the feature flag.

FEATURE_FLAG_RESULT_VARIANTsemconv_experimental

A semantic identifier for an evaluated flag value.

FEATURE_FLAG_SET_IDsemconv_experimental

The identifier of the flag set to which the feature flag belongs.

FEATURE_FLAG_VERSIONsemconv_experimental

The version of the ruleset used during the evaluation. This may be any stable value which uniquely identifies the ruleset.

GEN_AI_AGENT_DESCRIPTIONsemconv_experimental

Free-form description of the GenAI agent provided by the application.

GEN_AI_AGENT_IDsemconv_experimental

The unique identifier of the GenAI agent.

GEN_AI_AGENT_NAMEsemconv_experimental

Human-readable name of the GenAI agent provided by the application.

GEN_AI_CONVERSATION_IDsemconv_experimental

The unique identifier for a conversation (session, thread), used to store and correlate messages within this conversation.

GEN_AI_DATA_SOURCE_IDsemconv_experimental

The data source identifier.

GEN_AI_OPENAI_REQUEST_SERVICE_TIERsemconv_experimental

The service tier requested. May be a specific tier, default, or auto.

GEN_AI_OPENAI_RESPONSE_SERVICE_TIERsemconv_experimental

The service tier used for the response.

GEN_AI_OPENAI_RESPONSE_SYSTEM_FINGERPRINTsemconv_experimental

A fingerprint to track any eventual change in the Generative AI environment.

GEN_AI_OPERATION_NAMEsemconv_experimental

The name of the operation being performed.

GEN_AI_OUTPUT_TYPEsemconv_experimental

Represents the content type requested by the client.

GEN_AI_REQUEST_CHOICE_COUNTsemconv_experimental

The target number of candidate completions to return.

GEN_AI_REQUEST_ENCODING_FORMATSsemconv_experimental

The encoding formats requested in an embeddings operation, if specified.

GEN_AI_REQUEST_FREQUENCY_PENALTYsemconv_experimental

The frequency penalty setting for the GenAI request.

GEN_AI_REQUEST_MAX_TOKENSsemconv_experimental

The maximum number of tokens the model generates for a request.

GEN_AI_REQUEST_MODELsemconv_experimental

The name of the GenAI model a request is being made to.

GEN_AI_REQUEST_PRESENCE_PENALTYsemconv_experimental

The presence penalty setting for the GenAI request.

GEN_AI_REQUEST_SEEDsemconv_experimental

Requests with same seed value more likely to return same result.

GEN_AI_REQUEST_STOP_SEQUENCESsemconv_experimental

List of sequences that the model will use to stop generating further tokens.

GEN_AI_REQUEST_TEMPERATUREsemconv_experimental

The temperature setting for the GenAI request.

GEN_AI_REQUEST_TOP_Ksemconv_experimental

The top_k sampling setting for the GenAI request.

GEN_AI_REQUEST_TOP_Psemconv_experimental

The top_p sampling setting for the GenAI request.

GEN_AI_RESPONSE_FINISH_REASONSsemconv_experimental

Array of reasons the model stopped generating tokens, corresponding to each generation received.

GEN_AI_RESPONSE_IDsemconv_experimental

The unique identifier for the completion.

GEN_AI_RESPONSE_MODELsemconv_experimental

The name of the model that generated the response.

GEN_AI_SYSTEMsemconv_experimental

The Generative AI product as identified by the client or server instrumentation.

GEN_AI_TOOL_CALL_IDsemconv_experimental

The tool call identifier.

GEN_AI_TOOL_DESCRIPTIONsemconv_experimental

The tool description.

GEN_AI_TOOL_NAMEsemconv_experimental

Name of the tool utilized by the agent.

GEN_AI_USAGE_INPUT_TOKENSsemconv_experimental

The number of tokens used in the GenAI input (prompt).

GEN_AI_USAGE_OUTPUT_TOKENSsemconv_experimental

The number of tokens used in the GenAI response (completion).

GRAPHQL_DOCUMENTsemconv_experimental

The GraphQL document being executed.

GRAPHQL_OPERATION_NAMEsemconv_experimental

The name of the operation being executed.

GRAPHQL_OPERATION_TYPEsemconv_experimental

The type of the operation being executed.

HTTP_REQUEST_BODY_SIZEsemconv_experimental

The size of the request payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_REQUEST_HEADER

HTTP request headers, key being the normalized HTTP Header name (lowercase), the value being the header values.

HTTP_REQUEST_METHOD

HTTP request method.

HTTP_REQUEST_METHOD_ORIGINAL

Original HTTP method sent by the client in the request line.

HTTP_REQUEST_RESEND_COUNT

The ordinal number of request resending attempt (for any reason, including redirects).

HTTP_REQUEST_SIZEsemconv_experimental

The total size of the request in bytes. This should be the total number of bytes sent over the wire, including the request line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and request body if any.

HTTP_RESPONSE_BODY_SIZEsemconv_experimental

The size of the response payload body in bytes. This is the number of bytes transferred excluding headers and is often, but not always, present as the Content-Length header. For requests using transport encoding, this should be the compressed size.

HTTP_RESPONSE_HEADER

HTTP response headers, key being the normalized HTTP Header name (lowercase), the value being the header values.

HTTP_RESPONSE_SIZEsemconv_experimental

The total size of the response in bytes. This should be the total number of bytes sent over the wire, including the status line (HTTP/1.1), framing (HTTP/2 and HTTP/3), headers, and response body and trailers if any.

HTTP_RESPONSE_STATUS_CODE

HTTP response status code.

HTTP_ROUTE

The matched route, that is, the path template in the format used by the respective server framework.

IOS_APP_STATEsemconv_experimental

This attribute represents the state of the application.

NETWORK_LOCAL_ADDRESS

Local address of the network connection - IP address or Unix domain socket name.

NETWORK_LOCAL_PORT

Local port number of the network connection.

NETWORK_PEER_ADDRESS

Peer address of the network connection - IP address or Unix domain socket name.

NETWORK_PEER_PORT

Peer port number of the network connection.

NETWORK_PROTOCOL_NAME

OSI application layer or non-OSI equivalent.

NETWORK_PROTOCOL_VERSION

The actual version of the protocol used for network communication.

NETWORK_TRANSPORT

OSI transport layer or inter-process communication method.

NETWORK_TYPE

OSI network layer or non-OSI equivalent.

PROCESS_COMMAND_ARGSsemconv_experimental

All the command arguments (including the command/executable itself) as received by the process. On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according to the list of null-delimited strings extracted from proc/[pid]/cmdline. For libc-based executables, this would be the full argv vector passed to main. SHOULD NOT be collected by default unless there is sanitization that excludes sensitive data.

PROCESS_EXECUTABLE_NAMEsemconv_experimental

The name of the process executable. On Linux based systems, this SHOULD be set to the base name of the target of /proc/[pid]/exe. On Windows, this SHOULD be set to the base name of GetProcessImageFileNameW.

PROCESS_EXECUTABLE_PATHsemconv_experimental

The full path to the process executable. On Linux based systems, can be set to the target of proc/[pid]/exe. On Windows, can be set to the result of GetProcessImageFileNameW.

PROCESS_EXIT_CODEsemconv_experimental

The exit code of the process.

PROCESS_PIDsemconv_experimental

Process identifier (PID).

RPC_MESSAGE_COMPRESSED_SIZEsemconv_experimental

Compressed size of the message in bytes.

RPC_MESSAGE_IDsemconv_experimental

MUST be calculated as two different counters starting from 1 one for sent messages and one for received message.

RPC_MESSAGE_TYPEsemconv_experimental

Whether this is a received or sent message.

RPC_MESSAGE_UNCOMPRESSED_SIZEsemconv_experimental

Uncompressed size of the message in bytes.

RPC_METHODsemconv_experimental

The name of the (logical) method being called, must be equal to the $method part in the span name.

RPC_SERVICEsemconv_experimental

The full (logical) name of the service being called, including its package name, if applicable.

RPC_SYSTEMsemconv_experimental

A string identifying the remoting system. See below for a list of well-known identifiers.

SERVER_ADDRESS

Server domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.

SERVER_PORT

Server port number.

SESSION_IDsemconv_experimental

A unique id to identify a session.

SESSION_PREVIOUS_IDsemconv_experimental

The previous session.id for this user, when known.

TLS_PROTOCOL_NAMEsemconv_experimental

Normalized lowercase protocol name parsed from original string of the negotiated SSL/TLS protocol version

TLS_PROTOCOL_VERSIONsemconv_experimental

Numeric part of the version parsed from the original string of the negotiated SSL/TLS protocol version

URL_FULL

Absolute URL describing a network resource according to RFC3986

URL_PATH

The URI path component

URL_QUERY

The URI query component

URL_SCHEME

The URI scheme component identifying the used protocol.

URL_TEMPLATEsemconv_experimental

The low-cardinality template of an absolute path reference.

USER_AGENT_ORIGINAL

Value of the HTTP User-Agent header sent by the client.

USER_AGENT_SYNTHETIC_TYPEsemconv_experimental

Specifies the category of synthetic traffic, such as tests or bots.