rama::telemetry::tracing::metadata

Struct LevelFilter

pub struct LevelFilter(/* private fields */);

A filter comparable to a verbosity Level.

If a Level is considered less than or equal to a LevelFilter, it should be considered enabled; if greater than the LevelFilter, that level is disabled. See LevelFilter::current for more details.

Note that this is essentially identical to the Level type, but with the addition of an OFF level that completely disables all trace instrumentation.

See the documentation for the Level type to see how Levels and LevelFilters interact.

Implementations

impl LevelFilter

pub const OFF: LevelFilter

The “off” level.

Designates that trace instrumentation should be completely disabled.

pub const ERROR: LevelFilter

The “error” level.

Designates very serious errors.

pub const WARN: LevelFilter

The “warn” level.

Designates hazardous situations.

pub const INFO: LevelFilter

The “info” level.

Designates useful information.

pub const DEBUG: LevelFilter

The “debug” level.

Designates lower priority information.

pub const TRACE: LevelFilter

The “trace” level.

Designates very low priority, often extremely verbose, information.

pub const fn from_level(level: Level) -> LevelFilter

Returns a LevelFilter that enables spans and events with verbosity up to and including level.

pub const fn into_level(self) -> Option<Level>

Returns the most verbose Level that this filter accepts, or None if it is OFF.

pub fn current() -> LevelFilter

Returns a LevelFilter that matches the most verbose Level that any currently active Subscriber will enable.

User code should treat this as a hint. If a given span or event has a level higher than the returned LevelFilter, it will not be enabled. However, if the level is less than or equal to this value, the span or event is not guaranteed to be enabled; the subscriber will still filter each callsite individually.

Therefore, comparing a given span or event’s level to the returned LevelFilter can be used for determining if something is disabled, but should not be used for determining if something is enabled.