rama::tls::acme::proto

Module client

Available on (crate features rustls or boring or acme) and crate feature acme only.

Structs

CreateAccountOptions

Options used to create a new account, or request an identifier for an existing account, defined in rfc8555 section 7.3

FinalizePayload

Payload to request the certificate, defined in rfc8555 section 7.4

KeyAuthorization

KeyAuthorization concatenates the token for a challenge with key fingerprint, defined in rfc8555 section 8.1

NewOrderPayload

List of Identifier for which we want to issue certificate(s), defined in rfc8555 section 7.4