rama::tls::acme::proto

Module server

Available on (crate features rustls or boring or acme) and crate feature acme only.

Structs

Account

Account info returned by the acme server, defined in rfc8555 section 7.1.2

Authorization

All info about a specific Authorization, defined in rfc8555 section 7.1.4

Challenge

All info about a specific Challenge, defined in rfc8555 section 7.1.5

Directory

Directory containing all endpoints needed by an acme client, defined in rfc8555 section 7.1.1

DirectoryMeta

Extra metadata that acme server can return

Order

All info about a specific Order, defined in rfc8555 section 7.1.3

OrdersList

List of Orders currently opened by an account, defined in rfc8555 section 7.1.2.1

RawProblemResponse

Problem response such as described in rfc7808

Enums

AccountStatus

Current account status

AuthorizationStatus

ChallengeStatus

Current status of a Challenge

ChallengeType

Type of acme challenge

OrderStatus

Current status of an Order

Problem

Problem types used by acme server, as described in rfc8555 section 6.7

Constants

LOCATION_HEADER

REPLAY_NONCE_HEADER