rama::tls::boring

Module client

Available on (crate features rustls or boring or acme) and crate feature boring only.

TLS (Boring) client support for Rama.

Structs

AutoTlsStream

A stream which can be either a secure or a plain stream.

BoringTlsStream

A wrapper around an underlying raw stream which implements the SSL protocol.

ConnectorConfigClientAuth

ConnectorKindAuto

A connector which can be used to establish a connection to a server in function of the Request, meaning either it will be a seucre connector or it will be a plain connector.

ConnectorKindSecure

A connector which can only be used to establish a secure connection, regardless of the scheme of the request URI.

ConnectorKindTunnel

A connector which can be used to use this connector to support secure https tunnel connections.

EmulateTlsProfileLayer

EmulateTlsProfileService

TlsConnector

A connector which can be used to establish a connection to a server.

TlsConnectorData

TlsConnectorData that will be used by the connector

TlsConnectorDataBuilder

Use TlsConnectorDataBuilder to build a TlsConnectorData in an ergonomic way

TlsConnectorLayer

A Layer which wraps the given service with a TlsConnector.

TlsStream

A stream which can be either a secure or a plain stream.

Functions

tls_connect