rama::tls::boring::client

Struct TlsConnectorDataBuilder

pub struct TlsConnectorDataBuilder { /* private fields */ }

Available on (crate features rustls or boring or acme) and crate feature boring only.

Use TlsConnectorDataBuilder to build a TlsConnectorData in an ergonomic way

This builder is very powerful and is capable of stacking other builders. Using it this way gives each layer the option to modify what is needed in an efficient way.

Implementations

impl TlsConnectorDataBuilder

pub fn server_verify_mode(&self) -> Option<ServerVerifyMode>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn min_ssl_version(&self) -> Option<SslVersion>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn max_ssl_version(&self) -> Option<SslVersion>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn record_size_limit(&self) -> Option<u16>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn encrypted_client_hello(&self) -> Option<bool>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn store_server_certificate_chain(&self) -> Option<bool>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn grease_enabled(&self) -> Option<bool>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn ocsp_stapling_enabled(&self) -> Option<bool>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn signed_cert_timestamps_enabled(&self) -> Option<bool>

Get Copy of this field.

Will return Some(value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn server_verify_cert_store(&self) -> Option<&X509Store>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn cipher_list(&self) -> Option<&Vec<u16>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn extension_order(&self) -> Option<&Vec<u16>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn alpn_protos(&self) -> Option<&Bytes>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn curves(&self) -> Option<&Vec<SslCurve>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn verify_algorithm_prefs(&self) -> Option<&Vec<SslSignatureAlgorithm>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn client_auth(&self) -> Option<&ConnectorConfigClientAuth>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn certificate_compression_algorithms( &self, ) -> Option<&Vec<CertificateCompressionAlgorithm>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn delegated_credential_schemes( &self, ) -> Option<&Vec<SslSignatureAlgorithm>>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn server_name(&self) -> Option<&Domain>

Get reference to this field.

Will return Some(&value) if it is set on this builder. If not set on this builder base_builders will be checked starting from the end to see if one of them contains a value. The first match is returned.

pub fn keylog_filepath(&self) -> Option<Cow<'_, str>>

Return the SSL keylog file path if one exists.

pub fn new() -> TlsConnectorDataBuilder

pub fn new_http_auto() -> TlsConnectorDataBuilder

pub fn new_http_1() -> TlsConnectorDataBuilder

pub fn new_http_2() -> TlsConnectorDataBuilder

pub fn push_base_config( &mut self, config: Arc<TlsConnectorDataBuilder>, ) -> &mut TlsConnectorDataBuilder

Add [ConfigBuilder] to the end of our base builder

When evaluating builders we start from this builder (the last one) and work our way back until we find a value.

pub fn prepend_base_config( &mut self, config: Arc<TlsConnectorDataBuilder>, ) -> &mut TlsConnectorDataBuilder

Add [ConfigBuilder] to the start of our base builders

Builder in the start is evaluated as the last one when iterating over builders

pub fn with_base_config( self, config: Arc<TlsConnectorDataBuilder>, ) -> TlsConnectorDataBuilder

Add [ConfigBuilder] to the start of our base builders

pub fn set_base_config( &mut self, config: Arc<TlsConnectorDataBuilder>, ) -> &mut TlsConnectorDataBuilder

Add [ConfigBuilder] to the start of our base builders

pub fn maybe_with_server_verify_mode( self, server_verify_mode: Option<ServerVerifyMode>, ) -> TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn maybe_set_server_verify_mode( &mut self, server_verify_mode: Option<ServerVerifyMode>, ) -> &mut TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn with_server_verify_mode( self, server_verify_mode: ServerVerifyMode, ) -> TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn set_server_verify_mode( &mut self, server_verify_mode: ServerVerifyMode, ) -> &mut TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn without_server_verify_mode(self) -> TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn unset_server_verify_mode(&mut self) -> &mut TlsConnectorDataBuilder

Set the ServerVerifyMode that will be used by the tls client to verify the server

pub fn maybe_with_server_verify_cert_store( self, server_verify_cert_store: Option<Arc<X509Store>>, ) -> TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn maybe_set_server_verify_cert_store( &mut self, server_verify_cert_store: Option<Arc<X509Store>>, ) -> &mut TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn with_server_verify_cert_store( self, server_verify_cert_store: Arc<X509Store>, ) -> TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn set_server_verify_cert_store( &mut self, server_verify_cert_store: Arc<X509Store>, ) -> &mut TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn without_server_verify_cert_store(self) -> TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn unset_server_verify_cert_store(&mut self) -> &mut TlsConnectorDataBuilder

Set the X509Store that will be used by the tls client to verify the server certs

(unless the mode is ServerVerifyMode::Disable)

pub fn maybe_with_keylog_intent( self, intent: Option<KeyLogIntent>, ) -> TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn maybe_set_keylog_intent( &mut self, intent: Option<KeyLogIntent>, ) -> &mut TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn with_keylog_intent(self, intent: KeyLogIntent) -> TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn set_keylog_intent( &mut self, intent: KeyLogIntent, ) -> &mut TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn without_keylog_intent(self) -> TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn unset_keylog_intent(&mut self) -> &mut TlsConnectorDataBuilder

Set KeyLogIntent that will be used

pub fn maybe_with_cipher_list( self, list: Option<Vec<u16>>, ) -> TlsConnectorDataBuilder

Set the cipher list for this config

pub fn maybe_set_cipher_list( &mut self, list: Option<Vec<u16>>, ) -> &mut TlsConnectorDataBuilder

Set the cipher list for this config

pub fn with_cipher_list(self, list: Vec<u16>) -> TlsConnectorDataBuilder

Set the cipher list for this config

pub fn set_cipher_list( &mut self, list: Vec<u16>, ) -> &mut TlsConnectorDataBuilder

Set the cipher list for this config

pub fn without_cipher_list(self) -> TlsConnectorDataBuilder

Set the cipher list for this config

pub fn unset_cipher_list(&mut self) -> &mut TlsConnectorDataBuilder

Set the cipher list for this config

pub fn maybe_with_store_server_certificate_chain( self, store_server_certificate_chain: Option<bool>, ) -> TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn maybe_set_store_server_certificate_chain( &mut self, store_server_certificate_chain: Option<bool>, ) -> &mut TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn with_store_server_certificate_chain( self, store_server_certificate_chain: bool, ) -> TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn set_store_server_certificate_chain( &mut self, store_server_certificate_chain: bool, ) -> &mut TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn without_store_server_certificate_chain(self) -> TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn unset_store_server_certificate_chain( &mut self, ) -> &mut TlsConnectorDataBuilder

Store server certificate chain if enabled in [NegotiatedTlsParameters] extension

This will always clone the entire chain, so only enable this if needed.

pub fn maybe_with_alpn_protos( self, protos: Option<Bytes>, ) -> TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn maybe_set_alpn_protos( &mut self, protos: Option<Bytes>, ) -> &mut TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn with_alpn_protos(self, protos: Bytes) -> TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn set_alpn_protos(&mut self, protos: Bytes) -> &mut TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn without_alpn_protos(self) -> TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn unset_alpn_protos(&mut self) -> &mut TlsConnectorDataBuilder

Set alpn protos that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_maybe_with_rama_alpn_protos( self, protos: Option<&[ApplicationProtocol]>, ) -> Result<TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_maybe_set_rama_alpn_protos( &mut self, protos: Option<&[ApplicationProtocol]>, ) -> Result<&mut TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_with_rama_alpn_protos( self, protos: &[ApplicationProtocol], ) -> Result<TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_set_rama_alpn_protos( &mut self, protos: &[ApplicationProtocol], ) -> Result<&mut TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_without_rama_alpn_protos( self, ) -> Result<TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn try_unset_rama_alpn_protos( &mut self, ) -> Result<&mut TlsConnectorDataBuilder, OpaqueError>

Set ApplicationProtocol that this client will send to server

Order of protocols here is important. When server supports multiple protocols it will choose the first one it supports from this list.

pub fn maybe_with_min_ssl_version( self, version: Option<SslVersion>, ) -> TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn maybe_set_min_ssl_version( &mut self, version: Option<SslVersion>, ) -> &mut TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn with_min_ssl_version( self, version: SslVersion, ) -> TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn set_min_ssl_version( &mut self, version: SslVersion, ) -> &mut TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn without_min_ssl_version(self) -> TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn unset_min_ssl_version(&mut self) -> &mut TlsConnectorDataBuilder

Set the minimum ssl version that this connector will accept

pub fn maybe_with_max_ssl_version( self, version: Option<SslVersion>, ) -> TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn maybe_set_max_ssl_version( &mut self, version: Option<SslVersion>, ) -> &mut TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn with_max_ssl_version( self, version: SslVersion, ) -> TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn set_max_ssl_version( &mut self, version: SslVersion, ) -> &mut TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn without_max_ssl_version(self) -> TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn unset_max_ssl_version(&mut self) -> &mut TlsConnectorDataBuilder

Set the maxium ssl version that this connector will accept

pub fn maybe_with_record_size_limit( self, limit: Option<u16>, ) -> TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn maybe_set_record_size_limit( &mut self, limit: Option<u16>, ) -> &mut TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn with_record_size_limit(self, limit: u16) -> TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn set_record_size_limit( &mut self, limit: u16, ) -> &mut TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn without_record_size_limit(self) -> TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn unset_record_size_limit(&mut self) -> &mut TlsConnectorDataBuilder

Set the record size limit that will be set on this connector

pub fn maybe_with_encrypted_client_hello( self, value: Option<bool>, ) -> TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn maybe_set_encrypted_client_hello( &mut self, value: Option<bool>, ) -> &mut TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn with_encrypted_client_hello(self, value: bool) -> TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn set_encrypted_client_hello( &mut self, value: bool, ) -> &mut TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn without_encrypted_client_hello(self) -> TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn unset_encrypted_client_hello(&mut self) -> &mut TlsConnectorDataBuilder

Set if encrypted client hello should be enabled

pub fn maybe_with_grease_enabled( self, value: Option<bool>, ) -> TlsConnectorDataBuilder

Set if grease should be enable

pub fn maybe_set_grease_enabled( &mut self, value: Option<bool>, ) -> &mut TlsConnectorDataBuilder

Set if grease should be enable

pub fn with_grease_enabled(self, value: bool) -> TlsConnectorDataBuilder

Set if grease should be enable

pub fn set_grease_enabled( &mut self, value: bool, ) -> &mut TlsConnectorDataBuilder

Set if grease should be enable

pub fn without_grease_enabled(self) -> TlsConnectorDataBuilder

Set if grease should be enable

pub fn unset_grease_enabled(&mut self) -> &mut TlsConnectorDataBuilder

Set if grease should be enable

pub fn maybe_with_ocsp_stapling_enabled( self, value: Option<bool>, ) -> TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn maybe_set_ocsp_stapling_enabled( &mut self, value: Option<bool>, ) -> &mut TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn with_ocsp_stapling_enabled(self, value: bool) -> TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn set_ocsp_stapling_enabled( &mut self, value: bool, ) -> &mut TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn without_ocsp_stapling_enabled(self) -> TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn unset_ocsp_stapling_enabled(&mut self) -> &mut TlsConnectorDataBuilder

Set if ocsp stapling should be needed

pub fn maybe_with_signed_cert_timestamps_enabled( self, value: Option<bool>, ) -> TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn maybe_set_signed_cert_timestamps_enabled( &mut self, value: Option<bool>, ) -> &mut TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn with_signed_cert_timestamps_enabled( self, value: bool, ) -> TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn set_signed_cert_timestamps_enabled( &mut self, value: bool, ) -> &mut TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn without_signed_cert_timestamps_enabled(self) -> TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn unset_signed_cert_timestamps_enabled( &mut self, ) -> &mut TlsConnectorDataBuilder

Set if signed certificate timestamps should be enabled

pub fn maybe_with_extension_order( self, order: Option<Vec<u16>>, ) -> TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn maybe_set_extension_order( &mut self, order: Option<Vec<u16>>, ) -> &mut TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn with_extension_order(self, order: Vec<u16>) -> TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn set_extension_order( &mut self, order: Vec<u16>, ) -> &mut TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn without_extension_order(self) -> TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn unset_extension_order(&mut self) -> &mut TlsConnectorDataBuilder

Set the order of client hello extensions

pub fn maybe_with_curves( self, curves: Option<Vec<SslCurve>>, ) -> TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn maybe_set_curves( &mut self, curves: Option<Vec<SslCurve>>, ) -> &mut TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn with_curves(self, curves: Vec<SslCurve>) -> TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn set_curves( &mut self, curves: Vec<SslCurve>, ) -> &mut TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn without_curves(self) -> TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn unset_curves(&mut self) -> &mut TlsConnectorDataBuilder

Set the eliptic curves supported by this client

pub fn maybe_with_verify_algorithm_prefs( self, prefs: Option<Vec<SslSignatureAlgorithm>>, ) -> TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn maybe_set_verify_algorithm_prefs( &mut self, prefs: Option<Vec<SslSignatureAlgorithm>>, ) -> &mut TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn with_verify_algorithm_prefs( self, prefs: Vec<SslSignatureAlgorithm>, ) -> TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn set_verify_algorithm_prefs( &mut self, prefs: Vec<SslSignatureAlgorithm>, ) -> &mut TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn without_verify_algorithm_prefs(self) -> TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn unset_verify_algorithm_prefs(&mut self) -> &mut TlsConnectorDataBuilder

Set SslSignatureAlgorithms for verifying

pub fn maybe_with_client_auth( self, auth: Option<ConnectorConfigClientAuth>, ) -> TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn maybe_set_client_auth( &mut self, auth: Option<ConnectorConfigClientAuth>, ) -> &mut TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn with_client_auth( self, auth: ConnectorConfigClientAuth, ) -> TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn set_client_auth( &mut self, auth: ConnectorConfigClientAuth, ) -> &mut TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn without_client_auth(self) -> TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn unset_client_auth(&mut self) -> &mut TlsConnectorDataBuilder

Set client auth that will be used by this connector

pub fn maybe_with_certificate_compression_algorithms( self, algorithms: Option<Vec<CertificateCompressionAlgorithm>>, ) -> TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn maybe_set_certificate_compression_algorithms( &mut self, algorithms: Option<Vec<CertificateCompressionAlgorithm>>, ) -> &mut TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn with_certificate_compression_algorithms( self, algorithms: Vec<CertificateCompressionAlgorithm>, ) -> TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn set_certificate_compression_algorithms( &mut self, algorithms: Vec<CertificateCompressionAlgorithm>, ) -> &mut TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn without_certificate_compression_algorithms( self, ) -> TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn unset_certificate_compression_algorithms( &mut self, ) -> &mut TlsConnectorDataBuilder

Set certificate compression algorithms

pub fn maybe_with_delegated_credential_schemes( self, schemes: Option<Vec<SslSignatureAlgorithm>>, ) -> TlsConnectorDataBuilder

Set delegated credential schemes

pub fn maybe_set_delegated_credential_schemes( &mut self, schemes: Option<Vec<SslSignatureAlgorithm>>, ) -> &mut TlsConnectorDataBuilder

Set delegated credential schemes

pub fn with_delegated_credential_schemes( self, schemes: Vec<SslSignatureAlgorithm>, ) -> TlsConnectorDataBuilder

Set delegated credential schemes

pub fn set_delegated_credential_schemes( &mut self, schemes: Vec<SslSignatureAlgorithm>, ) -> &mut TlsConnectorDataBuilder

Set delegated credential schemes

pub fn without_delegated_credential_schemes(self) -> TlsConnectorDataBuilder

Set delegated credential schemes

pub fn unset_delegated_credential_schemes( &mut self, ) -> &mut TlsConnectorDataBuilder

Set delegated credential schemes

pub fn maybe_with_server_name( self, name: Option<Domain>, ) -> TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn maybe_set_server_name( &mut self, name: Option<Domain>, ) -> &mut TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn with_server_name(self, name: Domain) -> TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn set_server_name(&mut self, name: Domain) -> &mut TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn without_server_name(self) -> TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn unset_server_name(&mut self) -> &mut TlsConnectorDataBuilder

Set server name used for SNI extension

pub fn into_shared_builder(self) -> Arc<TlsConnectorDataBuilder>

pub fn build(&self) -> Result<TlsConnectorData, OpaqueError>

Build the TlsConnectorData used by the TlsConnector

NOTE: this method should in almost all circumstances never by called directly. The TlsConnector will call build only when needed and as late as possible. The only place where you manually need to build the TlsConnectorData is if you use tls_connect directly.

impl TlsConnectorDataBuilder

pub fn try_from_multiple_client_configs<'a>( cfg_it: impl Iterator<Item = &'a ClientConfig>, ) -> Result<TlsConnectorDataBuilder, OpaqueError>