rama::tls::boring::core::asn1

Struct Asn1Time

pub struct Asn1Time(/* private fields */);

Available on (crate features rustls or boring or acme) and crate feature boring only.

Time storage and comparison

Asn1Time should be used to store and share time information using certificates. If Asn1Time is set using a string, it must be in either YYMMDDHHMMSSZ, YYYYMMDDHHMMSSZ, or another ASN.1 format.

ASN_TIME_set documentation at OpenSSL explains the ASN.1 implementation used by OpenSSL.

Implementations

impl Asn1Time

pub fn days_from_now(days: u32) -> Result<Asn1Time, ErrorStack>

Creates a new time on specified interval in days from now

pub fn from_unix(time: i64) -> Result<Asn1Time, ErrorStack>

Creates a new time from the specified time_t value

This corresponds to ASN1_TIME_set.

pub fn from_str(s: &str) -> Result<Asn1Time, ErrorStack>

Creates a new time corresponding to the specified ASN1 time string.

This corresponds to ASN1_TIME_set_string.