rama::tls::boring::core::dsa

Struct Dsa

pub struct Dsa<T>(/* private fields */);

Available on (crate features rustls or boring or acme) and crate feature boring only.

Object representing DSA keys.

A DSA object contains the parameters p, q, and g. There is a private and public key. The values p, g, and q are:

p: DSA prime parameter

q: DSA sub-prime parameter

g: DSA base parameter

These values are used to calculate a pair of asymetrical keys used for signing.

OpenSSL documentation at DSA_new

Examples

use rama_boring::dsa::Dsa;
use rama_boring::error::ErrorStack;
use rama_boring::pkey::Private;

fn create_dsa() -> Result<Dsa<Private>, ErrorStack> {
    let sign = Dsa::generate(2048)?;
    Ok(sign)
}

Implementations

impl Dsa<Private>

pub fn generate(bits: u32) -> Result<Dsa<Private>, ErrorStack>

Generate a DSA key pair.

Calls DSA_generate_parameters_ex to populate the p, g, and q values. These values are used to generate the key pair with DSA_generate_key.

The bits parameter corresponds to the length of the prime p.

pub fn from_private_components( p: BigNum, q: BigNum, g: BigNum, priv_key: BigNum, pub_key: BigNum, ) -> Result<Dsa<Private>, ErrorStack>

Create a DSA key pair with the given parameters

p, q and g are the common parameters. priv_key is the private component of the key pair. pub_key is the public component of the key. Can be computed via g^(priv_key) mod p

impl Dsa<Public>

pub fn public_key_from_pem(pem: &[u8]) -> Result<Dsa<Public>, ErrorStack>

Decodes a PEM-encoded SubjectPublicKeyInfo structure containing a DSA key.

The input should have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_read_bio_DSA_PUBKEY.

pub fn public_key_from_der(der: &[u8]) -> Result<Dsa<Public>, ErrorStack>

Decodes a DER-encoded SubjectPublicKeyInfo structure containing a DSA key.

This corresponds to d2i_DSA_PUBKEY.

pub fn from_public_components( p: BigNum, q: BigNum, g: BigNum, pub_key: BigNum, ) -> Result<Dsa<Public>, ErrorStack>

Create a new DSA key with only public components.

p, q and g are the common parameters. pub_key is the public component of the key.