rama::tls::boring::core

Module tokio

Available on (crate features rustls or boring or acme) and crate feature boring only.

Full Re-export of the rama-boring-tokio crate.

Structs

AsyncPrivateKeyMethodError

A fatal error to be returned from async private key methods.

AsyncSelectCertError

A fatal error to be returned from async select certificate callbacks.

HandshakeError

The error type returned after a failed handshake.

HandshakeFuture

Future for an ongoing TLS handshake.

SslStream

A wrapper around an underlying raw stream which implements the SSL protocol.

SslStreamBuilder

A partially constructed SslStream, useful for unusual handshakes.

Traits

AsyncPrivateKeyMethod

Describes async private key hooks. This is used to off-load signing operations to a custom, potentially asynchronous, backend. Metadata about the key such as the type and size are parsed out of the certificate.

SslContextBuilderExt

Extensions to SslContextBuilder.

Functions

accept

Asynchronously performs a server-side TLS handshake over the provided stream.

connect

Asynchronously performs a client-side TLS handshake over the provided stream.

Type Aliases

BoxGetSessionFinish

The type of callbacks returned by BoxSelectCertFuture methods.

BoxGetSessionFuture

The type of futures to pass to SslContextBuilder::set_async_get_session_callback.

BoxPrivateKeyMethodFinish

The type of callbacks returned by BoxPrivateKeyMethodFuture.

BoxPrivateKeyMethodFuture

The type of futures returned by AsyncPrivateKeyMethod methods.

BoxSelectCertFinish

The type of callbacks returned by BoxSelectCertFuture methods.

BoxSelectCertFuture

The type of futures to pass to SslContextBuilder::set_async_select_certificate_callback.

ExDataFuture

Convenience alias for futures stored in Ssl ex data by SslContextBuilder methods.