rama::tls::boring::core::tokio

Function connect

pub async fn connect<S>(
    config: ConnectConfiguration,
    domain: &str,
    stream: S,
) -> Result<SslStream<S>, HandshakeError<S>>
where
    S: AsyncRead + AsyncWrite + Unpin,

Available on (crate features rustls or boring or acme) and crate feature boring only.

Asynchronously performs a client-side TLS handshake over the provided stream.

This function automatically sets the task waker on the Ssl from config to allow to make use of async callbacks provided by the boring crate.