rama::tls::boring::core::tokio

Struct SslStream

pub struct SslStream<S>(/* private fields */);

Available on (crate features rustls or boring or acme) and crate feature boring only.

A wrapper around an underlying raw stream which implements the SSL protocol.

A SslStream<S> represents a handshake that has been completed successfully and both the server and the client are ready for receiving and sending data. Bytes read from a SslStream are decrypted from S and bytes written to a SslStream are encrypted when passing through to S.

Implementations

impl<S> SslStream<S>

pub fn ssl(&self) -> &SslRef

Returns a shared reference to the Ssl object associated with this stream.

pub fn ssl_mut(&mut self) -> &mut SslRef

Returns a mutable reference to the Ssl object associated with this stream.

pub fn get_ref(&self) -> &S

Returns a shared reference to the underlying stream.

pub fn get_mut(&mut self) -> &mut S

Returns a mutable reference to the underlying stream.

impl<S> SslStream<S>
where S: AsyncRead + AsyncWrite + Unpin,

pub unsafe fn from_raw_parts(ssl: *mut ssl_st, stream: S) -> SslStream<S>

Constructs an SslStream from a pointer to the underlying OpenSSL SSL struct.

This is useful if the handshake has already been completed elsewhere.

Safety

The caller must ensure the pointer is valid.