rama::tls::boring::core

Module x509

Available on (crate features rustls or boring or acme) and crate feature boring only.

The standard defining the format of public key certificates.

An X509 certificate binds an identity to a public key, and is either signed by a certificate authority (CA) or self-signed. An entity that gets a hold of a certificate can both verify your identity (via a CA) and encrypt data with the included public key. X509 certificates are used in many Internet protocols, including SSL/TLS, which is the basis for HTTPS, the secure protocol for browsing the web.

Modules

extension

Add extensions to an X509 certificate or certificate request.

store

Describe a context in which to verify an X509 certificate.

verify

Structs

GeneralName

An X509 certificate alternative names.

GeneralNameRef

A borrowed reference to a GeneralName.

X509

An X509 public key certificate.

X509Algorithm

An X509 certificate signature algorithm.

X509AlgorithmRef

A borrowed reference to a X509Algorithm.

X509Builder

A builder used to construct an X509.

X509Extension

Permit additional fields to be added to an X509 v3 certificate.

X509ExtensionRef

A borrowed reference to a X509Extension.

X509Name

The names of an X509 certificate.

X509NameBuilder

A builder used to construct an X509Name.

X509NameEntries

A type to destructure and examine an X509Name.

X509NameEntry

A name entry associated with a X509Name.

X509NameEntryRef

A borrowed reference to a X509NameEntry.

X509NameRef

A borrowed reference to a X509Name.

X509Object

An X509 or an X509 certificate revocation list.

X509ObjectRef

A borrowed reference to a X509Object.

X509Ref

A borrowed reference to a X509.

X509Req

An X509 certificate request.

X509ReqBuilder

A builder used to construct an X509Req.

X509ReqRef

A borrowed reference to a X509Req.

X509StoreContext

An X509 certificate store context.

X509StoreContextRef

A borrowed reference to a X509StoreContext.

X509VerifyError

X509v3Context

A context object required to construct certain X509 extension values.

Type Aliases

X509VerifyResult

The result of peer certificate verification.