rama::tls::rustls

Module client

Available on (crate features rustls or boring or acme) and crate feature rustls only.

TLS client support for Rama.

Structs

AutoTlsStream

A stream which can be either a secure or a plain stream.

ConnectorKindAuto

A connector which can be used to establish a connection to a server in function of the input, meaning either it will be a secure connector or it will be a plain connector.

ConnectorKindSecure

A connector which can only be used to establish a secure connection, regardless of the scheme of the request URI.

ConnectorKindTunnel

A connector which can be used to use this connector to support secure tls tunnel connections.

TlsConnector

A connector which can be used to establish a connection to a server.

TlsConnectorData

Internal data used as configuration/input for the [super::HttpsConnector].

TlsConnectorDataBuilder

[ClientConfigBuilder] can be used to construct rustls::ClientConfig for most common use cases in Rama.

TlsConnectorLayer

A Layer which wraps the given service with a TlsConnector.

TlsStream

Functions

client_root_certs

self_signed_client_auth