rama::tls::rustls::dep

Module native_certs

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Re-export of the rustls-native-certs crate.

rustls-native-certs allows rustls to use the platform’s native certificate store when operating as a TLS client.

Structs

CertificateResult

Results from trying to load certificates from the platform’s native store.

Error

Enums

ErrorKind

Functions

load_certs_from_paths

Load certificates from the given paths.

load_native_certs

Load root certificates found in the platform’s native certificate store.