rama::tls::rustls::dep::rustls::crypto

Module aws_lc_rs

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature aws_lc_rs only.

aws-lc-rs-based CryptoProvider.

Modules

cipher_suite

All defined cipher suites supported by aws-lc-rs appear in this module.

hpke

Hybrid public key encryption (HPKE).

kx_group

All defined key exchange groups supported by aws-lc-rs appear in this module.

sign

Using software keys for authentication.

Structs

Ticketer

A concrete, safe ticket creation mechanism.

Statics

ALL_CIPHER_SUITES

A list of all the cipher suites supported by the rustls aws-lc-rs provider.

ALL_KX_GROUPS

A list of all the key exchange groups supported by this provider.

DEFAULT_CIPHER_SUITES

The cipher suite configuration that an application should use by default.

DEFAULT_KX_GROUPS

A list of the default key exchange groups supported by this provider.

Functions

default_provider

A CryptoProvider backed by aws-lc-rs.