rama::tls::rustls::dep::rustls::crypto::aws_lc_rs

Module kx_group

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature aws_lc_rs only.

All defined key exchange groups supported by aws-lc-rs appear in this module.

ALL_KX_GROUPS is provided as an array of all of these values. DEFAULT_KX_GROUPS is provided as an array of this provider’s defaults.

Statics

MLKEM768

This is the MLKEM key exchange.

SECP256R1

Ephemeral ECDH on secp256r1 (aka NIST-P256)

SECP256R1MLKEM768

This is the SECP256R1MLKEM768 key exchange.

SECP384R1

Ephemeral ECDH on secp384r1 (aka NIST-P384)

X25519

Ephemeral ECDH on curve25519 (see RFC7748)

X25519MLKEM768

This is the X25519MLKEM768 key exchange.