rama::tls::rustls::dep::rustls::crypto

Module hpke

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Hybrid public key encryption (RFC 9180).

Structs

EncapsulatedSecret

An encapsulated secret returned from setting up a sender or receiver context.

HpkeKeyPair

An HPKE key pair, made of a matching public and private key.

HpkePrivateKey

An HPKE private key.

HpkePublicKey

An HPKE public key.

HpkeSuite

An HPKE suite, specifying a key encapsulation mechanism and a symmetric cipher suite.

Traits

Hpke

An HPKE instance that can be used for base-mode single-shot encryption and decryption.

HpkeOpener

An HPKE opener context.

HpkeSealer

An HPKE sealer context.