rama::tls::rustls::dep::rustls::server

Trait ResolvesServerCert

pub trait ResolvesServerCert:
    Debug
    + Send
    + Sync {
    // Required method
    fn resolve(
        &self,
        client_hello: ClientHello<'_>,
    ) -> Option<Arc<CertifiedKey>>;

    // Provided method
    fn only_raw_public_keys(&self) -> bool { ... }
}

Available on (crate features rustls or boring or acme) and crate feature rustls only.

How to choose a certificate chain and signing key for use in server authentication.

This is suitable when selecting a certificate does not require I/O or when the application is using blocking I/O anyhow.

For applications that use async I/O and need to do I/O to choose a certificate (for instance, fetching a certificate from a data store), the Acceptor interface is more suitable.

Required Methods

fn resolve(&self, client_hello: ClientHello<'_>) -> Option<Arc<CertifiedKey>>

Choose a certificate chain and matching key given simplified ClientHello information.

Return None to abort the handshake.

Provided Methods

fn only_raw_public_keys(&self) -> bool

Return true when the server only supports raw public keys.