rama::tls::rustls::dep

Module tokio_rustls

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Full Re-export of the tokio-rustls crate.

Modules

client

rustls

Rustls - a modern TLS library

server

Structs

Accept

Future returned from TlsAcceptor::accept which will resolve once the accept handshake has finished.

Connect

Future returned from TlsConnector::connect which will resolve once the connection handshake has finished.

FallibleAccept

Like Accept, but returns IO on failure.

FallibleConnect

Like Connect, but returns IO on failure.

LazyConfigAcceptor

StartHandshake

An incoming connection received through LazyConfigAcceptor.

TlsAcceptor

A wrapper around a rustls::ServerConfig, providing an async accept method.

TlsConnector

A wrapper around a rustls::ClientConfig, providing an async connect method.

TlsConnectorWithAlpn

Enums

TlsStream

Unified TLS stream type