rama::tls::rustls::dep::tokio_rustls

Module client

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Structs

Connect

Future returned from TlsConnector::connect which will resolve once the connection handshake has finished.

FallibleConnect

Like Connect, but returns IO on failure.

TlsConnector

A wrapper around a rustls::ClientConfig, providing an async connect method.

TlsConnectorWithAlpn

TlsStream

A wrapper around an underlying raw stream which implements the TLS or SSL protocol.