rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom::bytes::streaming

Function take_till1

pub fn take_till1<F, Input, Error>(cond: F) -> impl Fn(Input)
where
    Error: ParseError<Input>,
    Input: InputTakeAtPosition,
    F: Fn(<Input as InputTakeAtPosition>::Item) -> bool,

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Returns the longest (at least 1) input slice till a predicate is met.

The parser will return the longest slice till the given predicate (a function that takes the input and returns a bool).

Streaming Specific

Streaming version will return a Err::Incomplete(Needed::new(1)) if the match reaches the end of input or if there was not match.

Example

use nom::bytes::streaming::take_till1;

fn till_colon(s: &str) -> IResult<&str, &str> {
  take_till1(|c| c == ':')(s)
}

assert_eq!(till_colon("latin:123"), Ok((":123", "latin")));
assert_eq!(till_colon(":empty matched"), Err(Err::Error(Error::new(":empty matched", ErrorKind::TakeTill1))));
assert_eq!(till_colon("12345"), Err(Err::Incomplete(Needed::new(1))));
assert_eq!(till_colon(""), Err(Err::Incomplete(Needed::new(1))));