rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom::lib::std::collections

Module vec_deque
·

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature std and non-no_global_oom_handling only.

A double-ended queue (deque) implemented with a growable ring buffer.

This queue has O(1) amortized inserts and removals from both ends of the container. It also has O(1) indexing like a vector. The contained elements are not required to be copyable, and the queue will be sendable if the contained type is sendable.

Structs

Drain

A draining iterator over the elements of a VecDeque.

IntoIter

An owning iterator over the elements of a VecDeque.

Iter

An iterator over the elements of a VecDeque.

IterMut

A mutable iterator over the elements of a VecDeque.

VecDeque

A double-ended queue implemented with a growable ring buffer.

ExtractIfExperimental

An iterator which uses a closure to determine if an element should be removed.