rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom::lib::std::iter

Trait FusedIterator
·

pub trait FusedIterator: Iterator { }

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature std only.

An iterator that always continues to yield None when exhausted.

Calling next on a fused iterator that has returned None once is guaranteed to return None again. This trait should be implemented by all iterators that behave this way because it allows optimizing Iterator::fuse().

Note: In general, you should not use FusedIterator in generic bounds if you need a fused iterator. Instead, you should just call Iterator::fuse() on the iterator. If the iterator is already fused, the additional Fuse wrapper will be a no-op with no performance penalty.