rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom::lib::std::ops

Trait SubAssign
·

pub trait SubAssign<Rhs = Self> {
    // Required method
    fn sub_assign(&mut self, rhs: Rhs);
}

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature std only.

The subtraction assignment operator -=.

Examples

This example creates a Point struct that implements the SubAssign trait, and then demonstrates sub-assigning to a mutable Point.

use std::ops::SubAssign;

#[derive(Debug, Copy, Clone, PartialEq)]
struct Point {
    x: i32,
    y: i32,
}

impl SubAssign for Point {
    fn sub_assign(&mut self, other: Self) {
        *self = Self {
            x: self.x - other.x,
            y: self.y - other.y,
        };
    }
}

let mut point = Point { x: 3, y: 3 };
point -= Point { x: 2, y: 3 };
assert_eq!(point, Point {x: 1, y: 0});

Required Methods

fn sub_assign(&mut self, rhs: Rhs)

Performs the -= operation.

Example

let mut x: u32 = 12;
x -= 1;
assert_eq!(x, 11);