rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom::lib::std::str

Function from_boxed_utf8_unchecked
·

pub unsafe fn from_boxed_utf8_unchecked(v: Box<[u8]>) -> Box<str>

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature std only.

Converts a boxed slice of bytes to a boxed string slice without checking that the string contains valid UTF-8.

Safety

The provided bytes must contain a valid UTF-8 sequence.

Examples

let smile_utf8 = Box::new([226, 152, 186]);
let smile = unsafe { std::str::from_boxed_utf8_unchecked(smile_utf8) };

assert_eq!("☺", &*smile);