rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom

Trait Offset

pub trait Offset {
    // Required method
    fn offset(&self, second: &Self) -> usize;
}

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Useful functions to calculate the offset between slices and show a hexdump of a slice

Required Methods

fn offset(&self, second: &Self) -> usize

Offset between the first byte of self and the first byte of the argument

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.