rama::tls::rustls::dep::x509_parser::prelude::asn1_rs

Struct SequenceIterator

pub struct SequenceIterator<'a, T, F, E = Error>
where
    F: ASN1Parser,
{ /* private fields */ }

Available on (crate features rustls or boring or acme) and crate feature rustls only.

An Iterator over binary data, parsing elements of type T

This helps parsing SEQUENCE OF items of type T. The type of parser (BER/DER) is specified using the generic parameter F of this struct.

Note: the iterator must start on the sequence contents, not the sequence itself.

Examples

use asn1_rs::{DerParser, Integer, SequenceIterator};

let data = &[0x30, 0x6, 0x2, 0x1, 0x1, 0x2, 0x1, 0x2];
for (idx, item) in SequenceIterator::<Integer, DerParser>::new(&data[2..]).enumerate() {
    let item = item.unwrap(); // parsing could have failed
    let i = item.as_u32().unwrap(); // integer can be negative, or too large to fit into u32
    assert_eq!(i as usize, idx + 1);
}

Implementations

impl<'a, T, F, E> SequenceIterator<'a, T, F, E>
where F: ASN1Parser,

pub fn new(data: &'a [u8]) -> SequenceIterator<'a, T, F, E>