rama::tls::rustls::dep::x509_parser::prelude::der_parser::error

Type Alias BerResult

pub type BerResult<'a, O = BerObject<'a>> = Result<(&'a [u8], O), Err<Error>>;

Available on (crate features rustls or boring or acme) and crate feature rustls only.

Holds the result of parsing functions

O is the output type, and defaults to a BerObject.

Note that this type is also a Result, so usual functions (map, unwrap etc.) are available.

This type is a wrapper around nom’s IResult type

Aliased Type

pub enum BerResult<'a, O = BerObject<'a>> {
    Ok((&'a [u8], O)),
    Err(Err<Error>),
}

Variants

Ok((&'a [u8], O))

Contains the success value

Err(Err<Error>)

Contains the error value