rama

Trait Service

pub trait Service<Input>:
    Sized
    + Send
    + Sync
    + 'static {
    type Output: Send + 'static;
    type Error: Send + 'static;

    // Required method
    fn serve(
        &self,
        input: Input,
    ) -> impl Future<Output = Result<Self::Output, Self::Error>> + Send;

    // Provided method
    fn boxed(self) -> BoxService<Input, Self::Output, Self::Error> { ... }
}

A Service that produces rama services, to serve given an input, be it transport layer Inputs or application layer http requests, or something else entirely.

Required Associated Types

type Output: Send + 'static

The type of the output returned by the service.

type Error: Send + 'static

The type of error returned by the service.

Required Methods

fn serve( &self, input: Input, ) -> impl Future<Output = Result<Self::Output, Self::Error>> + Send

Serve a an output or error for the given input, using the given context.

Provided Methods

fn boxed(self) -> BoxService<Input, Self::Output, Self::Error>

Box this service to allow for dynamic dispatch.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.