rama::ua::layer

Module emulate

Available on crate feature ua only.

Emulate user agent (UA) profiles.

This module contains the profiles for the User Agent (UA) that are used for emulation.

Learn more about User Agents (UA) and why Rama supports it at https://ramaproxy.org/book/intro/user_agent.html.

Ethics

At Plabayo, we support the principle that information wants to be free, provided it is pursued ethically and within the bounds of applicable law.

We do not endorse or support any malicious use of our technology. We advocate for the programmatic retrieval of publicly available data only when conducted responsibly — in a manner that is respectful, does not impose an undue burden on servers, and avoids causing disruption, harm, or degradation to third-party services.

Structs

SelectedUserAgentProfile

Extra information about the selected user agent profile, which isn’t already injected. E.g. http and tls information is already injected separately.

UserAgentEmulateHttpConnectModifier

UserAgentEmulateHttpConnectModifierLayer

UserAgentEmulateHttpRequestModifier

UserAgentEmulateHttpRequestModifierLayer

UserAgentEmulateLayer

A layer that emulates a user agent profile.

UserAgentEmulateService

Service to select a [UserAgentProfile] and inject its info into the input [Context].

Enums

UserAgentSelectFallback

Fallback strategy that can be injected into the context to customise what a provider can be requested to do in case the preconditions for UA selection were not fulfilled.

Traits

UserAgentProvider

A trait for providing user agent profiles for emulation.