rama::ua

Module profile

Available on crate feature ua only.

User Agent (UA) Profiles, mostly used for emulation.

See UserAgentProfile for the main profile type and UserAgentEmulateService for the service that triggers the emulation.

This module contains the profiles for the User Agent (UA) that are used for emulation.

Learn more about User Agents (UA) and why Rama supports it at https://ramaproxy.org/book/intro/user_agent.html.

Structs

Http1Profile

The HTTP/1.1 profile.

Http1Settings

The settings for the HTTP/1.1 profile.

Http2Profile

The HTTP/2 profile.

Http2Settings

The settings for the HTTP/2 profile.

HttpHeadersProfile

HttpProfile

A User Agent (UA) profile for HTTP.

JsProfile

Javascript information for a user-agent which supports Javascript and Web APIs.

JsProfileNavigator

Navigator information, if the user-agent supports the Navigator API.

JsProfileScreen

Screen information, if the user-agent supports the Screen API.

JsProfileWebApis

PreserveHeaderUserAgent

Runtime hint to request a user agent to be preserved, useful for systems that modify requests based on the context and request, but still wish to support preserving the original header user-agent.

TlsProfile

Profile of the user-agent’s TLS (client) configuration.

UserAgentDatabase

Reference implementation of a UserAgentProvider.

UserAgentProfile

The main profile for the user-agent.

UserAgentRuntimeProfile

Runtime (meta) info about the UA profile.

UserAgentSourceInfo

Source information injected by fingerprinting service.

WsClientConfigOverwrites

Client Config (overwrites) specific to WebSocket traffic.

Enums

RequestInitiator

The initiator of the (http) Request.

Statics

CUSTOM_HEADER_MARKER

Marker header name for custom headers.

Functions

try_load_embedded_profiles

Load the profiles embedded with the rama-ua crate.

Type Aliases

RequestClientHints

ClientHints requested for the (http) Request.