rama

Crate udp

Available on crate feature udp only.

UDP module for Rama.

Rama

Crate used by the end-user rama crate and rama crate authors alike.

Learn more about rama:

Github: https://github.com/plabayo/rama

Book: https://ramaproxy.org/book/

Structs

UdpFramed

A unified Stream and Sink interface to an underlying UdpSocket, using the Encoder and Decoder traits to encode and decode frames.

UdpSocketnet

A UDP socket.

Functions

bind_udp

Creates a new UdpSocket, which will be bound to the specified Interface.

bind_udp_with_address

Creates a new UdpSocket, which will be bound to the specified address.

bind_udp_with_deviceAndroid or Fuchsia or Linux

Creates a new UdpSocket, which will be bound to the specified (interface) device name).

bind_udp_with_socketWindows or target_family=unix

Creates a new UdpSocket, which will be bound to the specified socket.