rama

Crate unix

Available on target_family=unix and crate feature net only.

Unix (Domain) socket support for Rama.

Rama

Crate used by the end-user rama crate and rama crate authors alike.

Learn more about rama:

Github: https://github.com/plabayo/rama

Book: https://ramaproxy.org/book/

Modules

client

Unix (domain) socket client (connector) module for Rama.

server

Unix (domain) socket server module for Rama.

Structs

ClientUnixSocketInfo

Information about the socket on the egress end.

TokioSocketAddress

An address associated with a Tokio Unix socket.

TokioUnixStreamUnix

A structure representing a connected Unix socket.

UnixDatagramUnix

An I/O object representing a Unix datagram socket.

UnixDatagramFramed

A unified Stream and Sink interface to an underlying UnixDatagram, using the Encoder and Decoder traits to encode and decode frames.

UnixSocketUnix

A Unix socket that has not yet been converted to a UnixStream, UnixDatagram, or UnixListener.

UnixSocketAddress

An address associated with a Unix socket.

UnixSocketInfo

Connected unix socket information.

UnixStream

A stream which can be either a secure or a plain stream.