rama::unix

Module client

Available on target_family=unix and crate feature net only.

Unix (domain) socket client (connector) module for Rama.

Structs

UnixConnector

A connector which can be used to establish a Unix connection to a server.

UnixStreamConnectorCloneFactory

Utility implementation of a UnixStreamConnectorFactory which is implemented to allow one to use a Cloneable UnixStreamConnector as a UnixStreamConnectorFactory by cloning itself.

UnixTarget

Type of UnixConnector which connects to a fixed [Path].

Traits

UnixStreamConnector

Trait used by the UnixConnector to actually establish the UnixStream.

UnixStreamConnectorFactory

Factory to create a UnixStreamConnector. This is used by the Unix stream service to create a stream within a specific [Context].