rama

Module utils

utilities for rama

Modules

backoff

This module contains generic backoff utilities to be used with the retry and limit layers.

collections

Collections such as NonEmptyVec and NonEmptySmallVec provided by rama, mostly for internal usage, but available for others to use as well.

include_dir

An extension to the include_str!() and include_bytes!() macro for embedding an entire directory tree into your binary.

info

Info constants about the rama crate.

latency

latency utilities and common types

octets

rng

PRNG utilities for middleware.

str

string utilities

thirdparty

Thirdparty utilities.

towertower

Tower support for Rama.