rama::utils

Module backoff

This module contains generic backoff utilities to be used with the retry and limit layers.

The Backoff trait is a generic way to represent backoffs that can use any timer type.

ExponentialBackoff which implements the Backoff trait and provides a batteries included exponential backoff and jitter strategy.

Structs

ExponentialBackoff

A jittered exponential backoff strategy.

Traits

Backoff

A backoff trait where a single mutable reference represents a single backoff session.