rama::utils

Module collections

Collections such as NonEmptyVec and NonEmptySmallVec provided by rama, mostly for internal usage, but available for others to use as well.

Modules

smallvec

Small vectors in various sizes. These store a certain number of elements inline, and fall back to the heap for larger allocations. This can be a useful optimization for improving cache locality and reducing allocator traffic for workloads that fit within the inline buffer.

Structs

NonEmptySmallVec

A Non-empty stack vector which can grow to the heap.

NonEmptySmallVecEmptyError

empty value cannot be turned into a NonEmptySmallVec

NonEmptySmallVecIter

Iterator for NonEmptySmallVec.

NonEmptyVec

A Non-empty growable vector.

NonEmptyVecEmptyError

empty value cannot be turned into a NonEmptyVec

NonEmptyVecIter

Iterator for NonEmptyVec.