rama::utils

Module rng

PRNG utilities for middleware.

This module provides a generic Rng trait and a HasherRng that implements the trait based on RandomState or any other Hasher.

Structs

HasherRng

A Rng implementation that uses a Hasher to generate the random values. The implementation uses an internal counter to pass to the hasher for each iteration of Rng::next_u64.

Traits

Rng

A simple PRNG trait for use within middleware.