rama::utils::str

Module utf8

Incremental, zero-copy UTF-8 decoding with error handling

Forked from archived project: https://github.com/SimonSapin/rust-utf8/tree/218fea2b57b0e4c3de9fa17a376fcc4a4c0d08f3

Original License: https://github.com/SimonSapin/rust-utf8/blob/218fea2b57b0e4c3de9fa17a376fcc4a4c0d08f3/LICENSE-MIT

Credits to Simon Sapin (@SimonSapin)

Structs

BufReadDecoder

Wraps a std::io::BufRead buffered byte stream and decode it as UTF-8.

Incomplete

LossyDecoder

A push-based, lossy decoder for UTF-8. Errors are replaced with the U+FFFD replacement character.

Enums

BufReadDecoderError

DecodeError

Constants

REPLACEMENT_CHARACTER

The replacement character, U+FFFD. In lossy decoding, insert it for every decoding error.

Functions

decode