rama::http::header

Trait AsHeaderName

pub trait AsHeaderName: Sealed { }

Available on crate feature http only.

A marker trait used to identify values that can be used as search keys to a HeaderMap.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.