rama::http::headers::authorization

Trait AuthoritySync

pub trait AuthoritySync<C, L>:
    Send
    + Sync
    + 'static {
    // Required method
    fn authorized(&self, ext: &mut Extensions, credentials: &C) -> bool;
}

Available on crate feature http only.

A synchronous version of Authority, to be used for primitive implementations.

Required Methods

fn authorized(&self, ext: &mut Extensions, credentials: &C) -> bool

Returns true if the credentials are authorized, otherwise false.