rama::http::request

Trait HttpRequestParts

pub trait HttpRequestParts: ExtensionsRef {
    // Required methods
    fn method(&self) -> &Method;
    fn uri(&self) -> &Uri;
    fn version(&self) -> Version;
    fn headers(&self) -> &HeaderMap;
}

Available on crate feature http only.

HttpRequestParts is used in places where we don’t need the ReqBody of the HttpRequest

In those places we need to support using HttpRequest and Parts. By using this trait we can support both types behind a single generic that implements this trait.

Required Methods

fn method(&self) -> &Method

fn uri(&self) -> &Uri

fn version(&self) -> Version

fn headers(&self) -> &HeaderMap