rama::http::service::client::ext

Trait IntoHeaderName

pub trait IntoHeaderName: IntoHeaderNameSealed { }

Available on crate feature http only.

A trait to try to convert some type into a HeaderName.

This trait is “sealed”, such that only types within rama can implement it.