rama::http::service::client::ext

Trait IntoUrl

pub trait IntoUrl: IntoUrlSealed { }

Available on crate feature http only.

A trait to try to convert some type into a Url.

This trait is “sealed”, such that only types within rama can implement it.