rama::matcher

Trait IteratorMatcherExt

pub trait IteratorMatcherExt<'a, M, Input>: Iterator<Item = &'a M> + 'a
where
    M: Matcher<Input>,
{
    // Required methods
    fn matches_and(self, ext: Option<&mut Extensions>, input: &Input) -> bool;
    fn matches_or(self, ext: Option<&mut Extensions>, input: &Input) -> bool;
}

Extension to apply matcher operations to an Iterator of Matchers.

Required Methods

fn matches_and(self, ext: Option<&mut Extensions>, input: &Input) -> bool

Matches in case all Matcher elements match for the given Input within the specified Extensions.

fn matches_or(self, ext: Option<&mut Extensions>, input: &Input) -> bool

Matches in case any of the Matcher elements match for the given Input within the specified Extensions.