rama::net::address

Trait IntoDomain

pub trait IntoDomain: IntoDomainImpl { }

Available on crate feature net only.

A trait which can be use by crates where a Domain is expected, it can however only be implemented by the rama-net rate.