rama::proxy

Trait UsernameFormatter

pub trait UsernameFormatter:
    Send
    + Sync
    + 'static {
    // Required method
    fn fmt_username(
        &self,
        proxy: &Proxy,
        filter: &ProxyFilter,
        username: &str,
    ) -> Option<String>;
}

Available on (crate features proxy or haproxy or socks5) and crate feature proxy only.

Trait that is used to allow the formatting of a username, e.g. to allow proxy routers to have proxy config labels in the username.

Required Methods

fn fmt_username( &self, proxy: &Proxy, filter: &ProxyFilter, username: &str, ) -> Option<String>

format the username based on the root properties of the given proxy.