rama::telemetry::opentelemetry::propagation

Trait Injector

pub trait Injector {
    // Required method
    fn set(&mut self, key: &str, value: String);
}

Available on crate features opentelemetry and trace only.

Injector provides an interface for adding fields from an underlying struct like HashMap

Required Methods

fn set(&mut self, key: &str, value: String)

Add a key and value to the underlying data.