rama::telemetry::opentelemetry

Trait AttributesFactory

pub trait AttributesFactory:
    Send
    + Sync
    + 'static {
    // Required method
    fn attributes(&self, size_hint: usize, ext: &Extensions) -> Vec<KeyValue>;
}

Available on crate feature opentelemetry only.

Trait that can be used to implement your own attributes creator. It is used by layers as a starting point for attributes, and they will add their own attributes on top.

Required Methods

fn attributes(&self, size_hint: usize, ext: &Extensions) -> Vec<KeyValue>

Create an attributes Vec.

The size_hint indicates how many attributes the callee may wish to add on top