rama::tls::boring::core::stack

Trait Stackable

pub trait Stackable: ForeignType {
    type StackType;
}

Available on (crate features rustls or boring or acme) and crate feature boring only.

Trait implemented by types which can be placed in a stack.

It should not be implemented for any type outside of this crate.

Required Associated Types

type StackType

The C stack type for this element.

Generally called stack_st_{ELEMENT_TYPE}, normally hidden by the STACK_OF(ELEMENT_TYPE) macro in the OpenSSL API.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.