rama::tls::rustls::dep::x509_parser::prelude::asn1_rs::nom

Trait HexDisplay

pub trait HexDisplay {
    // Required methods
    fn to_hex(&self, chunk_size: usize) -> String;
    fn to_hex_from(&self, chunk_size: usize, from: usize) -> String;
}

Available on (crate features rustls or boring or acme) and crate feature rustls and crate feature std only.

Helper trait to show a byte slice as a hex dump

Required Methods

fn to_hex(&self, chunk_size: usize) -> String

Converts the value of self to a hex dump, returning the owned String.

fn to_hex_from(&self, chunk_size: usize, from: usize) -> String

Converts the value of self to a hex dump beginning at from address, returning the owned String.