rama::utils::thirdparty::wildcard

Trait WildcardSymbol

pub trait WildcardSymbol: Eq + Copy {
    const DEFAULT_METASYMBOL_ANY: Self;
    const DEFAULT_METASYMBOL_ONE: Self;
    const DEFAULT_METASYMBOL_ESCAPE: Self;

    // Required method
    fn eq_case_insensitive(a: Self, b: Self) -> bool;
}

This trait defined the alphabet a wildcard can be used on.

Required Associated Constants

const DEFAULT_METASYMBOL_ANY: Self

Default metasymbol that matches on sequences of arbitrary length.

This is typically *.

const DEFAULT_METASYMBOL_ONE: Self

Default metasymbol that matches on a single symbol.

This is typically ?.

const DEFAULT_METASYMBOL_ESCAPE: Self

Default metasymbol used for escaping.

This is typically \.

Required Methods

fn eq_case_insensitive(a: Self, b: Self) -> bool

Checks equality in a case-insensitive way.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.